% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_helper}
\alias{add_helper}
\title{Add helping text}
\usage{
add_helper(
  shiny_tag,
  content,
  title = "What do I have to do here?",
  type = "inline",
  size = "s",
  ...
)
}
\description{
Add helping text
}
