% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{add_vardenotation_to_cypro_object_shiny}
\alias{add_vardenotation_to_cypro_object_shiny}
\title{Adds variable denotation from pdl to object}
\usage{
add_vardenotation_to_cypro_object_shiny(object, amils)
}
\description{
Adds variable denotation from pdl to object
}
