% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAnalysisAspects}
\alias{addAnalysisAspects}
\alias{addAnalysisAspects,Cypro-method}
\alias{addAnalysisAspects,CyproTimeLapseMP-method}
\title{Add analysis aspects}
\usage{
addAnalysisAspects(object, ...)

\S4method{addAnalysisAspects}{Cypro}(object, fset_name, variables_numeric)

\S4method{addAnalysisAspects}{CyproTimeLapseMP}(object, fset_name, variables_numeric)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
The input object.
}
\description{
Adds new analysis aspects for a new feature set.
}
