% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addClusterVariables}
\alias{addClusterVariables}
\alias{addClusterVariables,Cypro-method}
\alias{addClusterVariables,CyproTimeLapseMP-method}
\title{Add cluster variables}
\usage{
addClusterVariables(object, ...)

\S4method{addClusterVariables}{Cypro}(
  object,
  input_df,
  variable_names,
  key = "cell_id",
  force = FALSE,
  verbose = TRUE
)

\S4method{addClusterVariables}{CyproTimeLapseMP}(
  object,
  phase,
  input_df,
  variable_names,
  key = "cell_id",
  verbose = TRUE
)

\S4method{addClusterVariables}{CyproTimeLapseMP}(
  object,
  phase,
  input_df,
  variable_names,
  key = "cell_id",
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{input_df}{Data.frame that contains the cluster variables that are supposed
to be added.}

\item{variable_names}{Character vector. The names of the variables that are
supposed to be added.}

\item{key}{Character value. The key variable by which the variables are added.
Defaults to \code{cell_id}.}

\item{force}{Logical value. Needs to be set to TRUE if you want to overwrite an already existing
set up or already existing results.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\description{
Add cluster variables to cluster data.frame of the \code{Cypro}
object.
}
