% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/add.R
\name{addClusterVarsHclust}
\alias{addClusterVarsHclust}
\alias{addClusterVarsHclust,Cypro-method}
\alias{addClusterVarsHclust,CyproTimeLapseMP-method}
\title{Add hierarchical cluster variables}
\usage{
addClusterVarsHclust(object, ...)

\S4method{addClusterVarsHclust}{Cypro}(
  object,
  fset = "all_features",
  ks = NULL,
  hs = NULL,
  methods_dist = "euclidean",
  method_aggl = "ward.D",
  prefix = "",
  naming_k = "{method_dist}_{method_aggl}_k{k}",
  naming_h = "{method_dist}_{method_aggl}_h{h}",
  verbose = TRUE
)

\S4method{addClusterVarsHclust}{CyproTimeLapseMP}(
  object,
  phase,
  fset = "all_features",
  ks = NULL,
  hs = NULL,
  methods_dist = "euclidean",
  method_aggl = "ward.D",
  prefix = "",
  naming_k = "{method_dist}_{method_aggl}_k{k}",
  naming_h = "{method_dist}_{method_aggl}_h{h}",
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\description{
Adds hierarchical cluster results in form of grouping variables
to the data of the input object.
}
\details{
Iterates over all possible combinations of the arguments \code{ks}, \code{hs},
\code{methods_dist} and \code{methods_aggl} and adds the corresponding grouping
variables to the objects data. Cluster variables are named corresponding to the input
of arguments \code{naming_k} and \code{naming_h}.
}
