% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/add.R
\name{addClusterVarsKmeans}
\alias{addClusterVarsKmeans}
\alias{addClusterVarsKmeans,Cypro-method}
\alias{addClusterVarsKmeans,CyproTimeLapseMP-method}
\title{Add kmeans cluster variables}
\usage{
addClusterVarsKmeans(object, ...)

\S4method{addClusterVarsKmeans}{Cypro}(
  object,
  ks,
  fset = "all_features",
  methods_kmeans = "Hartigan-Wong",
  prefix = "",
  naming = "{method_kmeans}_k{k}",
  force = FALSE,
  verbose = TRUE
)

\S4method{addClusterVarsKmeans}{CyproTimeLapseMP}(
  object,
  phase,
  ks,
  fset = "all_features",
  methods_kmeans = "Hartigan-Wong",
  prefix = "",
  naming = "{method_kmeans}_k{k}",
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\description{
Adds kmeans cluster results in form of grouping variables
to the data of the input object.
}
\details{
Iterates over all possible combinations of the arguments \code{ks}
and \code{methods_kmeans} and adds the corresponding grouping
variables to the objects data. Cluster variables are named corresponding to the input
of argument \code{naming}.
}
