% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addFeatureSet}
\alias{addFeatureSet}
\alias{addFeatureSet,CyproScreening-method}
\alias{addFeatureSet,CyproTimeLapse-method}
\alias{addFeatureSet,CyproTimeLapseMP-method}
\title{Add feature set}
\usage{
addFeatureSet(object, ...)

\S4method{addFeatureSet}{CyproScreening}(object, fset_name, features, force = FALSE, verbose = TRUE)

\S4method{addFeatureSet}{CyproTimeLapse}(object, fset_name, stat_features, force = FALSE, verbose = TRUE)

\S4method{addFeatureSet}{CyproTimeLapseMP}(object, fset_name, stat_features, force = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{fset_name}{Character value. The name of the feature set.}

\item{features}{Character vector. Denotes the features of interest. \emph{Features} are numeric
variables in the cell data (slot @cdata) of the \code{Cypro} object.

Use \code{getFeatureNames()} to obtain all valid input options.}

\item{force}{Logical value. Needs to be set to TRUE if you want to overwrite an already existing
set up or already existing results.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{stat_features}{Character vector. Denotes the stat features of interest. \emph{Stat features}
are the summarized features by cell ID in the statistics data.frame of slot @cdata.

Use \code{getStatFeatureNames()} to obtain all valid input options.}
}
\description{
With this function defined sets of cell features are set. See
details for more.
}
\details{
The features specified of argument \code{features} or \code{stat_features}
are stored as a new feature set. Based on that clustering, correlation, etc.
can be conducted.
}
\section{Concept}{

The concept of \emph{feature sets} is used in \code{cypro} to facilitate
the analysis of data sets that contain an abundance of features as is often obtained
by high content screening. It allows to conveniently analyze the data from several points of
view as machine learning results (clustering, correlation, dimensional reduction, etc.)
are stored by feature set. E.g. you can cluster cells according to feature set \emph{area_shape}
that you might have defined as a set of features that only measure shape related properties
of the cell and do the same for a feature set \emph{migration}.
}

\seealso{
\code{getFeatureSet()}, \code{getFeautureSets()}, \code{getFeatureSetNames()}
}
