% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addHierarchicalClusterVariables}
\alias{addHierarchicalClusterVariables}
\title{Add hierarchical clustering results to overall data}
\usage{
addHierarchicalClusterVariables(
  object,
  variable_set,
  phase = NULL,
  method_dist = NULL,
  method_aggl = NULL,
  k = NULL,
  h = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{method_dist}{Character vector (or value see details for more.) Denotes the distance method(s)
of interest (e.g. \emph{'euclidean'} or \emph{'manhattan'}).

Use \code{validDistanceMethods()} to obtain all valid input options.}

\item{method_aggl}{Character vector (or value see details for more.) Denotes the agglomeration
method(s) of interest according to which the existing distance matrices are agglomerated to hierarchical
trees.

Use \code{validAgglomerationMethods()} to obtain all valid input options.}

\item{k}{Numeric vector. Denotes the exact number of clusters in which the tree created
according to the distance- and agglomeration method is supposed to be cut.}

\item{h}{Numeric vector. Denotes the heights at which the hierarchical tree created
according to the distance- and agglomeration method is supposed to be cut.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated cypro object that contains the data added.
}
\description{
Adds hierarchical clustering results in form of
grouping variables to the object's overall data - making them available for the
\code{across}-argument..
}
\details{
The last step of the hierarchical clustering pipeline. This function iterates
over all combinations of \code{method_dist}, \code{method_aggl}, \code{k} and \code{h} and
adds the respective clustering variables to the object's overall data named according to
the following syntax: \emph{hcl_\code{method_dist}_\code{method_aggl}_k/h_\code{k}/\code{h}_\code{(variable_set)}}.
This naming concept results in somewhat bulky but unambiguous clustering names. You can always
rename grouping variables with \code{renameClusterDf()}.

Use \code{getGroupingVariableNames()} afterwards to obtain all grouping variables.
}
