% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addMetaVariables}
\alias{addMetaVariables}
\alias{addMetaVariables,Cypro-method}
\title{Add meta variables}
\usage{
addMetaVariables(object, ...)

\S4method{addMetaVariables}{Cypro}(
  object,
  input_df,
  variable_names,
  key = "cell_id",
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{input_df}{Data.frame that contains the meta variables that are supposed
to be added.}

\item{variable_names}{Character vector. The names of the variables that are
supposed to be added.}

\item{key}{Character value. The key variable by which the variables are added.
Defaults to \code{cell_id}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Add meta variables to meta data.frame of the \code{Cypro}
object.
}
