% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addRoiInfoVarNames}
\alias{addRoiInfoVarNames}
\alias{addRoiInfoVarNames.layout_df}
\title{Add variabels to layout nesting}
\usage{
addRoiInfoVarNames(df, vars)

\method{addRoiInfoVarNames}{layout_df}(df, vars)
}
\arguments{
\item{df}{A layout data.frame.}

\item{vars}{Character vector. Names of variables.}
}
\value{
Input data.frame with adjusted attributes.
}
\description{
Adds variable names to the attributes with with
\code{nestLayoutDf()} / \code{unnestLayoutDf()} shift between
the layout data.frame scops.
}
\details{
A call to \code{base::unique()} silently prevents
duplicated values.
}
