% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addStatVariables}
\alias{addStatVariables}
\title{Add numeric variables}
\usage{
addStatVariables(
  object,
  input_df,
  variable_names,
  phase = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{input_df}{A data.frame that contains the variables denoted in argument \code{variable_names}
as well as a character variable named according to input of argument \code{by} that is used to match
and join both data.frames.}

\item{variable_names}{Character vector. The name of the variables that are to be joined.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{stat_df}{A data.frame that contains the variables denoted in argument \code{variable_names}
as well as a character variable named \emph{cell_id} matching the cell ids of the input object.}
}
\value{
An updated cypro object that contains the data added.
}
\description{
Allows to join numeric variables that can be referred
to via the \code{variables}-argument of many functions.
}
