% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addSubset}
\alias{addSubset}
\alias{addSubset,Cypro-method}
\title{Add subset information}
\usage{
addSubset(object, subset_object)

\S4method{addSubset}{Cypro}(object, subset_object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{subset_object}{An object of class \code{CyproSubset}.}
}
\value{
An updated cypro object that contains the data added.
}
\description{
Adds a new object of class \code{CyproSubset} to @slot
subset.
}
