% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addVariableSet}
\alias{addVariableSet}
\title{Add predefined set of variables}
\usage{
addVariableSet(
  object,
  variable_names,
  set_name,
  overwrite = FALSE,
  phase = NULL
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_names}{Character vector. Denotes the variable names that are part
of the variable set.}

\item{set_name}{Character value. The name of the set of variable names.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Allows to store predefined set of variable names in the object that
can be easily accessed via \code{getVariableSet()}.
}
