% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addWellPlate}
\alias{addWellPlate}
\alias{addWellPlate,ExperimentDesign-method}
\title{Add well plate S4 object}
\usage{
addWellPlate(object, ...)

\S4method{addWellPlate}{ExperimentDesign}(object, name, layout, ...)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{methods::new(Class = \emph{'WellPlate'}, ...)}.}

\item{name, layout}{Input as described in documentation of class \code{WellPlate}.
See details for more information.}
}
\value{
The input object.
}
\description{
Creates and adds a new object of class \code{WellPlate} in the respective
slot of input for argument \code{object}.
}
\details{
Slot @index of the new well plate is automatically defined by adding 1 to the sum of all well plates found.
Slot @experiment of the new well plate is inherited by the

Input for slot @name must not be already used by another well plate.
}
\seealso{
\code{WellPlate-class}
}
