% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{adjust_well_roi_vec}
\alias{adjust_well_roi_vec}
\title{Function to make well roi vectors compatible with cypros terminology}
\usage{
adjust_well_roi_vec(well_roi_vec)
}
\description{
Function to make well roi vectors compatible with cypros terminology
}
