% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{adjust_well_vec}
\alias{adjust_well_vec}
\title{Function to make well vectors compatible with cypros terminology}
\usage{
adjust_well_vec(well_vec)
}
\description{
Function to make well vectors compatible with cypros terminology
}
