% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-input-options.R
\name{adjustableDefaultInstructions}
\alias{adjustableDefaultInstructions}
\title{Valid default arguments}
\usage{
adjustableDefaultInstructions()
}
\value{
Character vector.
}
\description{
Obtain a character vector of argument names whose
default input you can adjust via \code{adjustDefaultInstructions()}.
}
