% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-instructions.R
\name{adjustDefaultInstructions}
\alias{adjustDefaultInstructions}
\title{Adjust object based default}
\usage{
adjustDefaultInstructions(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{The argument names with their new default value specified with the
following syntax 'argument_name' = 'argument_default'.

Use \code{adjustableDefaultInstructions()} to obtain all argument names
whose default input you can manipulate this way.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
The default input for many recurring arguments is carried with
the cypro object. Use this function to savely adjust the default
according to your preference.
}
