% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjustWellInfo}
\alias{adjustWellInfo}
\alias{adjustWellRoiInfo}
\alias{adjustRoiInfo}
\title{Adjust well naming conventions}
\usage{
adjustWellInfo(vec)

adjustWellRoiInfo(vec)

adjustRoiInfo(vec)
}
\arguments{
\item{wells}{Character vector of wells.}

\item{well_rois}{Character vector of well-rois.}
}
\description{
Converts all values of the provided vectors to
the proper naming convention.
}
\details{
The naming convention that the \code{cypro}-package
follows is \emph{A1, A2, A3, ... A10, A11, ..., B1, B2} \bold{not}
\emph{A01, A02, A03, ..., A10, A11}. The inherent order of
the wells is ensured by storing well and well ROIs as factors
in the \code{Cypro} object with corresponding levels.
}
\examples{

well_vec <- c('B01', 'B02', 'B3')

adjustWellInfo(vec = well_vec)

well_rois_vec <- c('B1_1', 'B2_1', 'B3_01')

adjustWellRoiInfo(vec = well_rois_vec)

}
\seealso{
\code{getWellLevels()}
}
