% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R,
%   R/clustering-hierarchical.R
\name{agglomerateHierarchicalTrees}
\alias{agglomerateHierarchicalTrees}
\alias{agglomerateHierarchicalTrees,Cypro-method}
\alias{agglomerateHierarchicalTrees,CyproTimeLapseMP-method}
\title{Agglomerate hierarchical trees}
\usage{
agglomerateHierarchicalTrees(object, ...)

\S4method{agglomerateHierarchicalTrees}{Cypro}(
  object,
  fset = "all_features",
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  verbose = TRUE
)

\S4method{agglomerateHierarchicalTrees}{CyproTimeLapseMP}(
  object,
  phase,
  fset = "all_features",
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{stats::hclust()}.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{methods_dist}{Character vector. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{methods_aggl}{Character vector. Specifies the agglomerative
methods of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
The input object.
}
\description{
Agglomerates hierarchical trees from a list
of distance matrices.
}
