% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{AnalysisModuleTimeLapse-class}
\alias{AnalysisModuleTimeLapse-class}
\alias{AnalysisModuleTimeLapse}
\title{The AnalysisModuleTimeLapse Class}
\description{
Subclass of S4-Class \code{AnalysisModule}. Specific class for modules
that can only be used in time lapse experiments.
}
\section{Slots}{

\describe{
\item{\code{variables_summary}}{list. Named list of summary variables represented by the
S4-class \code{SummaryVariable}. Names of list correspond to the respective
slot @name_in_cypro. Variable must be ordered according to the order in which they
are supposed to be summarized. E.g. as the migration efficiency (\emph{mgr_eff}) requires
the variable total distance (\emph{total_dist}) the total distance must be listed
before migration efficiency.}
}}

