% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animateWellImage}
\alias{animateWellImage}
\title{Animate cellular movement}
\usage{
animateWellImage(
  object,
  well_plate,
  well,
  image,
  phase = "all",
  style = "normal",
  add_on_list = list(),
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}

\item{well}{Character value. The well of interest (e.g. \emph{'A1'}, \emph{'B12'})}

\item{image}{Numeric value. The well-image of interest.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{style}{Character value. If set to \emph{'cell_tracker'} imitates the tracking process of the CellTracker software.}

\item{add_on_list}{A list of ggplot2-add-ons that are supposed to be integrated in
the visualization process.}

\item{...}{Additional arguments given to \code{gganimate::animate()}.}
}
\value{
An animated GIF.
}
\description{
This function animates the movement of cells that were covered
within one well-image.
}
