% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{argument_dummy}
\alias{argument_dummy}
\title{argument_dummy}
\usage{
argument_dummy(object, phase, pt_alpha, pt_clr, pt_clrp, pt_clrsp, pt_size)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector, list of character vectors or NULL (see details for
more). Specifies the particular groups of interest the grouping variable specified in argument
\code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{add_on_list}{A list of ggplot2-add-ons that are supposed to be integrated in
the visualization process.}

\item{cell_ids}{Character vector. Denotes the cell ids of interest.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{cypro_nc}{Logical value. Denotes which regular expression to use and thus, the valid
naming conventions. If TRUE, only the naming convention of the \code{cypro}-package is considered valid.}

\item{discrete_feature}{Character value. Specifies the name of the grouping variable
of interest. Use \code{getGroupingOptions()} to obtain all valid input options.}

\item{display_cols}{Logical value. If set to TRUE columns are used additionally
to display the results.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{display_legend}{Logical value. If set to TRUE a legend is displayed.}

\item{display_line}{Logical value. If set to TRUE a line is used additionally
to display the results.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{display_smooth}{Logical value. If set to TRUE a smoothed line is displayed to emphasize
the overall trend of the data. Use the \code{smooth_*} arguments to additionally adjust they way
the trend is plotted.}

\item{display_title}{Logical value. If set to TRUE an informative title is displayed.}

\item{features}{Character vector. Denotes the features of interest. \emph{Features} are numeric
variables in the cell data (slot @cdata) of the \code{Cypro} object.

Use \code{getFeatureNames()} to obtain all valid input options.}

\item{stat_features}{Character vector. Denotes the stat features of interest. \emph{Stat features}
are the summarized features by cell ID in the statistics data.frame of slot @cdata.

Use \code{getStatFeatureNames()} to obtain all valid input options.}

\item{feature_set}{Character value. The name of the feature set of interest.}

\item{force}{Logical value. Needs to be set to TRUE if you want to overwrite an already existing
set up or already existing results.}

\item{grouping_variable}{Character value. Denotes the categorical variable - the grouping of cells -
of interest. Use \code{getGroupingVariableNames()} to obtain all valid input options.}

\item{image}{Numeric value. The well-image of interest.}

\item{info_status}{Character vector. Valid options are \emph{'Complete', 'Incomplete', 'Missing'}. Refers
to the information status of a well regarding cell and condition.}

\item{linesize}{Numeric value. Denotes the size of the lines drawn.}

\item{linetype}{Character value. Valid options are \emph{'solid', 'twodash', 'longdash', 'dotted'}
and \emph{'dotdash'}.}

\item{method_aggl}{Character vector (or value see details for more.) Denotes the agglomeration
method(s) of interest according to which the existing distance matrices are agglomerated to hierarchical
trees.

Use \code{validAgglomerationMethods()} to obtain all valid input options.}

\item{method_corr}{Character value. Denotes the correlation method of interest. Either \code{'pearson'}
or \code{'spearman'}.}

\item{method_dist}{Character vector (or value see details for more.) Denotes the distance method(s)
of interest (e.g. \emph{'euclidean'} or \emph{'manhattan'}).

Use \code{validDistanceMethods()} to obtain all valid input options.}

\item{method_kmeans}{Character vector (or value see details for more.) Denotes the algorithms of interest. Defaults
to \emph{'Hartigan-Wong'}.

Use \code{validKmeansMethods()} to obtain all valid input options.}

\item{method_outlier}{Character vector. Specifies the method/algorithm of interest.

Use \code{validOutlierDetectionMethods()} to obtain all valid input options.}

\item{method_pam}{Character vector (or value see details for more.) Denotes the algorithm of interest.
Valid input options are \emph{'euclidean'} and \emph{'manhattan'}.}

\item{module_name}{Character value. The name of the analysis module. Use \code{validModuleNames()} or
\code{getModuleNames()} to obtain valid input options.}

\item{n_cells}{Numeric calue. Determines the number of cells that are randomly chosen from
every group to be displayed. Useful to keep plots insightful and aesthetically pleasing.}

\item{overwrite}{Logical value. Must be set to TRUE in case of overlapping
variable names.

(Note that overwriting stat variables leads to all analysis progress of affected
variable sets being discarded.)}

\item{pretty_names}{Logical. If set to TRUE the function attempts to convert the concisely named
variables into more aesthetically pleasing ones.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the choice of color.}

\item{scales, space, ncol, nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}

\item{smooth}{Logical. If set to TRUE the values are smoothed.}

\item{smooth_clr}{Character value. Denotes the color of the smoothed line.}

\item{smooth_method}{Character value. Denotes the model type used to display the line. Defaults to
\emph{'lm'} (linear model). Given to argument \code{method} of function \code{ggplot2::geom_smooth()}.}

\item{smooth_se}{Logical. If set to TRUE the standard error will be displayed.}

\item{smooth_size}{Numeric value. Denotes the size of the smoothed line.}

\item{smooth_span}{NUmeric value. Denotes the smoothing span used.}

\item{stop_if_false}{Logical value. If set to TRUE and the test evaluates to FALSE the
function is stopped via \code{stop()} printing an informative message. If FALSE the
resulting logical value is returned.}

\item{well}{Character value. The well of interest (e.g. \emph{'A1'}, \emph{'B12'})}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}

\item{well_plates}{Character vector. The names of the well plates of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}

\item{with_cluster}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{with_meta}{Logical. If set to TRUE the discrete variables of the cluster data slot are added
to the output data.frame.}

\item{with_stats}{Logical. If set to TRUE the numeric variables of the stat data slots are added
tot the output data.frame.}

\item{with_cluster, with_meta, with_well_plate}{Logical values. Denoting
if the respective grouping information should be joined to the output data.frame
or not.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{vec}{A vector.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
argument_dummy
}
