% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_.R
\name{as_cluster_df}
\alias{as_cluster_df}
\alias{as_cluster_df.cypro_df}
\title{Coerce to \code{cluster_df} data.frame}
\usage{
as_cluster_df(df)

\method{as_cluster_df}{cypro_df}(df)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}
}
\value{
The input data.frame with adjusted attributes and - if necessary -
adjusted variables.
}
\description{
Coerces input data.frame to \code{cluster_df} data.frame.
}
\details{
Numeric variables of type \emph{double} are silently dropped.
Apart from variable \emph{cell_id} all non-\emph{double} variables
are coerced to factors.
}
\seealso{
\code{as_cypro_df()}, \code{as_cluster_df()}, \code{as_stats_df()}, \code{as_tracks_df()}
}
