% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_.R
\name{as_cypro_df}
\alias{as_cypro_df}
\alias{as_cypro_df.data.frame}
\alias{asCyproDf}
\alias{asCyproDf.data.frame}
\title{Coerce to \code{cypro_df} data.frame}
\usage{
as_cypro_df(df)

\method{as_cypro_df}{data.frame}(df)

asCyproDf(df)

\method{asCyproDf}{data.frame}(df)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}
}
\value{
The input data.frame with adjusted attributes and - if necessary -
adjusted variables.
}
\description{
Coerces input data.frame to \code{cypro_df} data.frame.
}
\details{
Variable \emph{cell_id} is coerced to class \code{character}.
}
\seealso{
\code{as_cypro_df()}, \code{as_cluster_df()}, \code{as_stats_df()}, \code{as_tracks_df()}
}
