% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_.R
\name{as_df_dummy}
\alias{as_df_dummy}
\title{Dummy}
\usage{
as_df_dummy(df)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}
}
\value{
The input data.frame with adjusted attributes and - if necessary -
adjusted variables.
}
\description{
Dummy
}
\seealso{
\code{as_cypro_df()}, \code{as_cluster_df()}, \code{as_stats_df()}, \code{as_tracks_df()}
}
