% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_.R
\name{as_layout_df}
\alias{as_layout_df}
\alias{as_layout_df_mp}
\alias{as_layout_df.data.frame}
\alias{as_layout_df_mp.data.frame}
\title{Coerce to \code{layout_df} data.frames}
\usage{
as_layout_df(df, well_plate_type)

as_layout_df_mp(df, well_plate_type, n_phases)

\method{as_layout_df}{data.frame}(df, well_plate_type)

\method{as_layout_df_mp}{data.frame}(df, well_plate_type = NULL, n_phases)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
The input data.frame with adjusted attributes.
}
\description{
Converts input data.frame to layout data.frames.
}
