% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_.R
\name{as_stats_df}
\alias{as_stats_df}
\alias{as_stats_df.cypro_df}
\alias{asStatsDf}
\alias{asStatsDf.cypro_df}
\title{Coerce to \code{stats_df}.}
\usage{
as_stats_df(df)

\method{as_stats_df}{cypro_df}(df, completed = FALSE)

asStatsDf(df)

\method{asStatsDf}{cypro_df}(df, completed = FALSE)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}
}
\value{
The input data.frame with adjusted attributes and - if necessary -
adjusted variables.
}
\description{
Coerces input data.frame to \code{stats_df} data.frame.
}
\seealso{
\code{as_cypro_df()}, \code{as_cluster_df()}, \code{as_stats_df()}, \code{as_tracks_df()}
}
