% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_.R
\name{as_well_plate_df}
\alias{as_well_plate_df}
\alias{as_well_plate_df.cypro_df}
\alias{asWellPlateDf}
\alias{asWellPlateDf.cypro_df}
\title{Coerce to \code{well_plate_df} data.frame}
\usage{
as_well_plate_df(df)

\method{as_well_plate_df}{cypro_df}(df)

asWellPlateDf(df)

\method{asWellPlateDf}{cypro_df}(df)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}
}
\value{
The input data.frame with adjusted attributes and - if necessary -
adjusted variables classes.
}
\description{
Coerces input data.frame to \code{well_plate_df} data.frame.
}
