% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{assemble_additional_data_variables_module_info_shiny}
\alias{assemble_additional_data_variables_module_info_shiny}
\title{Title}
\usage{
assemble_additional_data_variables_module_info_shiny(input_list, object)
}
\arguments{
\item{input_list}{shiny server input list.}

\item{object}{the cypro object of the session}

\item{example_df}{the example data file uploaded}
}
\value{
the list describing the 'additional data variables' module. it contains
information about the names the additional variables carry in
the example data.frame as well as whether they are grouping or numeric
variables
}
\description{
Title
}
