% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{assemble_analysis_module_info_shiny}
\alias{assemble_analysis_module_info_shiny}
\title{Title}
\usage{
assemble_analysis_module_info_shiny(input_list, object)
}
\arguments{
\item{input_list}{shiny server input list.}

\item{object}{the cypro object of the session}

\item{example_df}{the example data file uploaded}
}
\value{
the list \code{cypro_modules} subsetted by modules to be used. It
contains information about how the variables needed are named in the
example data.frame as well as which of the computable variables
have to be computed or are already part of the data to be loaded.
}
\description{
Title
}
