% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{assemble_id_module_info_shiny}
\alias{assemble_id_module_info_shiny}
\title{Title}
\usage{
assemble_id_module_info_shiny(
  input_list,
  object,
  example_df = NULL,
  ed_vars = NULL
)
}
\arguments{
\item{input_list}{shiny server input list.}

\item{object}{the cypro object of the session}

\item{example_df}{the example data file uploaded}
}
\value{
the list describing the experiment-type depending identification module containing
information about the names the identification variables carry in
the example data.frame
}
\description{
Title
}
