% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{assemble_track_list_shiny}
\alias{assemble_track_list_shiny}
\title{Integrate cell track tables}
\usage{
assemble_track_list_shiny(track_data_list, well_plate_list, object)
}
\arguments{
\item{track_data_list}{The output of \code{load_track_files_shiny()}.}
}
\value{
A list of data.frames named by
}
\description{
Subsets the successfully loaded files. In case of timelapse
experiments with multiple phases it sorts the data into multiple
data.frames and returns a list of multiple data.frames named according
to the ordinal number of the phase ("first", "second" etc.). In case of
non timelapse experimernts the list contains only one data.frame in a
slot named "first".
}
