% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{assemble_tracks_one_time_imaging_shiny}
\alias{assemble_tracks_one_time_imaging_shiny}
\title{Integrate cell stat tables}
\usage{
assemble_tracks_one_time_imaging_shiny(stat_data_list, well_plate_list, object)
}
\arguments{
\item{track_data_list}{The output of \code{load_track_files_shiny()}.}
}
\value{
A list of one slot 'only' with the stat data.frame
}
\description{
Subsets the successfully loaded files
}
