% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignVariables.R
\name{assignVariablesManually}
\alias{assignVariablesManually}
\alias{assignVariablesManually,Cypro-method}
\title{Manual option of assigning variables}
\usage{
assignVariablesManually(object, ...)

\S4method{assignVariablesManually}{Cypro}(
  object,
  module_name,
  var_names_required = NULL,
  var_names_optional = NULL,
  var_names_computable = NULL,
  fdb_if_false = TRUE,
  fdb_fn = "warning",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{module_name}{Character value. The name of the analysis module. Use \code{validModuleNames()} or
\code{getModuleNames()} to obtain valid input options.}

\item{var_names_required, var_names_optional, var_names_computable}{Named lists of character values where the names
correspond to the slot @name_in_cypro and the character value to slot @name_in_example}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
something
}
