% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{Cdata-class}
\alias{Cdata-class}
\alias{Cdata}
\title{The Cdata Class (Cell Data)}
\description{
The parent S4-class of every \code{Cypro}-objects \code{@cdata}
slot. Contains the experiment data in form of data.frames. All data.frames
are functionally connected by their common data variable \emph{cell_id} with which
they can be joined via \code{dplyr::left_join()}.
}
\section{Slots}{

\describe{
\item{\code{well_plate}}{data.frame. Contains data variables that give
information about a cells well plate belonging: \emph{well_plate_name,
well_plate_index, well} and \emph{well_roi}.}
}}

\seealso{
CdataScreening, CdataTimeLapse, CdataTimeLapseMP
}
