% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{CdataScreening-class}
\alias{CdataScreening-class}
\alias{CdataScreening}
\title{The Cdata Class (Cell Data - Screening Experiments)}
\description{
S4 object that contains the slots needed to represent the
data obtained by screening experiments (one time imaging).
}
\section{Slots}{

\describe{
\item{\code{cluster}}{data.frame. Contains data variables that group cells
according to clustering results. These cluster variables can either
be added from cypro intern clustering via \code{add*Algorithm*ClusterVariables()}
or from cypro extern clustering via \code{addClusterVariables()}.}

\item{\code{features}}{data.frame. Contains the cellular features (quantified by e.g. CellProfiler or ImageJ)
in form of numeric data variables.}

\item{\code{meta}}{data.frame. Contains data variables that group cells according
to the experiment design: \emph{cell_line} and \emph{condition}. Additional
grouping variables can be added via \code{addMetaVariables()}.}
}}

