% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{CdataTimeLapse-class}
\alias{CdataTimeLapse-class}
\alias{CdataTimeLapse}
\title{The Cdata Class (Cell Data - Time Lapse Experiments)}
\description{
S4 object that contains the slots needed to represent the
data obtained by screening experiments (one time imaging).
}
\section{Slots}{

\describe{
\item{\code{cluster}}{data.frame. Contains data variables that group cells
according to clustering results. These cluster variables can either
be added from cypro intern clustering via \code{add*Algorithm*ClusterVariables()}
or from cypro extern clustering via \code{addClusterVariables()}.}

\item{\code{meta}}{data.frame. Contains data variables that group cells according
to the experiment design: \emph{cell_line} and \emph{condition}. Additional
grouping variables can be added via \code{addMetaVariables()}.}

\item{\code{features_stats}}{data.frame. Contains data variables that represent
summarized, cellular profiles of the time lapse data of slot @features_tracks.}

\item{\code{features_tracks}}{data.frame. Contains the cellular features (quantified by e.g. CellProfiler or ImageJ)
in form of numeric data variables. In addition to the \emph{cell_id}-variable three
variables (\emph{frame_num, frame_time} and \emph{frame_itvl} indicate the point of time during which
the observation was made. The variable \emph{frame_added} indicates whether the
observation was actually missing due to imaging artifacts and was added artificially by imputation.}

\item{\code{scaled}}{data.frame. Contains scaled data. In case of time lapse experiments it's
the data from slot @features_stats.}
}}

