% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{CdataTimeLapseMP-class}
\alias{CdataTimeLapseMP-class}
\alias{CdataTimeLapseMP}
\title{Cypro Cell Data - Time Lapse Experiments - Multiple Phases}
\description{
S4 object that contains the slots needed to represent the
data obtained by screening experiments (multiple phase time lapse experiments).
All slots are lists named according to the ordinal number of the phase (\emph{first,
second, third...}).
}
\section{Slots}{

\describe{
\item{\code{cluster}}{list. Contains data.frames as described in slot @cluster of
S4-class \code{CdataTimeLapse} for every phase of the experiment.}

\item{\code{meta}}{list. Contains data.frames as described in slot @meta of S4-class
\code{CdataTimeLapse} for every phase of the experiment.}

\item{\code{features_stats}}{list. Contains data.frames as described in slot @features_stats
of S4-class \code{CdataTimeLapse} for every phase of the experiment.}

\item{\code{features_tracks}}{list. Contains data.frames as described in slot @features_tracks
of S4-class \code{CdataTimeLapse} for every phase of the experiment.}
}}

