% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_availability}
\alias{check_availability}
\title{Check content availability}
\usage{
check_availability(evaluate, phase, ref_input, ref_fun)
}
\arguments{
\item{evaluate}{The test that must be TRUE.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{ref_input}{glue reference}

\item{ref_fun}{glue reference}
}
\description{
Check content availability
}
