% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_nrow}
\alias{check_nrow}
\title{Detect poorly added variables}
\usage{
check_nrow(df, n_rows, ref)
}
\description{
Makes sure that the data.frame that is added to the cypro object
after a call to add*Variables() contains as many rows as there are cells
in the cypro object.
}
