% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_phase}
\alias{check_phase}
\title{Check phase input}
\usage{
check_phase(object, phase, max_phases = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{max_phase}{Numeric value or NULL. If numeric it regulates the maximal number of phases allowed.}
}
\description{
Check phase input
}
