% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_phase_manually}
\alias{check_phase_manually}
\title{Check that phase is specified manually}
\usage{
check_phase_manually(object, phase = NULL)
}
\description{
Check that phase is specified manually
}
