% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_renamed_variables}
\alias{check_renamed_variables}
\title{Detect protected variables}
\usage{
check_renamed_variables(cnames)
}
\description{
Makes sure that renaming variables does not conflict with
protected variable names,
}
