% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designExperiment-shiny-module-server.R
\name{check_set_up_progress}
\alias{check_set_up_progress}
\title{Title}
\usage{
check_set_up_progress(
  progress_input,
  subset = c("overall_information", "imaging_set_up", "experiment_phases"),
  proceed_with = "the well plate set up"
)
}
\description{
Make sure that information input is sufficient before proceeding with well plates.
}
