% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_track_df}
\alias{check_track_df}
\title{Check track data.frame}
\usage{
check_track_df(track_df)
}
\arguments{
\item{track_df}{A data.frame in which each observation refers to a cell at a given frame and
that contains the following variables:

\describe{
\item{\emph{cell_id}}{Character. Refers to the cell id's.}
\item{\emph{x_coords}}{Numeric. Refers to the cell's x-coordinates.}
\item{\emph{y_coords}}{Numeric. Refers to the cell's y-coordinates.}
\item{\emph{dfo}}{Numeric. The distances from the first position tracked. (Distance from origin).}
\item{\emph{dfpl}}{Numeric. The distances from the last point.}
\item{\emph{speed}}{Numeric. Refers to the instantaneuous speed.}
\item{\emph{afo}}{Numeric. The angle from the position of origin.}
\item{\emph{aflp}}{Numeric. The angle from the last position.}
\item{\emph{frame_num}}{Numeric. The frame number the observation refers to.}
\item{\emph{frame_time}}{Numeric. The frame number multiplied with the interval between two frames}
\item{\emph{frame_itvl}}{Character. The frame time combined with the intervals unit}
}}
}
\description{
Check track data.frame
}
