% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_var_angle}
\alias{check_var_angle}
\alias{check_var_cell_id}
\title{Variable specific content checks}
\usage{
check_var_angle(df, name_in_cypro, object = NULL)

check_var_cell_id(df, name_in_cypro, object = NULL)
}
\arguments{
\item{df}{The input data.frame.}

\item{name_in_cypro}{Character value. The name of the data variable
as \code{cypro} knows it.}

\item{object}{The \code{Cypro} object.}
}
\description{
These functions make sure that the content of the data.frame
variables assigned to variables known by \code{cypro}. See details for more.
}
\details{
These functions are added to the @slot check_content of each
\code{DataVariable} object. They are called mainly during \code{validateInputDf()}.
See it's documentation for more information of how it works.

Every \code{check_var_*()} function must take the same arguments: \code{df},
\code{name_in_cypro} and \code{object} in that order. It must do the following:
\enumerate{
\item Extract the assigned variable from the data.frame.
\item Make sure that the variable fits the basic requirements regarding class / type. If it
does not fit the class and type requirements but is convertible the function converts it.
\item Beyond class/type requirements the function checks the content for variable required
specifics. E.g. the content of \code{var_well} can not be a collection of any strings but has to
meet the proper well naming conventions.
}

Output: The output is one of the following two options:

a) If the variable passed everything (or could be adjusted to be valid) and is valid
the \code{check_var_()*}-function only returns the variable valid variable (not the data.frame.)
b) If anything did not work out or the output variable is invalid the function
returns an informative message about what went wrong in form of a character value.
}
\section{Function details}{

Constructs and outputs a preliminary cell id from all assigned optional identifier
variables to make sure that the variables identify cells uniquely across all regions
of interest if the example data.frame loaded contains data from more than one ROI.
}

