% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_wp_df_shiny}
\alias{check_wp_df_shiny}
\title{Check well plate data.frame (shiny)}
\usage{
check_wp_df_shiny(wp_df)
}
\value{
A named list containing the argument input for \code{checkpoint()}.
}
\description{
Makes sure that the well plate data.frame's information
status is sufficient to proceed with data loading.
}
