% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family.R
\name{check_wp_name}
\alias{check_wp_name}
\title{Check well plate name}
\usage{
check_wp_name(object, well_plate)
}
\arguments{
\item{object}{A valid cypro object.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\description{
Check well plate name
}
