% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete_tracks}
\alias{complete_tracks}
\title{Completes track data.frame:}
\usage{
complete_tracks(df, phase, object, verbose)
}
\description{
\itemize{
\item inserts NA observations for every cell id/frame combination that is missing
\item adds frame_itvl and frame_time variables
\item computes all track variables that derive from modules
}
}
\details{
outputs a data.frame that contains variables cell id, frame variables and
all additional numeric variables found

to be used as input for .f in purrr::map()
}
