% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{completeTracksDf}
\alias{completeTracksDf}
\alias{completeTracksDf,CyproTimeLapse-method}
\title{Complete tracks data.frame}
\usage{
completeTracksDf(object, verbose = TRUE)

\S4method{completeTracksDf}{CyproTimeLapse}(object, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Completes the track data.frame by adding missing \emph{cell_id} +
\emph{frame} observations.
}
