% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_distances_from_origin}
\alias{compute_distances_from_origin}
\alias{compute_distances_from_last_point}
\title{Compute the distances from origin}
\usage{
compute_distances_from_origin(x_coords, y_coords, object)

compute_distances_from_last_point(x_coords, y_coords)
}
\arguments{
\item{x_coords}{Numeric vector. Refers to x-coordinates of object.}

\item{y_coords}{Numeric vector. Refers to y-coordinates of object.}
}
\value{
A numeric vector that corresponds to the respective distances from
the first position.
}
\description{
The first values of both input-vectors are taken as
the origin position. For each subsequent position the respective
distance to the position of origin is computed.

To be used in \code{dplyr::mutate()} to recalculate the distances
of origin if the experiment includes time displaced treatment.
}
