% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-family.R
\name{compute_migration_efficiency}
\alias{compute_migration_efficiency}
\title{Calculate migration efficiency}
\usage{
compute_migration_efficiency(x_coords, y_coords, actual_distance)
}
\arguments{
\item{x_coords}{Numeric vector. Refers to x-coordinates of object.}

\item{y_coords}{Numeric vector. Refers to y-coordinates of object.}

\item{actual_distance}{Numeric vector/value. Refers to the total distance
the object of interested has traveled.}
}
\value{
A numeric value assessing how efficient the object of interested has
migrated.
}
\description{
Calculate migration efficiency
}
