% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-family.R
\name{compute_n_missing_values}
\alias{compute_n_missing_values}
\title{Compute n missing values}
\usage{
compute_n_missing_values(object, phase = NULL, verbose = NULL)
}
\description{
Counts the number of missing values a cell id has for all
track variables.
}
