% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-kmeans.R
\name{computeClusteringKmeans}
\alias{computeClusteringKmeans}
\alias{computeClusteringKmeans,Cypro-method}
\alias{computeClusteringKmeans,CyproTimeLapseMP-method}
\title{Compute cluster with kmeans}
\usage{
computeClusteringKmeans(object, ...)

\S4method{computeClusteringKmeans}{Cypro}(
  object,
  ks,
  methods_kmeans = "Hartigan-Wong",
  fset = "all_features",
  verbose = TRUE
)

\S4method{computeClusteringKmeans}{CyproTimeLapseMP}(
  object,
  phase,
  ks,
  methods_kmeans = "Hartigan-Wong",
  fset = "all_features",
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{methods_kmeans}{Character vector. Specifies the kmeans methods
of interest. Use \code{validMethodsKmeans()} to obtain all valid input options.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
The input object.
}
\description{
Computes cluster with method \emph{kmeans}.
}
