% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-pam.R
\name{computeClusteringPam}
\alias{computeClusteringPam}
\alias{computeClusteringPam,Cypro-method}
\alias{computeClusteringPam,CyproTimeLapseMP-method}
\title{Compute cluster with pam}
\usage{
computeClusteringPam(object, ...)

\S4method{computeClusteringPam}{Cypro}(
  object,
  ks,
  methods_pam = "euclidean",
  fset = "all_features",
  verbose = TRUE
)

\S4method{computeClusteringPam}{CyproTimeLapseMP}(
  object,
  phase,
  ks,
  methods_pam = "euclidean",
  fset = "all_features",
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{methods_pam}{Character value. Specifies the pam-method of interest.
Use \code{validMethodsPam()} to obtain all valid input options.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
The input object.
}
\description{
Computes cluster with method \emph{pam}.
}
