% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-hierarchical.R
\name{computeDistanceMatrices}
\alias{computeDistanceMatrices}
\title{Compute distance matrices}
\usage{
computeDistanceMatrices(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Calculates distance matrices for every valid input value of argument \code{method_dist} and stores
the results in the celltracer object. Requires that \code{initiateHierarchicalClustering()} has been called.
}
\details{
\code{computeDistanceMatrices()} is the second step in the convenient hierarchical clustering
of celltracer. It calculates the distance matrices according to all methods you are interested in in a
loop. Argument \code{method_dist} therefore either takes a value or a vector of valid character strings.
Each input value is given to argument \code{method} of function \code{stats::dist()}.

Based on the distance matrices computed and saved this way the function \code{agglomerateHierarchicalCluster()} can be used as the third
step of the hierarchical clustering pipeline.
}
