% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-kmeans.R
\name{computeKmeansCluster}
\alias{computeKmeansCluster}
\title{Compute cluster with kmeans}
\usage{
computeKmeansCluster(
  object,
  variable_set,
  k,
  phase = NULL,
  method_kmeans = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{k}{Numeric vector. For every value the kmeans cluster assignment is calculated via
\code{stats::kmeans()}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{method_kmeans}{Character vector (or value see details for more.) Denotes the algorithms of interest. Defaults
to \emph{'Hartigan-Wong'}.

Use \code{validKmeansMethods()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Performs kmeans clustering for every combination of \code{method_kmeans} and
\code{k} and saves the results in the cypro object.
}
\details{
As this function iterates over all valid combinations of \code{k} and \code{method_kmeans}
both inputs can be specified as vectors.
}
