% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{computeModuleVariables}
\alias{computeModuleVariables}
\alias{computeModuleVariables,CyproScreening-method}
\alias{computeModuleVariables,CyproTimeLapse-method}
\title{Compute module variables}
\usage{
computeModuleVariables(object, ...)

\S4method{computeModuleVariables}{CyproScreening}(object, verbose = TRUE)

\S4method{computeModuleVariables}{CyproTimeLapse}(object, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Computes the computable variables of all active analysis modules
in the \code{Cypro} object and adds them to \code{Cypro} objects data.
}
