% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-pam.R
\name{computePamCluster}
\alias{computePamCluster}
\title{Compute cluster with partitioning around medoids (PAM)}
\usage{
computePamCluster(
  object,
  variable_set,
  k,
  phase = NULL,
  method_pam = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{k}{Numeric vector. Denotes the numbers of clusters the pam-algorithm
is supposed to assign return. Values must be bigger than 2.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{method_pam}{Character vector (or value see details for more.) Denotes the algorithm of interest.
Valid input options are \emph{'euclidean'} and \emph{'manhattan'}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{cluster::pam()}.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Performs partitioning around medoids for every combination of
\code{method_pam} and \code{k} and saves the results in the cypro object.
}
\details{
As this function iterates over all valid combinations of \code{method_pam}
and \code{k} both inputs can be specified as vectors.
}
