% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/dimensional-reduction.R
\name{computePCA}
\alias{computePCA}
\alias{computePCA,Cypro-method}
\alias{computePCA,CyproTimeLapseMP-method}
\title{Compute PCA}
\usage{
computePCA(object, ...)

\S4method{computePCA}{Cypro}(object, n_dims, fset = "all_features", verbose = TRUE, ...)

\S4method{computePCA}{CyproTimeLapseMP}(object, phase, n_dims, fset = "all_features", verbose = TRUE, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{irlba::prcomp_irlba()}.}

\item{n_dims}{Numeric value. Integer that indicates the number of
principal components to compute. Must be lower than the number
of numeric variables based on which principal components are
computed.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\description{
Computes principal components and stores the embedding
in form of a data.frame.
}
