% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/dimensional-reduction.R
\name{computeTSNE}
\alias{computeTSNE}
\alias{computeTSNE,Cypro-method}
\alias{computeTSNE,CyproTimeLapseMP-method}
\title{Compute TSNE}
\usage{
computeTSNE(object, ...)

\S4method{computeTSNE}{Cypro}(object, n_dims, fset = "all_features", ...)

\S4method{computeTSNE}{CyproTimeLapseMP}(object, phase, n_dims, fset = "all_features", ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{tsne::tsne()}.}

\item{n_dims}{Numeric value. Integer that indicates the number of
dimensions of the resulting embedding. Must be lower than the number
of numeric variables based on which principal components are
computed. Given to argument \code{k} of function \code{tsne::tsne()}.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\description{
Computes t stochastic neighbour embedding stores the embedding
in form of a data.frame.
}
