% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/dimensional-reduction.R
\name{computeUMAP}
\alias{computeUMAP}
\alias{computeUMAP,Cypro-method}
\alias{computeUMAP,CyproTimeLapseMP-method}
\title{Compute UMAP}
\usage{
computeUMAP(object, ...)

\S4method{computeUMAP}{Cypro}(object, fset = "all_features", ...)

\S4method{computeUMAP}{CyproTimeLapseMP}(object, phase, fset = "all_features", ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{umap::umap()}.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
The input object.
}
\description{
Computes a manifold approximation and projection.
}
