% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-contains.R
\name{containsComputableVariables}
\alias{containsComputableVariables}
\alias{containsComputableVariables,AnalysisModule-method}
\alias{containsOptionalVariables}
\alias{containsOptionalVariables,AnalysisModule-method}
\title{Tests for analysis modules}
\usage{
containsComputableVariables(object)

\S4method{containsComputableVariables}{AnalysisModule}(object)

containsOptionalVariables(object)

\S4method{containsOptionalVariables}{AnalysisModule}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
TRUE or FALSE
}
\description{
A variety of tests that check for certain module specifics.
}
