% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-contains.R
\name{containsData}
\alias{containsData}
\alias{containsData,DataFile-method}
\alias{containsData,Cypro-method}
\title{Test for valid directory slot}
\usage{
containsData(object)

\S4method{containsData}{DataFile}(object)

\S4method{containsData}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the object contains valid content in slot @directory.
}
