% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-contains.R
\name{containsDirectory}
\alias{containsDirectory}
\alias{containsDirectory,DataFile-method}
\alias{containsDirectory,WellPlate-method}
\title{Test for valid directory slot}
\usage{
containsDirectory(object)

\S4method{containsDirectory}{DataFile}(object)

\S4method{containsDirectory}{WellPlate}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the object contains valid content in slot @directory.
}
