% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-contains.R
\name{containsErrors}
\alias{containsErrors}
\alias{containsErrors,DataFile-method}
\title{Test for validation errors}
\usage{
containsErrors(object)

\S4method{containsErrors}{DataFile}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the return value of \code{validateInputDf()} contained
any error messages.
}
