% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-contains.R
\name{containsModules}
\alias{containsModules}
\alias{containsModules,Cypro-method}
\title{Test for used modules}
\usage{
containsModules(object, ...)

\S4method{containsModules}{Cypro}(object, modules = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{modules}{Character vector NULL. If character vector every mentioned
module is checked for.}
}
\value{
If \code{modules} is character vector a named logical vector
of the same length. Else a single TRUE or FALSE.
}
\description{
Tests if the object contains module set ups.
}
