% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-contains.R
\name{containsStatsDf}
\alias{containsStatsDf}
\alias{containsStatsDf,CdataTimeLapse-method}
\alias{containsStatsDf,CdataTimeLapseMP-method}
\alias{containsStatsDf,Cypro-method}
\title{Test for stats data.frame}
\usage{
containsStatsDf(object, stop_if_false = FALSE)

\S4method{containsStatsDf}{CdataTimeLapse}(object, stop_if_false = FALSE)

\S4method{containsStatsDf}{CdataTimeLapseMP}(object, stop_if_false = FALSE)

\S4method{containsStatsDf}{Cypro}(object, stop_if_false = FALSE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{stop_if_false}{Logical value. If set to TRUE and the test evaluates to FALSE the
function is stopped via \code{stop()} printing an informative message. If FALSE the
resulting logical value is returned.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the respective slot is not empty.
}
