% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-contains.R
\name{containsSummaryVariables}
\alias{containsSummaryVariables}
\title{Tests for analysis modules}
\usage{
containsSummaryVariables(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
TRUE or FALSE
}
\description{
A variety of tests that check for certain module specifics.
}
