% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-contains.R
\name{containsTracksDf}
\alias{containsTracksDf}
\alias{containsTracksDf,CdataTimeLapse-method}
\alias{containsTracksDf,CdataTimeLapseMP-method}
\alias{containsTracksDf,Cypro-method}
\title{Tests for tracks data.frame}
\usage{
containsTracksDf(object, stop_if_false = FALSE)

\S4method{containsTracksDf}{CdataTimeLapse}(object, stop_if_false = FALSE)

\S4method{containsTracksDf}{CdataTimeLapseMP}(object, stop_if_false = FALSE)

\S4method{containsTracksDf}{Cypro}(object, stop_if_false = FALSE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{stop_if_false}{Logical value. If set to TRUE and the test evaluates to FALSE the
function is stopped via \code{stop()} printing an informative message. If FALSE the
resulting logical value is returned.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the respective slot is not empty.
}
