% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-analysis.R
\name{correlateAll}
\alias{correlateAll}
\alias{correlateAcross}
\title{Compute correlation between variables}
\usage{
correlateAll(
  object,
  variable_set,
  method_corr = NULL,
  phase = NULL,
  print_errors = FALSE,
  verbose = NULL
)

correlateAcross(
  object,
  variable_set,
  phase = NULL,
  across = NULL,
  method_corr = NULL,
  verbose = NULL,
  print_errors = TRUE
)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{method_corr}{Character value. Denotes the correlation method of interest. Either \code{'pearson'}
or \code{'spearman'}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{across}{Character vector or NULL. Specifies the grouping variables across which
correlation is computed. Meaning that before correlation is computed the data is split
into the groups suggested by the respective grouping variable.

If set to NULL defaults to all grouping variables obtained via \code{getGroupingVariableNames()}.}
}
\value{
updated_object return
}
\description{
Performs correlation analysis on the variables with which
the correlation analysis has been initiated via \code{initiateCorrelation()}.
}
\details{
\code{correlateAcross()} iterates over all grouping variables and computes the
correlation analysis for every group the respective grouping variables suggests. Use the
\code{across}-argument of \code{plotCorrplot()} to visualize the results of \code{correlateAcross()}.
}
