% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_well_add_on}
\alias{create_well_add_on}
\title{Create well ggplot layer}
\usage{
create_well_add_on(
  df = NULL,
  plot_type,
  alpha,
  fill_by,
  fill,
  size,
  stroke,
  color_by,
  color,
  display_border,
  border_color,
  border_size
)
}
\description{
Create well ggplot layer
}
