% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createCyproObject}
\alias{createCyproObject}
\title{Create \code{Cypro} object}
\usage{
createCyproObject(class = "CyproTimeLapse", exp_name = "test", ...)
}
\arguments{
\item{class}{Character value. The Cypro class of the object.}

\item{exp_name}{Character value. The name of the experiment.}

\item{...}{Additional arguments given to the constructor function
\code{methods::new(Class = \code{class})}.}
}
\value{
An object of the denoted class.
}
\description{
Wrapper for manual construction of the \code{Cypro} object.
}
