% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createEmptyCyproObject}
\alias{createEmptyCyproObject}
\title{Create empty Cypro object}
\usage{
createEmptyCyproObject(class = "CyproTimeLapse", exp_name = "test", ...)
}
\arguments{
\item{class}{Character value. The Cypro class of the object.}

\item{exp_name}{Character value. The name of the experiment.}

\item{...}{Additional arguments given to the constructor function \code{methods::new()}.}
}
\value{
An object of the denoted class.
}
\description{
Wrapper for manual construction of the \code{Cypro} object.
}
