% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createLayoutDf}
\alias{createLayoutDf}
\title{Create well plate layout}
\usage{
createLayoutDf(
  well_plate_name = "wp",
  well_plate_type = "8x12",
  n_rois = 1,
  n_phases = 0
)
}
\arguments{
\item{well_plate_name}{Character value. The name of the well plate.}

\item{well_plate_type}{Character value. One of \emph{'2x3, '3x4', '4x6, '6x8', '8x12'}. Denotes
overall structure of the well plate.}

\item{phases}{Numeric value. Denotes the number of phases the
experiment design includes. If not 0, The data.frame is set up
to contain an additional column that keeps track of the changes in condition
from phase to phase.}
}
\value{
A data.frame of class \code{layout_df} or \code{layout_df_mp} if
\code{phase} is numeric.
}
\description{
Creates a data.frame of class \code{layout_df} or \code{layout_df_mp}
representing the well plate layout.
}
