% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{Cypro-class}
\alias{Cypro-class}
\alias{Cypro}
\title{The Cypro Class}
\description{
The \code{Cypro} object is a representation of high-content-screening (HCS)
and time-lapse-imaging experiment for R. In both cases images from cells have been
made and the features of the captured cells have been quantified by image analyzing
software such as \emph{CellProfer, CellTracker, ImageJ}.

The \code{Cypro}-class is only the parent of the actual \code{Cypro}-classes that are
used, namely \code{CyproScreening}, \code{CyproTimeLapse} and \code{CyproTimeLapseMP}.
}
\section{Slots}{

\describe{
\item{\code{analysis}}{list. A list of three slots containing results of statistical and machine-learning
techniques used on the numeric data of slot @cdata. Currently implemented analysis pipelines
occupy the slots \emph{clustering}, \emph{correlation} and \emph{dim_red}.}

\item{\code{commands}}{list. A list of commands run on this \code{Cypro}-object.}

\item{\code{compatibility}}{list. A list that can be used as a short term deposit for anything that isn't implemented
in \code{cypro}.}

\item{\code{default}}{list. Contains the object specific default input for recurring arguments. Can
be safely modified via \code{adjustDefaultInstructions()}.}

\item{\code{directory}}{character. Directory under which the \code{Cypro}-object is stored by default
using \code{saveCyproObject()}.}

\item{\code{experiment}}{character. The name of the experiment.}

\item{\code{experiment_design}}{ExperimentDesign. An S4-class of class \code{ExperimentDesign}.}

\item{\code{feature_sets}}{list. Each slot contains a character vector of names of numeric variables forming
a \emph{feature_set}. Based on these features clustering and dimensional reduction can be
performed. This allows to store several clustering results based on different features in one and the
same \code{Cypro} object.}

\item{\code{information}}{list. Miscellaneous information around the object.}

\item{\code{modules}}{list. Each slot is again a list representing one of the modules that has been denoted
during \code{assignVariables()}. In addition to the variable denotation it can contain
module specific data, computation or analysis results.}

\item{\code{progress}}{Progress. S4-class of class \code{Progress.}}

\item{\code{quality_checks}}{list. Contains results of quality check related results such as outlier detection.}

\item{\code{subsets}}{list. Contains information of each subsetting process the \code{Cypro}-object
has gone through. See functions prefixed with \code{subsetBy*()}.}

\item{\code{version}}{list. Three slots named \emph{patch, minor} and \emph{major} that keep
track of the version of the object and the latest version of the cypro package.}
}}

\seealso{
\code{CyproScreening}-class, \code{CyproTimeLapse}-class, \code{CyproTimeLapseMP}-class
}
