% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{CyproScreening-class}
\alias{CyproScreening-class}
\alias{CyproScreening}
\title{The CyproScreening Class}
\description{
The abstraction of high content screening experiments in which cells were imaged
only one time in contrast to time lapse experiments where cells are imaged several times in defined
intervals to track them.

In addition to slot @cdata it contains all slots of class \code{Cypro}.
}
\section{Slots}{

\describe{
\item{\code{cdata}}{Object of class \code{CdataScreening}.}
}}

\seealso{
\code{Cypro}-class, \code{CdataScreening}-class
}
