% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{CyproSubset-class}
\alias{CyproSubset-class}
\alias{CyproSubset}
\title{The CyproSubset Class}
\description{
Subsetting allows to conveniently split data sets by certain
characteristics such as cell lines, conditions, cluster etc. or for specific
cell ids. All outlier removal functions of \code{cypro} base on this system
and it might be additionally useful if you want apply some machine learning
algorithms such as clustering and correlation on only a subset of cells.

The \code{CyproSubset}-class contains information about the subsetting the
\code{Cypro} object it is part of has undergone. This allows to trace
all subsetting steps back to original data set with which downstream analysis
began.

object after the subsetting.
}
\section{Slots}{

\describe{
\item{\code{new_name}}{character. The content of slot @experiment of the new \code{Cypro}
object.}

\item{\code{parent_name}}{character. The content of slot @experiment of the \code{Cypro}
object that underwent subsetting.}

\item{\code{reasoning}}{character. Character string of length one that describes in
a human readable way the reasoning behind the subsetting.}
}}

\seealso{
\code{CyproSubsetByCellID},\code{CyproSubsetByGroup}, \code{CyproSubsetByNumber}
}
