% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{CyproSubsetByCellID-class}
\alias{CyproSubsetByCellID-class}
\alias{CyproSubsetByCellID}
\title{The CyproSubsetByCellID Class}
\description{
Every subsetting in \code{Cypro} happens by cell id. The
\code{CyproSubsetByCellID} class contains information about
the cell IDs that were kept and that were discarded.
}
\details{
Inherits from class \code{CyproSubset}.
}
\section{Slots}{

\describe{
\item{\code{ids_discarded}}{character. The cell IDs that were discarded.}

\item{\code{ids_remaining}}{character. The cell IDs that remained in the \code{Cypro}}
}}

\section{Subset mechanism}{


Cells with cell IDs that were not part of the input ids are discarded.
}

