% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{CyproSubsetByGroup-class}
\alias{CyproSubsetByGroup-class}
\alias{CyproSubsetByGroup}
\title{The CyproSubsetByGroup Class}
\description{
Subsetting by group allows to only denote the groups
from a grouping variable that are supposed to be kept. The cell IDs
are then filtered and subsetting by cell ID takes place.

The class \code{CyproSubsetByGroup} contains the denoted information
that lead to the cell IDs that were used to subset the \code{Cypro}
object eventually.
}
\details{
Inherits from classes \code{CyproSubset} and \code{CyproSubsetByCellID}.
}
\section{Slots}{

\describe{
\item{\code{grouping_variable}}{character. The grouping variable that contains
the groups that are supposed to be kept.}

\item{\code{groups}}{character. The groups whose cells are supposed to be kept.}
}}

\section{Subset mechanism}{


The cell IDs that belong to the groups denoted are gathered. The object is
then subsetted by the mechanism described in the documentation for S4-class
\code{CyproSubsetByCellID}.
}

