% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{CyproSubsetByNumber-class}
\alias{CyproSubsetByNumber-class}
\alias{CyproSubsetByNumber}
\title{The CyproSubsetByNumber Class}
\description{
The class \code{CyproSubsetByGroup} contains the denoted information
that lead to the cell IDs that were used to subset the \code{Cypro}
object eventually. See section 'Subset mechanism' for more information about the options.
}
\details{
Inherits from classes \code{CyproSubset} and \code{CyproSubsetByCellID}.
}
\section{Slots}{

\describe{
\item{\code{across}}{character. The grouping variables across which
to reduce the cell number.}

\item{\code{n_by_group}}{numeric. The number of cells that are selected by
group in case of option \emph{groupwise}.}

\item{\code{n_total}}{numeric. The number of cells the \code{Cypro} object
is supposed to contain after the subsetting in case of either option \emph{total_fixed}
or option \emph{total_weighted}.}

\item{\code{weighted}}{logical. Decides if the proportion of cells in each group is
maintained (weighted = TRUE, option \emph{total_weighted}) or is equalized
(weighted = FALSE, option \emph{total_fixed}).}

\item{\code{option}}{character. The option according to which the subsettign took place.}
}}

\section{Subset mechanism}{


Subsetting by number does not subset by anything
specific but simply reduces the number of cells in the object by random
selection across certain grouping variables. There are three ways to subset by number:

\itemize{
\item{\emph{groupwise:}}{ A fixed number of cells is picked from every group. The
resulting total number of cells in the new \code{Cypro} object is the sum of groups
of the chosen grouping variable multiplied by the denoted number.}
\item{\emph{total_fixed:}}{ The number of cells in the whole data set is reduced
to a fixed number. In the new \code{Cypro} object very group of the grouping variable
across which the data is subsetted contains the same number of cells.}
\item{\emph{total_weighted:}}{ The number of cells in the whole data set is reduced
to a fixed number. In the new \code{Cypro} object the original proportion of cells in
the groups of the grouping variable across which the object is subsetted stays the same.}
}
}

