% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{CyproTimeLapse-class}
\alias{CyproTimeLapse-class}
\alias{CyproTimeLapse}
\title{The CyproTimeLapse Class}
\description{
The abstraction of time lapse imaging experiments in which cells were
imaged several times in defined intervals to track them. In contrast to the class
\code{CyproTimeLapseMP} where the suffix MP stands for multiple phases the condition
in which cells existed did not change over the course of the experiment. Meaning that
they were treated right before the imaging started.

In addition to slot @cdata it contains all slots of class \code{Cypro}.
}
\section{Slots}{

\describe{
\item{\code{cdata}}{Object of class \code{CdataTimeLapse}.}
}}

\seealso{
\code{Cypro}-class, \code{CdataTimeLapse}-class
}
