% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{CyproTimeLapseMP-class}
\alias{CyproTimeLapseMP-class}
\alias{CyproTimeLapseMP}
\title{The CyproTimeLapseMP Class}
\description{
The abstraction of time lapse imaging experiments in which cells were
imaged several times in defined intervals to track them. The suffix MP stands for multiple
phases. This describes a special kind of experiment design in which the condition of the cells
changes over time. E.g cells were treated with compound x right before the imaging started
and after 10 hours compound y was added.

In addition to slot @cdata it contains all slots of class \code{Cypro}.
}
\section{Slots}{

\describe{
\item{\code{cdata}}{Object of class \code{CdataTimeLapseMP}.}
}}

\seealso{
\code{Cypro}-class, \code{CdataTimeLapseMP}-class
}
