% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{DataFile-class}
\alias{DataFile-class}
\alias{DataFile}
\title{The DataFile Class}
\description{
S4-class that contains the information about every file loaded.
}
\section{Slots}{

\describe{
\item{\code{added}}{logical. TRUE if all variables of slot @content were valid and
the data has been integrated in the \code{Cypro} object slot @cdata.}

\item{\code{content}}{list. Every file is loaded and every variable is checked for its
content according to what has been assigned during \code{AssignVariables()}.
Slot @content contains the output of all check functions of slot @check_var.}

\item{\code{directory}}{character. The directory from which the file was loaded.}

\item{\code{valid}}{logical. TRUE if no variable of slot @content is a value of class
\code{glue} - meaning that all @check_var functions returned a valid variable
and no error/feedback message.}

\item{\code{well_plate}}{character. The well plate belonging.}
}}

