% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{DataVariable-class}
\alias{DataVariable-class}
\alias{DataVariable}
\title{The Variable Class}
\description{
The S4-class \code{DataVariable} is an object oriented approach to
solve the problem of different image analysis software following different
naming conventions.
}
\details{
The 'Variable-Ecosystem' of \code{cypro} attempts to solve the problem of different
image analysis software following different naming conventions. The reasoning behind this goes
as follows: Columns that refer to certain data variables such as x- and y-coordinates, frame number, cell perimeter,
etc. complicate writing specific code. E.g. cellular migration is
commonly visualized by rose plots (see function \code{plotRosePlot()}). This function refers to
x- and y-coordinates. If the input data, however, can refer to coordinates in different ways like
\emph{x}, \emph{x_coordinates}, \emph{x-coordinates}, the code becomes unreliable. While
unspecific data variables visualized by boxplots or density plots can follow arbitrary
naming conventions variables that are used repeatedly in \code{cypro}-intern code must follow a
consistent system to ensure flexibility regarding the data input while maintaining reliable code.
}
\section{Slots}{

\describe{
\item{\code{name_in_cypro}}{character. Short name to be used as reference throughout
\code{cypro}. Should follow underscore-naming-convention (e.g. \emph{x_coords} instead of
\emph{xCoords} or \emph{x-coords})}
}}

\seealso{
\code{AssignableVariable}, \code{RequiredVariable}, \code{ComputableVariable},
\code{SummarizableVariable}
}
