% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designExperiment.R
\name{designExperiment}
\alias{designExperiment}
\title{Object initiation: Step 1}
\usage{
designExperiment()
}
\value{
An empty cypro object.
}
\description{
Opens an interactive application in which the experiment can be set
up. This includes in particular the design of the well plates.
}
