% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{dim_red_df}
\alias{dim_red_df}
\title{dim_red_df}
\usage{
dim_red_df(dim_red_df)
}
\arguments{
\item{dim_red_df}{A data.frame that contains at least two variables:

\describe{
\item{\emph{cell_id}}{Character. The cells ids that are used to join the cluster variable unambiguously}
\item{\emph{x}}{Numeric. The cell id's position on the x-axis.}
\item{\emph{y}}{Numeric. The cell id's position on the y-axis.}
}}
}
\description{
dim_red_df
}
