% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard.R
\name{discardCellIDs}
\alias{discardCellIDs}
\alias{discardCellIDs,Cypro-method}
\title{Discard cells from object}
\usage{
discardCellIDs(object, ...)

\S4method{discardCellIDs}{Cypro}(object, cell_ids, reasoning = NA_character_, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{cell_ids}{Character vector. Denotes the cells to keep unambiguously with their cell IDs.}

\item{reasoning}{Character value. Allows two write a short description of how the cell ids according
to which the object is filtered were selected. This description is included in the output of \code{printSubsetHistory()}.
Ignored if set to NULL.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Discards the cells specified in \code{cell_ids} from the object.
}
\details{
This is done by an inverse application of \code{subsetByCellID()}. From
all valid cell IDs obtained via \code{getCellIDs()} only those are kept
that are \bold{not} in the input vector of \code{cell_ids}.
}
\seealso{
\code{subsetByCellID()}
}
