% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard.R
\name{discardDataFiles}
\alias{discardDataFiles}
\alias{discardDataFiles,WellPlate-method}
\alias{discardDataFiles,ExperimentDesign-method}
\alias{discardDataFiles,Cypro-method}
\title{Discard data files}
\usage{
discardDataFiles(object, ...)

\S4method{discardDataFiles}{WellPlate}(object)

\S4method{discardDataFiles}{ExperimentDesign}(object)

\S4method{discardDataFiles}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
The adjusted input object.
}
\description{
Discards objects of class \code{DataFile} that were
created for data loading.
}
