% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard.R
\name{discardOutliers}
\alias{discardOutliers}
\alias{discardOutliers,Cypro-method}
\title{Discard outlier IDs}
\usage{
discardOutliers(object, ...)

\S4method{discardOutliers}{Cypro}(
  object,
  method,
  features = NULL,
  across = NULL,
  across_subset = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{features}{Character vector. Denotes the features of interest. \emph{Features} are numeric
variables in the cell data (slot @cdata) of the \code{Cypro} object.

Use \code{getFeatureNames()} to obtain all valid input options.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector, list of character vectors or NULL (see details for
more). Specifies the particular groups of interest the grouping variable specified in argument
\code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Discards cell IDs that were identified as outliers.
}
