% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard-family.R
\name{discardStatVariables}
\alias{discardStatVariables}
\title{Discard unwanted variables from your data}
\usage{
discardStatVariables(object, stat_variables, phase = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{stat_variables}{Character vector. The variables from the cell statistics data you want to discard.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{cluster_variables}{Character vector. The cluster variables you want to discard.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Discard unwanted variables from your data
}
