% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard.R
\name{discardStorageDirectory}
\alias{discardStorageDirectory}
\alias{discardStorageDirectory,Cypro-method}
\title{Discard storage directory}
\usage{
discardStorageDirectory(object, ...)

\S4method{discardStorageDirectory}{Cypro}(object, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Discards the currently set storage directory
}
