% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard-family.R
\name{discardWellPlate}
\alias{discardWellPlate}
\alias{discardWellPlate,ExperimentDesign-method}
\title{Discard well plate}
\usage{
discardWellPlate(object, well_plate, ...)

\S4method{discardWellPlate}{ExperimentDesign}(object, well_plate, verbose = FALSE)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Discards the well plate from the provided object.
}
