% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-a-m.R
\name{emptyDataFiles}
\alias{emptyDataFiles}
\alias{emptyDataFiles,WellPlate-method}
\alias{emptyDataFiles,ExperimentDesign-method}
\alias{emptyDataFiles,Cypro-method}
\title{Empty data files}
\usage{
emptyDataFiles(object, ...)

\S4method{emptyDataFiles}{WellPlate}(object, transferred = TRUE)

\S4method{emptyDataFiles}{ExperimentDesign}(object, transferred = TRUE)

\S4method{emptyDataFiles}{Cypro}(object, transferred = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{transferred}{Logical value. Indicates that the emptying takes place due to
the content beeing transferred to an object of class \code{Cdata}.}
}
\value{
The adjusted input object.
}
\description{
Empties slot @content of objects of class \code{DataFile}.
}
\details{
Empties slot @content of objects of class \code{DataFile} that were
created for data loading and sets slot @transferred to TRUE indicating
that the content has been merged via \code{transferData()} and transferred
to the \code{Cypro} objects @cdata slot in form of proper \code{cypro_df} data.frames.
}
