% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-experiment-design-server.R
\name{example_selected}
\alias{example_selected}
\title{Title}
\usage{
example_selected(lst, slot = "denoted_columns", col, nth, choices = NULL)
}
\description{
Helps to generate the rendered UIs with which to denote the
x-, y-, etc. colums if the data has been generated by cell profiler.
}
