% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests.R
\name{existOutlierResults}
\alias{existOutlierResults}
\alias{isTimeLapseExp}
\alias{isUsable}
\alias{multiplePhases}
\title{Logical Tests}
\usage{
existOutlierResults(object)

isTimeLapseExp(object)

isUsable(object, module)

multiplePhases(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
TRUE or FALSE
}
\description{
Logical tests that return TRUE if the specified functions/algorithms
have been used on the object and FALSE if not.
}
