% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{ExperimentDesign-class}
\alias{ExperimentDesign-class}
\alias{ExperimentDesign}
\title{The ExperimentDesign Class}
\description{
S4-class that contains the information around the
experiment design.
}
\section{Slots}{

\describe{
\item{\code{experiment}}{character. The name of the \code{Cypro} object it is part of.}

\item{\code{example_df}}{data.frame. The example spreadsheet uploaded during \code{assignVariables()}.}

\item{\code{example_dir}}{character. The file directory of the loaded example spreadsheet.}

\item{\code{ias}}{character. The image analysis software used to quantify the images.}

\item{\code{variables_grouping}}{character. Contains variable names that were denoted as
additional grouping variables. All variables denoted like this are converted to factors.}

\item{\code{variables_numeric}}{character. Contains variable names that were denoted as
additional numeric variables. All variables denoted like this are converted to numeric.}

\item{\code{well_plates}}{list. Contains as many S4-objects of class \code{WellPlate}
as defined during the interactive session of \code{designExperiment()}.}
}}

\seealso{
\code{ExperimentDesignScreening}, \code{ExperimentDesignTimeLapse},
\code{ExperimentDesignTimeLapseMP}
}
