% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{ExperimentDesignScreening-class}
\alias{ExperimentDesignScreening-class}
\alias{ExperimentDesignSreening}
\title{The ExperimentDesignScreening Class}
\description{
S4-class that contains the information around the
experiment design in a \code{CyproScreening} object.
}
\details{
Inherits from class \code{ExperimentDesign}. Does not contain any
additional information.
}
\seealso{
\code{ExperimentDesign}, \code{ExperimentDesignTimeLapse},
\code{ExperimentDesignTimeLapseMP}
}
