% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{ExperimentDesignTimeLapse-class}
\alias{ExperimentDesignTimeLapse-class}
\alias{ExperimentDesignTimeLapse}
\title{The ExperimentDesignTimeLapse Class}
\description{
S4-class that contains the information around the
experiment design in a \code{CyproTimeLapse} object.
}
\details{
Inherits from class \code{ExperimentDesign} and contains additional
information about the imaging process.
}
\section{Slots}{

\describe{
\item{\code{n_frames}}{numeric. Only relevant in case of time lapse experiments. The total
number of images that have been made for every well-region-of-interest (well-roi).}

\item{\code{interval}}{numeric. Only relevant in case of time lapse experiments.
The interval in between which the images were made.}

\item{\code{interval_unit}}{character. Only relevant in case of time lapse experiments.
The unit of the numeric value specified in slot @interval. One of
\emph{'days', 'hours', 'minutes', 'seconds'}.}
}}

\seealso{
\code{ExperimentDesign}, \code{ExperimentDesignScreening},
\code{ExperimentDesignTimeLapseMP}
}
