% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{ExperimentDesignTimeLapseMP-class}
\alias{ExperimentDesignTimeLapseMP-class}
\alias{ExperimentDesignTimeLapseMP}
\title{The ExperimentDesignTimeLapseMP Class}
\description{
S4-class that contains the information around the
experiment design in a \code{CyproTimeLapseMP} object.
}
\details{
Inherits from class \code{ExperimentDesignTimeLapse} and contains
additional information about the phase spans.
}
\section{Slots}{

\describe{
\item{\code{phases}}{list. A list of numeric vectors.}
}}

\seealso{
\code{ExperimentDesignScreening}, \code{ExperimentDesignTimeLapse},
\code{ExperimentDesignTimeLapseMP}
}
