% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractWellInfo}
\alias{extractWellInfo}
\alias{extractRoiInfo}
\alias{extractWellRoiInfo}
\alias{extractWellRoiFileInfo}
\title{Extract well plate information}
\usage{
extractWellInfo(vec, cypro_nc = FALSE)

extractRoiInfo(vec, cypro_nc = FALSE)

extractWellRoiInfo(vec, cypro_nc = FALSE)

extractWellRoiFileInfo(vec, cypro_nc = FALSE)
}
\arguments{
\item{vec}{A vector.}

\item{cypro_nc}{Logical value. Denotes which regular expression to use and thus, the valid
naming conventions. If TRUE, only the naming convention of the \code{cypro}-package is considered valid.}
}
\value{
A character vector of the same length as the input vector.
}
\description{
Extracts information from the input vector regarding
the well plate layout by using regular expressions.
}
