% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{filterTracks}
\alias{filterTracks}
\title{Filter track data.frame}
\usage{
filterTracks(track_df, n_cells = 100, condition = NULL, cell_line = NULL)
}
\arguments{
\item{track_df}{A data.frame in which each observation refers to a cell at a given frame and
that contains the following variables:

\describe{
\item{\emph{cell_id}}{Character. Refers to the cell id's.}
\item{\emph{x_coords}}{Numeric. Refers to the cell's x-coordinates.}
\item{\emph{y_coords}}{Numeric. Refers to the cell's y-coordinates.}
\item{\emph{dfo}}{Numeric. The distances from the first position tracked. (Distance from origin).}
\item{\emph{dfpl}}{Numeric. The distances from the last point.}
\item{\emph{speed}}{Numeric. Refers to the instantaneuous speed.}
\item{\emph{afo}}{Numeric. The angle from the position of origin.}
\item{\emph{aflp}}{Numeric. The angle from the last position.}
\item{\emph{frame_num}}{Numeric. The frame number the observation refers to.}
\item{\emph{frame_time}}{Numeric. The frame number multiplied with the interval between two frames}
\item{\emph{frame_itvl}}{Character. The frame time combined with the intervals unit}
}}

\item{n_cells}{Numeric value. The number of cells the resulting data.frame
contains.}

\item{condition}{Character vector. The conditions to be kept.}

\item{cell_line}{Character vector. The cell lines to be kept.}
}
\value{
A filtered track data.frame
}
\description{
Filters a track data.frame according to the parameters
set.
}
