% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{get_dim_red_obj}
\alias{get_dim_red_obj}
\title{Title}
\usage{
get_dim_red_obj(object, variable_set, phase = NULL, method_dim_red)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_dim_red}{}
}
\value{

}
\description{
Title
}
