% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getF.R
\name{getActiveModuleNames}
\alias{getActiveModuleNames}
\alias{getActiveModuleNames,Cypro-method}
\title{Obtain names of active modules}
\usage{
getActiveModuleNames(object)

\S4method{getActiveModuleNames}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A character vector or NULL if no modules are active.
}
\description{
Extracts the names of the modules for which variables
have been assigned during \code{assignVariables()}.
}
