% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getAdditionalVariableNames}
\alias{getAdditionalVariableNames}
\alias{getAdditionalVariableNames,Cypro-method}
\title{Obtain additional variable names}
\usage{
getAdditionalVariableNames(object)

\S4method{getAdditionalVariableNames}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Character vector.
}
\description{
Extracts variable names that have been denoted during
\code{assignVariables()} as unknown to \code{cypro} but of interest
to the user.
}
