% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R
\name{getAvgSilWidthsDf}
\alias{getAvgSilWidthsDf}
\title{Obtain average sil-width data}
\usage{
getAvgSilWidthsDf(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame that contains information
about the average silhouette widths of different clustering
results with mehod \emph{pam}.
}
