% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getBatchEffectDf}
\alias{getBatchEffectDf}
\alias{getBatchEffectDist}
\title{Obtain batch effect computation results}
\usage{
getBatchEffectDf(object, reduce = TRUE, verbose = NULL)

getBatchEffectDist(object, verbose = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{reduce}{Logical value. If set to TRUE (the default) the data.frame
is reduced to the unique combinations of well plate well, else each combination
appears to times in the output data.frame.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame or an object of class \emph{dist}.
}
\description{
Returns the distances across well plate wells computed
by \code{detectBatchEffects()} either as a data.frame or as
a distance matrix.
}
