% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getCategoricalVariablesNames}
\alias{getCategoricalVariablesNames}
\alias{getNumericVariableNames}
\alias{getFrameSeq}
\alias{getFrameTimeSeq}
\alias{getFrameLevels}
\alias{getInterval}
\alias{getIntervalUnit}
\alias{getPhases}
\alias{getVariableNames}
\title{Helper functions to extract information from the}
\usage{
getCategoricalVariablesNames(object, phase = NULL)

getNumericVariableNames(object)

getFrameSeq(object, phase = NULL)

getFrameTimeSeq(object, phase = NULL)

getFrameLevels(object, phase = NULL)

getInterval(object)

getIntervalUnit(object)

getPhases(object)

getVariableNames(
  object,
  phase = NULL,
  variable_classes = c("cluster", "meta", "stats", "well_plate"),
  flatten = TRUE
)
}
\value{

}
\description{
Helper functions to extract information from the
}
