% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-extractors.R
\name{getCdata}
\alias{getCdata}
\alias{getCdata,Cypro-method}
\title{Basic extractor functions}
\usage{
getCdata(object)

\S4method{getCdata}{Cypro}(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
The respective slot content.
}
\description{
Basic extractor functions
}
