% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getClusterDf}
\alias{getClusterDf}
\alias{getClusterDf,Cypro-method}
\alias{getClusterDf,CyproTimeLapseMP-method}
\title{Obtain cluster data (by cell id)}
\usage{
getClusterDf(object, ...)

\S4method{getClusterDf}{Cypro}(object, verbose = NULL, ...)

\S4method{getClusterDf}{CyproTimeLapseMP}(object, phase = NULL, verbose = NULL, ...)
}
\arguments{
\item{object}{A valid cypro object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
This function lets you extract a data.frame that contains variables
with which cells are clustered.
}
