% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/getA-getF.R
\name{getClustering}
\alias{getClustering}
\alias{getClustering,Cypro-method}
\alias{getClustering,CyproTimeLapseMP-method}
\title{Obtain \code{Clustering} object}
\usage{
getClustering(object, ...)

\S4method{getClustering}{Cypro}(object, fset = "all_features", ...)

\S4method{getClustering}{CyproTimeLapseMP}(object, fset = "all_features", phase = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
S4-object of class \code{Clustering}.
}
\description{
Extracts an object of class \code{Clustering} containing
clustering results for all used methods so far.
}
