% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R
\name{getClusteringHclust}
\alias{getClusteringHclust}
\title{Obtain object of class \code{ClusteringHclust}}
\usage{
getClusteringHclust(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
An object of class \code{ClusteringHclust}.
}
\description{
Extracts an object of class \code{ClusteringHclust}.
}
