% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R
\name{getClusteringKmeans}
\alias{getClusteringKmeans}
\title{Obtain object of class \code{ClusteringKmeans}}
\usage{
getClusteringKmeans(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
An object of class \code{ClusteringKmeans}.
}
\description{
Extracts an object of class \code{ClusteringKmeans}.
}
