% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getF.R
\name{getClusterVariableNames}
\alias{getClusterVariableNames}
\alias{getClusterVariableNames,Cypro-method}
\alias{getClusterVariableNames,CyproTimeLapseMP-method}
\title{Extract cluster variable names of cell data}
\usage{
getClusterVariableNames(object, ...)

\S4method{getClusterVariableNames}{Cypro}(object, verbose = NULL)

\S4method{getClusterVariableNames}{CyproTimeLapseMP}(object, named = FALSE, phase = NULL, verbose = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{named}{Logical value. If set to TRUE the grouping variables are named
according to their grouping type (cluster, meta or well_plate).}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A character vector.
}
\description{
Obtain the names of the cluster variables of your cell data. Useful to
obtain vectors of variable names as input for recurring arguments like \code{across} or
\code{grouping_variable}.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
