% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R,
%   R/clustering-hierarchical.R
\name{getClusterVarsHclust}
\alias{getClusterVarsHclust}
\alias{getClusterVarsHclust,Cypro-method}
\alias{getClusterVarsHclust,CyproTimeLapseMP-method}
\title{Obtain cluster variables (hclust)}
\usage{
getClusterVarsHclust(object, ...)

\S4method{getClusterVarsHclust}{Cypro}(
  object,
  fset = "all_features",
  ks = NULL,
  hs = NULL,
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  prefix = "",
  naming_k = "{method_dist}_{method_aggl}_k{k}",
  nmaing_h = "{method_dist}_{method_aggl}_h{h}"
)

\S4method{getClusterVarsHclust}{CyproTimeLapseMP}(
  object,
  phase,
  fset = "all_features",
  ks = NULL,
  hs = NULL,
  methods_dist = "euclidean",
  methdos_aggl = "ward.D",
  prefix = "",
  naming_k = "{method_dist}_{method_aggl}_k{k}",
  nmaing_h = "{method_dist}_{method_aggl}_h{h}"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{hs}{Numeric vector. Denotes the heights at which the dendrogram is cut.}

\item{methods_dist}{Character vector. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{methods_aggl}{Character vector. Specifies the agglomerative
methods of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
Extracts a data.frame that contains grouping variables
according to clustering results.
}
