% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-kmeans.R
\name{getClusterVarsKmeans}
\alias{getClusterVarsKmeans}
\alias{getClusterVarsKmeans,Cypro-method}
\alias{getClusterVarsKmeans,CyproTimeLapseMP-method}
\title{Obtain cluster variables (kmeans)}
\usage{
getClusterVarsKmeans(object, ...)

getClusterVarsKmeans(object, ...)

\S4method{getClusterVarsKmeans}{Cypro}(
  object,
  ks,
  fset = "all_features",
  methods_kmeans = "Hartigan-Wong",
  prefix = "",
  naming = "{method_kmeans}_k{k}"
)

\S4method{getClusterVarsKmeans}{CyproTimeLapseMP}(
  object,
  phase,
  ks,
  fset = "all_features",
  methods_kmeans = "Hartigan-Wong",
  prefix = "",
  naming = "{method_kmeans}_k{k}"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{methods_kmeans}{Character vector. Specifies the kmeans methods
of interest. Use \code{validMethodsKmeans()} to obtain all valid input options.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
Extracts a data.frame that contains grouping variables
according to clustering results.
}
