% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-pam.R
\name{getClusterVarsPam}
\alias{getClusterVarsPam}
\alias{getClusterVarsPam,Cypro-method}
\alias{getClusterVarsPam,CyproTimeLapseMP-method}
\title{Obtain cluster variables (PAM)}
\usage{
getClusterVarsPam(object, ...)

\S4method{getClusterVarsPam}{Cypro}(
  object,
  ks,
  methods_pam = "euclidean",
  fset = "all_features",
  prefix = "",
  naming = "{method_pam}_k{k}"
)

\S4method{getClusterVarsPam}{CyproTimeLapseMP}(
  object,
  phase,
  ks,
  methods_pam = "euclidean",
  fset = "all_features",
  prefix = "",
  naming = "{method_pam}_k{k}"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{methods_pam}{Character value. Specifies the pam-method of interest.
Use \code{validMethodsPam()} to obtain all valid input options.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
Extracts a data.frame that contains grouping variables
according to clustering results.
}
