% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getCorrConv}
\alias{getCorrConv}
\alias{getCorrObject}
\title{Obtain cypros correlation objects}
\usage{
getCorrConv(object, feature_set, phase = NULL)

getCorrObject(object, feature_set, phase = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{feature_set}{Character value. The name of the feature set of interest.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An S4 object of class \emph{'corr_conv'}
}
\description{
Obtain cypros correlation objects
}
