% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/getA-getF.R
\name{getCorrelation}
\alias{getCorrelation}
\alias{getCorrelation,Cypro-method}
\alias{getCorrelation,CyproTimeLapseMP-method}
\title{Obtain \code{Correlation} object}
\usage{
getCorrelation(object, ...)

\S4method{getCorrelation}{Cypro}(object, fset = "all_features")

\S4method{getCorrelation}{CyproTimeLapseMP}(object, fset = "all_features", phase = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
S4-bject of class \code{Correlation}.
}
\description{
Extracts \code{Correlation} object that contains all correlation
results computed so far.
}
