% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/correlation.R
\name{getCorrMtr}
\alias{getCorrMtr}
\alias{getCorrMtr,Cypro-method}
\alias{getCorrMtr,CyproTimeLapseMP-method}
\title{Obtain correlation matrix}
\usage{
getCorrMtr(object, ...)

\S4method{getCorrMtr}{Cypro}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  type = "complete",
  diagonal = TRUE,
  flatten = TRUE
)

\S4method{getCorrMtr}{CyproTimeLapseMP}(
  object,
  phase,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  type = "complete",
  diagonal = TRUE,
  flatten = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method_corr}{Character value. Specifies the correlation
method of interest. Use \code{validMethodsCorrelation()} to obtain all valid
input options.}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{across_subset}{Character vector, list of character vectors or NULL (see details for
more). Specifies the particular groups of interest the grouping variable specified in argument
\code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{type}{Character value. Denotes how the underlying correlation matrix is
handled. Three options:

\itemize{
 \item{\emph{'complete'}:}{ The matrix stays as is.},
 \item{\emph{'lower'}:}{ The part below the diagonal is used. The upper part is set to NA.},
 \item{\emph{'upper'}:}{The part above the diagonal is used. the lower part is set to NA.}
 }}

\item{diagonal}{Logical value.  Indicates if the diagonal values of the
matrices are kept ( = TRUE) or set to NA (= FALSE). Ignored if \code{type}
= \emph{'complete'}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\description{
Extracts correlation results in form of correlation
matrices.
}
\details{
If \code{across} is a character a list of correlation matrices
is returned. Slots are named according to the groups of the denoted
grouping variable. If \code{across_subset} is NULL all groups are considered.
If \code{across_subset} is of length one and \code{flatten} is TRUE the
output list is flattened and a single correlation matrix is returned.
}
