% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getData}
\alias{getData}
\title{Obtain data slots}
\usage{
getData(object, data_slot, phase)
}
\arguments{
\item{object}{A valid cypro object.}

\item{data_slot}{Character value. One of \emph{'stats', 'tracks', 'meta'} or \emph{'cluster'}.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
The data.frame of interest.
}
\description{
A wrapper around \code{purr::map_df()} and the respective
list of the data slot of interest.
}
