% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getF.R
\name{getDataFileDf}
\alias{getDataFileDf}
\alias{getDataFileDf,DataFile-method}
\title{Obtain loaded file as data.frame}
\usage{
getDataFileDf(object)

\S4method{getDataFileDf}{DataFile}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A data.frame.
}
\description{
Extracts the loaded content of slot @content,
converts it into a data.frame and adds the missing identification
variables depending on the loading modality.
}
