% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getDefaultInstructions}
\alias{getDefaultInstructions}
\alias{getDefaultInstructions,Cypro-method}
\title{Obtain the object's default instructions}
\usage{
getDefaultInstructions(object, ...)

\S4method{getDefaultInstructions}{Cypro}(object, return = "list")
}
\arguments{
\item{object}{A valid cypro object.}

\item{return}{Character value. Either \emph{'list'} (default) or \emph{'S4'}.}
}
\value{
A list or an S4 object.
}
\description{
Convenient access to what is currently defined as the default.
}
