% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R,
%   R/clustering-hierarchical.R
\name{getDendroSegmentDf}
\alias{getDendroSegmentDf}
\alias{getDendroSegmentDf,Cypro-method}
\alias{getDendroSegmentDf,CyproTimeLapseMP-method}
\title{Obtain dendrogram segments}
\usage{
getDendroSegmentDf(object, ...)

\S4method{getDendroSegmentDf}{Cypro}(
  object,
  fset = "all_features",
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle"
)

\S4method{getDendroSegmentDf}{CyproTimeLapseMP}(
  object,
  phase,
  fset = "all_features",
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{methods_dist}{Character vector. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{methods_aggl}{Character vector. Specifies the agglomerative
methods of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{k}{Numeric value. Denotes the number of clusters. Must be bigger than 1. Must
not be bigger than the number of observations of the data set.}

\item{h}{Numeric value. Denotes the height at which the dendrogram is cut.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame that contains data to plot
a dendrogram with \code{ggplot2::geom_segment()}.
}
