% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R
\name{getDf}
\alias{getDf}
\title{Obtain object data}
\usage{
getDf(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A data.frame with subclass \code{tibble}.
}
\description{
Extracts the objects data as a data.frame.
}
