% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R,
%   R/clustering-hierarchical.R
\name{getDistMtr}
\alias{getDistMtr}
\alias{getDistMtr,Cypro-method}
\alias{getDistMtr,CyproTimeLapseMP-method}
\title{Obtain distance matrix}
\usage{
getDistMtr(object, ...)

\S4method{getDistMtr}{Cypro}(
  object,
  fset = "all_features",
  method_dist = "euclidean",
  stop_if_null = TRUE
)

\S4method{getDistMtr}{CyproTimeLapseMP}(
  object,
  fset = "all_features",
  method_dist = "euclidean",
  stop_if_null = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{method_dist}{Character value. Specifies the distance method
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{stop_if_null}{Logical value. If TRUE and the function does not find the object to
extract an informative error is raised. Else the empty value - usually NULL - is returned.}
}
\value{
A distance matrix or NULL (if none has been computed yet).
}
\description{
Extracts distance matrix from S4 object.
}
