% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/dimensional-reduction.R
\name{getEmbeddingDf}
\alias{getEmbeddingDf}
\alias{getEmbeddingDf,Cypro-method}
\alias{getEmbeddingDf,CyproTimeLapseMP-method}
\title{Obtain \code{DimRed} embedding}
\usage{
getEmbeddingDf(object, ...)

\S4method{getEmbeddingDf}{Cypro}(
  object,
  fset = "all_features",
  method_dimred = "pca",
  numeric = FALSE,
  numeric_scaled = FALSE,
  grouping = FALSE,
  logical = FALSE,
  complete = FALSE,
  shift = FALSE
)

\S4method{getEmbeddingDf}{CyproTimeLapseMP}(
  object,
  phase,
  fset = "all_features",
  method_dimred = "pca",
  numeric = FALSE,
  numeric_scaled = FALSE,
  grouping = FALSE,
  logical = FALSE,
  complete = FALSE,
  shift = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
Extracts the dimensional reduction embedding in form
of a data.frame. Naming of the variables corresponds
to the chosen method.
}
