% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getErrors}
\alias{getErrors}
\alias{getErrors,DataFile-method}
\alias{getErrors,Cypro-method}
\title{Obtain loading error messages}
\usage{
getErrors(object, ...)

\S4method{getErrors}{DataFile}(object)

\S4method{getErrors}{Cypro}(object, well_plates = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{well_plates}{Character vector. The names of the well plates of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
A list of \code{glue} objects or NULL if no errors were thrown.
}
\description{
Extracts the error messages that were thrown
during data loading.
}
