% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getF.R
\name{getExampleDf}
\alias{getExampleDf}
\alias{getExampleDf,ExperimentDesign-method}
\alias{getExampleDf,Cypro-method}
\title{Obtain example data.frame}
\usage{
getExampleDf(object)

\S4method{getExampleDf}{ExperimentDesign}(object)

\S4method{getExampleDf}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A data.frame.
}
\description{
Extracts the data.frame that has been loaded as an
input example during \code{assingVariables()}.
}
