% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getExperimentDesign}
\alias{getExperimentDesign}
\alias{getExperimentDesign,Cypro-method}
\title{Extract experiment design}
\usage{
getExperimentDesign(object, ...)

\S4method{getExperimentDesign}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
An object of class \code{ExperimentDesign}
}
\description{
Obtain the experiment design in form of an S4-class called
\code{ExperimentDesign.}
}
