% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getFeatureNames}
\alias{getFeatureNames}
\alias{getFeatureNames,CyproScreening-method}
\title{Obtain names of numeric features}
\usage{
getFeatureNames(object, ...)

\S4method{getFeatureNames}{CyproScreening}(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A character vector.
}
\description{
Extracts the names of variables stored in
slot @features of the \code{Cypro} objects cell data.
}
