% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getFeatureSet}
\alias{getFeatureSet}
\alias{getFeatureSet,Cypro-method}
\alias{getFeatureSets,Cypro-method}
\alias{getFeatureSetNames,Cypro-method}
\title{Obtain a feature set}
\usage{
getFeatureSet(object, feature_set)

\S4method{getFeatureSet}{Cypro}(object)

\S4method{getFeatureSets}{Cypro}(object)

\S4method{getFeatureSetNames}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{feature_set}{Character value. The name of the feature set of interest.}
}
\value{
A character vector.
}
\description{
Extracts the names of features that were pooled
under the name of the feature set.
}
