% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family-A-M.R
\name{getFeatureSetNames}
\alias{getFeatureSetNames}
\title{Extract names of defined sets of data variables}
\usage{
getFeatureSetNames(object)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}
}
\value{
A character vector.
}
\description{
Obtain the names off all currently defined feature sets.
}
