% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getFileDirectories}
\alias{getFileDirectories}
\alias{getFileDirectories,layout_df-method}
\alias{getFileDirectories,WellPlate-method}
\title{Obtain file directories}
\usage{
getFileDirectories(object, ...)

\S4method{getFileDirectories}{layout_df}(object, ...)

\S4method{getFileDirectories}{WellPlate}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Character vector
}
\description{
Extracts the directories that were used to load the cell data.
}
