% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getFrames}
\alias{getFrames}
\alias{getFrames,ExperimentDesign-method}
\alias{getFrames,CyproTimeLapse-method}
\alias{getFrames,CyproTimeLapseMP-method}
\title{Obtain frame range}
\usage{
getFrames(object, ...)

\S4method{getFrames}{ExperimentDesign}(object)

\S4method{getFrames}{CyproTimeLapse}(object)

\S4method{getFrames}{CyproTimeLapseMP}(object, by_phase = FALSE)
}
\arguments{
\item{object}{}
}
\value{
Numeric vector.
}
\description{
Extracts a numeric vector from one till the number of
frames that was denoted during the experiment design steps.
}
