% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getFrameTimeSeq}
\alias{getFrameTimeSeq}
\alias{getFrameTimeSeq,CyproTimeLapse-method}
\alias{getFrameTimeSeq,CyproTimeLapseMP-method}
\title{Extract frame time sequence}
\usage{
getFrameTimeSeq(object, ...)

\S4method{getFrameTimeSeq}{CyproTimeLapse}(object)

\S4method{getFrameTimeSeq}{CyproTimeLapseMP}(object, by_phase = FALSE, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A numeric vector.
}
\description{
Obtain the time line of the time lapse experiment as a numeric vector.
}
