% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGroupingDf}
\alias{getGroupingDf}
\alias{getClusterDf}
\alias{getMetaDf}
\alias{getWellPlateDf}
\title{Obtain grouping information}
\usage{
getGroupingDf(object, phase = NULL, verbose = NULL)

getClusterDf(object, phase = NULL, verbose = NULL)

getMetaDf(object, phase = NULL)

getWellPlateDf(object)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame that contains the cell ids and their group belonging.
}
\description{
These functions let you extract a data.frame that contain variables
with which cells are grouped.
}
