% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGroupingVariableNames}
\alias{getGroupingVariableNames}
\alias{getClusterVariableNames}
\alias{getMetaVariableNames}
\alias{getWellPlateVariableNames}
\title{Obtain grouping variable names of cell data}
\usage{
getGroupingVariableNames(
  object,
  ...,
  named = FALSE,
  phase = NULL,
  verbose = TRUE
)

getClusterVariableNames(object, ..., phase = NULL)

getMetaVariableNames(object, ..., phase = NULL)

getWellPlateVariableNames(object, ...)
}
\arguments{
\item{object}{A valid cypro object.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}

\item{named}{Logial value. If set to TRUE the grouping variables are named
according to their grouping type (cluster, meta or well_plate).}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A character vector.
}
\description{
Convenient access to the names of your objects data variables. Useful to
obtain vectors of variable names as input for recurring arguments like \code{across} or
\code{grouping_variable}.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
