% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/getG-getM.R
\name{getGroups}
\alias{getGroups}
\alias{getGroupNames}
\alias{getGroupNames,Cypro-method}
\alias{getGroupNames,CyproTimeLapseMP-method}
\title{Obtain group names a grouping variable contains}
\usage{
getGroups(object, option)

getGroupNames(object, ...)

\S4method{getGroupNames}{Cypro}(object, grouping_variable, ...)

\S4method{getGroupNames}{CyproTimeLapseMP}(object, grouping_variable, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{grouping_variable}{Character value. Denotes the categorical variable - the grouping of cells -
of interest. Use \code{getGroupingVariableNames()} to obtain all valid input options.}
}
\value{
A character vector.
}
\description{
Extracts the names of the groups in which a specific grouping
variable groups the cells. Useful to obtain input options for arguments like \code{across_subset}.
}
