% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R,
%   R/clustering-hierarchical.R
\name{getHclust}
\alias{getHclust}
\alias{getHclust,Cypro-method}
\title{Obtain object of class \code{hclust}}
\usage{
getHclust(object, ...)

\S4method{getHclust}{Cypro}(
  object,
  phase,
  fset = "all_features",
  method_dist = "euclidean",
  method_aggl = "ward.D",
  stop_if_null = TRUE
)

\S4method{getHclust}{Cypro}(
  object,
  phase,
  fset = "all_features",
  method_dist = "euclidean",
  method_aggl = "ward.D",
  stop_if_null = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{method_dist}{Character value. Specifies the distance method
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{method_aggl}{Character value. Specifies the agglomerative method
of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{stop_if_null}{Logical value. If TRUE and the function does not find the object to
extract an informative error is raised. Else the empty value - usually NULL - is returned.}
}
\value{
An object of class \code{hclust.}
}
\description{
Extracts object of class \code{hclust}
as described in \code{stats::hclust()}.
}
