% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getHclustConv}
\alias{getHclustConv}
\alias{getHclustObject}
\alias{getKmeansConv}
\alias{getKmeansObject}
\alias{getPamConv}
\alias{getPamObject}
\title{Obtain cypros clustering objects}
\usage{
getHclustConv(object, feature_set, phase = NULL, with_data = TRUE)

getHclustObject(object, feature_set, phase = NULL, with_data = TRUE)

getKmeansConv(object, feature_set, phase = NULL, with_data = TRUE)

getKmeansObject(object, feature_set, phase = NULL, with_data = TRUE)

getPamConv(object, feature_set, phase = NULL, with_data = TRUE)

getPamObject(object, feature_set, phase = NULL, with_data = TRUE)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{feature_set}{Character value. The name of the feature set of interest.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An S4 object of \emph{'hclust_conv'}, \emph{'kmeans_conv'} or \emph{'pam_conv'}.
}
\description{
Obtain cypros clustering objects
}
