% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-hierarchical.R
\name{getHclustObj}
\alias{getHclustObj}
\title{Obtain the original clustering objects}
\usage{
getHclustObj(
  object,
  variable_set,
  phase = NULL,
  method_dist = NULL,
  method_aggl = NULL
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{method_dist}{Character vector (or value see details for more.) Denotes the distance method(s)
of interest (e.g. \emph{'euclidean'} or \emph{'manhattan'}).

Use \code{validDistanceMethods()} to obtain all valid input options.}

\item{method_aggl}{Character vector (or value see details for more.) Denotes the agglomeration
method(s) of interest according to which the existing distance matrices are agglomerated to hierarchical
trees.

Use \code{validAgglomerationMethods()} to obtain all valid input options.}
}
\value{
The respective clustering results as their original objects.
}
\description{
These functions extract the resulting objects
of the respective algorithms calculated with the respective methods denoted
by the \code{method_*}-arguments.
}
\details{
In this case all \code{method_*}-arguments must be specified as
character values as the objects are returned one by one.
}
