% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getG-getM.R
\name{getImageDirDf}
\alias{getImageDirDf}
\alias{getImageDirDf,Cypro-method}
\title{Obtain image directories}
\usage{
getImageDirDf(object)

\S4method{getImageDirDf}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A data.frame.
}
\description{
Extracts a data.frame in which the added directories
to the original images are set by well-roi.
}
