% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getG-getM.R
\name{getInterval}
\alias{getInterval}
\alias{getInterval,ExperimentDesignTimeLapse-method}
\alias{getInterval,CyproTimeLapse-method}
\alias{getInterval,CyproTimeLapseMP-method}
\title{Obtain frame interval}
\usage{
getInterval(object)

\S4method{getInterval}{ExperimentDesignTimeLapse}(object)

\S4method{getInterval}{CyproTimeLapse}(object)

\S4method{getInterval}{CyproTimeLapseMP}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A numeric value.
}
\description{
Extracts the interval between the frames as the numeric
value.
}
