% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getIntervalUnit}
\alias{getIntervalUnit}
\alias{getIntervalUnit,ExperimentDesignTimeLapse-method}
\alias{getIntervalUnit,CyproTimeLapse-method}
\title{Obtain frame interval unit}
\usage{
getIntervalUnit(object)

\S4method{getIntervalUnit}{ExperimentDesignTimeLapse}(object)

\S4method{getIntervalUnit}{CyproTimeLapse}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A character value.
}
\description{
Extracts the unit of the interval between the frames as
a character value.
}
