% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-kmeans.R
\name{getKmeans}
\alias{getKmeans}
\alias{getKmeans,Cypro-method}
\alias{getKmeans,CyproTimeLapseMP-method}
\title{Obtain object of class \code{kmeans}}
\usage{
getKmeans(object, ...)

\S4method{getKmeans}{Cypro}(
  object,
  fset = "all_features",
  method_kmeans = "Hartigan-Wong",
  stop_if_null = TRUE
)

\S4method{getKmeans}{CyproTimeLapseMP}(
  object,
  phase,
  fset = "all_features",
  method_kmeans = "Hartigan-Wong",
  stop_if_null = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{method_kmeans}{Character value. Specifies the kmeans method
of interest. Use \code{validMethodsKmeans()} to obtain all valid input options.}

\item{stop_if_null}{Logical value. If TRUE and the function does not find the object to
extract an informative error is raised. Else the empty value - usually NULL - is returned.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An object of class \code{kmeans}.
}
\description{
Extracts an object of class \code{kmeans}
as described in \code{?stats::kmeans}.
}
