% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getG-getM.R
\name{getLayoutAttributes}
\alias{getLayoutAttributes}
\title{Obtain layout attributes}
\usage{
getLayoutAttributes(df)
}
\arguments{
\item{df}{A data.frame of class \code{layout_df}.}
}
\value{
A named list.
}
\description{
Extracts the well plate layout specific attributes
from the input data.frame.
}
