% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family-A-M.R
\name{getLayoutDf}
\alias{getLayoutDf}
\alias{getLayoutDf,WellPlate-method}
\alias{getLayoutDf,ExperimentDesign-method}
\alias{getLayoutDf,Cypro-method}
\title{Extract well plate layout data.frame}
\usage{
getLayoutDf(object, ...)

\S4method{getLayoutDf}{WellPlate}(object, ...)

\S4method{getLayoutDf}{ExperimentDesign}(object, well_plate, ...)

\S4method{getLayoutDf}{Cypro}(object, well_plate, ...)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
A \code{layout_df} - data.frame.
}
\description{
Obtain the well plate layout in form of a data.frame.
}
