% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getLayoutVariableNames}
\alias{getLayoutVariableNames}
\alias{getLayoutVariableNames,layout_df-method}
\alias{getLayoutVariableNames,Cypro-method}
\title{Obtain layout variable names}
\usage{
getLayoutVariableNames(object, named = TRUE, ...)

\S4method{getLayoutVariableNames}{layout_df}(object, named = TRUE)

\S4method{getLayoutVariableNames}{Cypro}(object, named = TRUE, well_plate = NULL, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{named}{Logical. If TRUE, the returned vector of names is named by
the class of the variable.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
Character vector.

@export
}
\description{
Extracts the variable names of the layout data.frame
as a character vector.
}
