% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getLoadingStatusDf}
\alias{getLoadingStatusDf}
\alias{getLoadingStatusDf,Cypro-method}
\title{Obtain loading status summary}
\usage{
getLoadingStatusDf(object, ...)

\S4method{getLoadingStatusDf}{Cypro}(
  object,
  well_plates = NULL,
  with_transferred = TRUE,
  n_letters = 40
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{well_plates}{Character vector. The names of the well plates of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
A data.frame.
}
\description{
Extracts a data.frame that summarizes the loading status
of the data by well plate.
}
