% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getMetaDf}
\alias{getMetaDf}
\alias{getMetaDf,Cypro-method}
\alias{getMetaDf,CyproTimeLapseMP-method}
\title{Obtain meta data (by cell id)}
\usage{
getMetaDf(object, ...)

\S4method{getMetaDf}{Cypro}(object)

\S4method{getMetaDf}{CyproTimeLapseMP}(object, phase = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
This function lets you extract a data.frame that contain variables
with which cells are grouped.
}
