% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getMetaVariableNames}
\alias{getMetaVariableNames}
\alias{getMetaVariableNames,Cypro-method}
\alias{getMetaVariableNames,CyproTimeLapseMP-method}
\title{Obtain meta variable names of cell data}
\usage{
getMetaVariableNames(object, ...)

\S4method{getMetaVariableNames}{Cypro}(object, ..., named = FALSE, verbose = NULL)

\S4method{getMetaVariableNames}{CyproTimeLapseMP}(object, ..., named = FALSE, phase = NULL, verbose = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}

\item{named}{Logical value. If set to TRUE the grouping variables are named
according to their grouping type (cluster, meta or well_plate).}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A character vector.
}
\description{
Convenient access to the names of the data variables of your cell data. Useful to
obtain vectors of variable names as input for recurring arguments like \code{across} or
\code{grouping_variable}.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
