% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getMissingValuesDf}
\alias{getMissingValuesDf}
\title{Obtain missing value counts}
\usage{
getMissingValuesDf(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
A data.frame.
}
\description{
This function returns a data.frame giving insight into
the number of missing values every cell has across all variables.
}
