% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getMissingValuesMax}
\alias{getMissingValuesMax}
\title{Obtain missing value counts}
\usage{
getMissingValuesMax(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
A data.frame.
}
\description{
This function returns the maximum number of missing values
every cell has across all variables.
}
