% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getModule}
\alias{getModule}
\title{Obtain analysis module}
\usage{
getModule(object, module_name, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{module_name}{Character value. The name of the analysis module.}
}
\value{
An object of class \code{AnalysisModule}.
}
\description{
Extracts the complete S4-object of class \code{AnalysisModule} from
the \code{Cypro} object with all its content.
}
