% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R
\name{getMtr}
\alias{getMtr}
\title{Obtain data matrix}
\usage{
getMtr(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\description{
Extracts the numeric data of the object in form
of a matrix. The key variable is used as rownames.
}
