% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getS.R
\name{getOutlierDetection}
\alias{getOutlierDetection}
\alias{getOutlierDetection,CyproScreening-method}
\alias{getOutlierDetection,CyproTimeLapse-method}
\alias{getOutlierDetection,CyproTimeLapseMP-method}
\title{Obtain object of class \code{OutlierDetection}}
\usage{
getOutlierDetection(object, ...)

\S4method{getOutlierDetection}{CyproScreening}(object, verbose = TRUE, ...)

\S4method{getOutlierDetection}{CyproTimeLapse}(object, verbose = TRUE, ...)

\S4method{getOutlierDetection}{CyproTimeLapseMP}(object, phase, verbose = TRUE, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An object of S4-class \code{OutlierDetection}.
}
\description{
Extracts the complete \code{OutlierDetection} object from
slot @quality_checks.
}
