% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getOutlierIDs}
\alias{getOutlierIDs}
\alias{getOutlierIDs,Cypro-method}
\title{Obtain outlier cell IDs}
\usage{
getOutlierIDs(object, ...)

\S4method{getOutlierIDs}{Cypro}(
  object,
  method = "IQR",
  features = NULL,
  across = NULL,
  across_subset = NULL,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method}{Character value. The outlier detection method. Valid
input options are \emph{'IQR'} and \emph{'Mahalanobis'}.}

\item{features}{Character vector. Denotes the features of interest. \emph{Features} are numeric
variables in the cell data (slot @cdata) of the \code{Cypro} object.

Use \code{getFeatureNames()} to obtain all valid input options.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector, list of character vectors or NULL (see details for
more). Specifies the particular groups of interest the grouping variable specified in argument
\code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}
}
\value{
Character vector or named list of character vectors.
}
\description{
Extracts the cell IDs that were identified as outliers. See
details for more.
}
\details{
The class of the return value depends on the input for argument
\code{across}. If \code{across} = NULL, a character vector of cell IDs
is returned. If \code{across} is a character, a list of character vectors
is returned. Each character vector in the list represents the results for
one of the group the grouping variable denoted in \code{across} contains.

Depending on the input for argument \code{method} additional arguments
can be specified to refine the output.

\itemize{
\item{\code{method} = \emph{'IQR'}: }{Input for argument \code{features} determines
the numeric features for which outlier detection has been conducted. If NULL, all
are considered. If character, only cell IDs that were identified as outliers within
the specified features/variables are included in the return value.}
}
}
\seealso{
\code{getOutlierResults()}
}
