% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getOutlierWells}
\alias{getOutlierWells}
\title{Obtain possible outlier wells}
\usage{
getOutlierWells(object, threshold = 0.75, verbose = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{threshold}{Numeric value. The minimum percentage of wells to which
a wells distance must be assumed to be an outlier. Defaults to 0.75.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame of well plate wells.
}
\description{
Returns a data.frame of well plate wells that indicates
how often a wells distance value to other wells is assumed to
be an outlier by \code{grDevices::boxplot.stats()}.
}
