% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-pam.R
\name{getPam}
\alias{getPam}
\alias{getPam,Cypro-method}
\alias{getPam,CyproTimeLapseMP-method}
\title{Obtain object of class \code{pam}}
\usage{
getPam(object, ...)

\S4method{getPam}{Cypro}(
  object,
  k,
  method_pam = "euclidean",
  fset = "all_features",
  stop_if_null = TRUE
)

\S4method{getPam}{CyproTimeLapseMP}(
  object,
  phase,
  k,
  method_pam = "euclidean",
  fset = "all_features",
  stop_if_null = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{k}{Numeric value. Denotes the number of clusters. Must be bigger than 1. Must
not be bigger than the number of observations of the data set.}

\item{method_pam}{Character value. Specifies the pam-method of interest.
Use \code{validMethodsPam()} to obtain all valid input options.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{stop_if_null}{Logical value. If TRUE and the function does not find the object to
extract an informative error is raised. Else the empty value - usually NULL - is returned.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An object of class \code{pam}.
}
\description{
Extracts an object of class \code{pam}
as described in \code{?cluster::pam}.
}
