% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{getPcaConv}
\alias{getPcaConv}
\alias{getPcaObject}
\alias{getTsneConv}
\alias{getTsneObject}
\alias{getUmapConv}
\alias{getUmapObject}
\title{Obtain dimensional reduction objects}
\usage{
getPcaConv(object, variable_set, phase = NULL)

getPcaObject(object, variable_set, phase = NULL)

getTsneConv(object, variable_set, phase = NULL)

getTsneObject(object, variable_set, phase = NULL)

getUmapConv(object, variable_set, phase = NULL)

getUmapObject(object, variable_set, phase = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An S4 object of class \emph{'dim_red_conv'}.
}
\description{
Returns the S4 objects in which the dimensional reduction
results are stored.
}
