% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getPhaseNames}
\alias{getPhaseNames}
\alias{getPhaseNames,CyproTimeLapseMP-method}
\title{Obtain phase start points}
\usage{
getPhaseNames(object, ...)

\S4method{getPhaseNames}{CyproTimeLapseMP}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Character vector.
}
\description{
Extracts the names of the phases.
}
