% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getPhases}
\alias{getPhases}
\alias{getPhases,CyproTimeLapseMP-method}
\title{Extract phase names}
\usage{
getPhases(object)

\S4method{getPhases}{CyproTimeLapseMP}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Character vector.
}
\description{
Obtain the names of all phases the experiment design
contains in form of a vector.
}
