% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getS.R
\name{getPhaseStarts}
\alias{getPhaseStarts}
\alias{getPhaseStarts,CyproTimeLapseMP-method}
\title{Obtain phase start points}
\usage{
getPhaseStarts(object, ...)

\S4method{getPhaseStarts}{CyproTimeLapseMP}(object, numeric = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Numeric vector.
}
\description{
Extracts the frames with which the respective
experiment phases started.
}
