% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/correlation.R
\name{getRcorr}
\alias{getRcorr}
\alias{getRcorr,Cypro-method}
\alias{getRcorr,CyproTimeLapseMP-method}
\title{Obtain object of class \code{rcorr}}
\usage{
getRcorr(object, ...)

\S4method{getRcorr}{Cypro}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  as_list = FALSE
)

\S4method{getRcorr}{CyproTimeLapseMP}(
  object,
  phase,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  as_list = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method_corr}{Character value. Specifies the correlation
method of interest. Use \code{validMethodsCorrelation()} to obtain all valid
input options.}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{across_subset}{Character vector, list of character vectors or NULL (see details for
more). Specifies the particular groups of interest the grouping variable specified in argument
\code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{as_list}{Logical value. If TRUE, \code{rcorr} objects are returned
as simple lists.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
\code{rcorr} returns a list with elements \code{r}, the
matrix of correlations, \code{n} the
matrix of number of observations used in analyzing each pair of variables,
and \code{P}, the asymptotic P-values.
Pairs with fewer than 2 non-missing values have the r values set to NA.
The diagonals of \code{n} are the number of non-NAs for the single variable
corresponding to that row and column.
}
\description{
Extracts correlation results in form of class \code{rcorr}.
}
\details{
If \code{across} is NULL one object of class \code{rcorr} is
returned - the one for the complete data set. If \code{across} is a character the list of \code{rcorr} is
extracted and subsetted by the groups of interest denoted in \code{across_subset}.
If \code{across_subset} is NULL the complete list is returned. If \code{across_subset}
is of length one and \code{flatten} is TRUE the output list is flattened and
a single \code{rcorr} object is returned.
}
