% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R
\name{getResults}
\alias{getResults}
\title{Obtain analysis results}
\usage{
getResults(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Depends on the objects class.
}
\description{
Generic extractor for results mainly used for
programming purpose as it provides informative error
messages if the requested content is missing.
}
