% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/getN-getZ.R
\name{getScaledDf}
\alias{getScaledDf}
\alias{getScaledDf,Cypro-method}
\alias{getScaledDf,CyproTimeLapseMP-method}
\title{Obtain scaled data}
\usage{
getScaledDf(object, ...)

\S4method{getScaledDf}{Cypro}(
  object,
  with_cluster = FALSE,
  with_meta = FALSE,
  with_well_plate = FALSE,
  with_everything = FALSE,
  ...
)

\S4method{getScaledDf}{CyproTimeLapseMP}(
  object,
  phase,
  with_cluster = FALSE,
  with_meta = FALSE,
  with_well_plate = FALSE,
  with_everything = FALSE,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{with_cluster}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{with_meta}{Logical. If set to TRUE the discrete variables of the cluster data slot are added
to the output data.frame.}

\item{with_well_plate}{Logical values. Denoting
if the respective grouping information should be joined to the output data.frame
or not.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
Extracts scaled data as a data.frame. In case of screening experiments
it's the scaled data from slot @features. In case of time lapse
experiments its the scaled data from slot @stat_features.
}
