% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/getN-getZ.R
\name{getScaledMtr}
\alias{getScaledMtr}
\alias{getScaledMtr,Cypro-method}
\alias{getScaledMtr,CyproTimeLapseMP-method}
\title{Obtain scaled matrix}
\usage{
getScaledMtr(object, ...)

\S4method{getScaledMtr}{Cypro}(object, ...)

\S4method{getScaledMtr}{CyproTimeLapseMP}(object, phase, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A matrix.
}
\description{
Extracts scaled data as a matrix. In case of screening experiments
it's the scaled data from slot @features. In case of time lapse
experiments its the scaled data from slot @stat_features.
}
