% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSetUpDf}
\alias{getSetUpDf}
\title{Obtain well plate set up}
\usage{
getSetUpDf(object, well_plate_name = NULL)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
A data.frame in which each observation represents the well of a well plate
}
\description{
Access function to the experiment set up in a tidy-data fashion.
}
