% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-pam.R
\name{getSilWidthsDf}
\alias{getSilWidthsDf}
\alias{getSilWidthsDf,Cypro-method}
\alias{getSilWidthsDf,CyproTimeLapseMP-method}
\title{Obtain sil-width data}
\usage{
getSilWidthsDf(object, ...)

\S4method{getSilWidthsDf}{Cypro}(object, ks, method_pam = "euclidean", fset = "all_features")

\S4method{getSilWidthsDf}{CyproTimeLapseMP}(
  object,
  phase,
  ks,
  method_pam = "euclidean",
  fset = "all_features"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{method_pam}{Character value. Specifies the pam-method of interest.
Use \code{validMethodsPam()} to obtain all valid input options.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame that contains information
about the silhouette width of every observation in different
clustering results with method \emph{pam}.
}
