% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getStatVariableNames}
\alias{getStatVariableNames}
\alias{getTrackVariableNames}
\title{Obtain numeric variables of cell data}
\usage{
getStatVariableNames(object, ...)

getTrackVariableNames(object, ...)
}
\arguments{
\item{object}{A valid cypro object.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}
}
\value{
A character vector.
}
\description{
Convenient access to the names of your objects data variables. Useful to
obtain vectors of variable names as input for recurring arguments like \code{variables}.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
