% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getStorageDirectory}
\alias{getStorageDirectory}
\title{Obtain storage directory}
\usage{
getStorageDirectory(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
Character value.
}
\description{
Obtain storage directory
}
