% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family-N-Z.R
\name{getSubsetList}
\alias{getSubsetList}
\alias{getSubsetList,Cypro-method}
\title{Extract subset information}
\usage{
getSubsetList(object, nth)

\S4method{getSubsetList}{Cypro}(object, nth = 1)
}
\arguments{
\item{nth}{Numeric value. Denotes the ordinal number of the subsetting
you are interested in.}
}
\value{
A list.
}
\description{
Whenever a \code{Cypro} object is subsetted by one of it's
\code{subsetBy*()} methods information about the subsetting is stored in
the resulting object. This functions uses as an extractor for these
information.
}
