% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getTracksDf}
\alias{getTracksDf}
\alias{getTracksDf,CdataTimeLapse-method}
\alias{getTracksDf,CyproTimeLapse-method}
\alias{getTracksDf,CyproTimeLapseMP-method}
\title{Extract track data.frame}
\usage{
getTracksDf(
  object,
  with_cluster = FALSE,
  with_meta = FALSE,
  with_well_plate = FALSE,
  ...
)

\S4method{getTracksDf}{CdataTimeLapse}(
  object,
  with_cluster = FALSE,
  with_meta = FALSE,
  with_well_plate = FALSE,
  ...
)

\S4method{getTracksDf}{CyproTimeLapse}(
  object,
  with_cluster = FALSE,
  with_meta = FALSE,
  with_well_plate = FALSE,
  ...
)

\S4method{getTracksDf}{CyproTimeLapseMP}(
  object,
  phase = 1,
  with_cluster = FALSE,
  with_meta = FALSE,
  with_well_plate = FALSE,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{with_cluster}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{with_meta}{Logical. If set to TRUE the discrete variables of the cluster data slot are added
to the output data.frame.}

\item{with_well_plate}{Logical values. Denoting
if the respective grouping information should be joined to the output data.frame
or not.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame in which each observation refers to a cell at a given frame.
}
\description{
Obtain numeric cell features of time lapse experiments in
form of a data.frame.
}
