% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTrackVariableNames}
\alias{getTrackVariableNames}
\alias{getTrackVariableNames,CyproTimeLapse-method}
\alias{getTrackVariableNames,CyproTimeLapseMP-method}
\title{Obtain numeric variables of cell track data}
\usage{
getTrackVariableNames(object, ...)

\S4method{getTrackVariableNames}{CyproTimeLapse}(object, ...)

\S4method{getTrackVariableNames}{CyproTimeLapseMP}(object, phase = NULL, ...)
}
\arguments{
\item{object}{A valid cypro object.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A character vector.
}
\description{
Convenient access to the names of the numeric data variables of your
cell data. Useful to obtain vectors of variable names as input for recurring
arguments like \code{variables}.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
