% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getVariableAssignment}
\alias{getVariableAssignment}
\alias{getVariableAssignment,AnalysisModule-method}
\alias{getVariableAssignment,Cypro-method}
\title{Obtain variable assignment}
\usage{
getVariableAssignment(object, ...)

\S4method{getVariableAssignment}{AnalysisModule}(
  object,
  variables = NULL,
  required = TRUE,
  optional = TRUE,
  computable = TRUE
)

\S4method{getVariableAssignment}{Cypro}(
  object,
  required = TRUE,
  optional = TRUE,
  computable = TRUE,
  modules = NULL,
  variables = NULL,
  drop_na = FALSE,
  flatten = FALSE,
  reverse = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{variables, modules}{Character vector or NULL. If character vector only
the variables and modules mentioned are included.}

\item{required, optional, computable}{Logical values. Indicate if the respective
variable type should be included or not.}

\item{drop_na}{Logical. If set to TRUE all empty values are discarded.}

\item{flatten}{Logical. If set to TRUE the list is flattened to a character vector
via \code{purrr::flatten_chr()}.}
}
\value{
A named list or a named character vector.
}
\description{
Extracts the variable names of the experiment data that has been
assigned to the data variables that \code{cypro} knows.
}
