% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getVariableAssignmentID}
\alias{getVariableAssignmentID}
\alias{getVariableAssignmentID,CyproScreening-method}
\alias{getVariableAssignmentID,CyproTimeLapse-method}
\alias{getVariableAssignmentID,CyproTimeLapseMP-method}
\title{Obtain variable assignment}
\usage{
getVariableAssignmentID(object, drop_na = FALSE)

\S4method{getVariableAssignmentID}{CyproScreening}(object, drop_na = FALSE)

\S4method{getVariableAssignmentID}{CyproTimeLapse}(object, drop_na = FALSE)

\S4method{getVariableAssignmentID}{CyproTimeLapseMP}(object, drop_na = FALSE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{drop_na}{Logical. If set to TRUE all empty values are discarded.}
}
\value{
A character vector.
}
\description{
Extracts the variable names of the experiment data that has been
assigned to the data variables that \code{cypro} knows.
}
