% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getVariableSet}
\alias{getVariableSet}
\alias{getVariableSets}
\alias{getVariableSetNames}
\title{Obtain defined sets of variables}
\usage{
getVariableSet(object, variable_set)

getVariableSets(object)

getVariableSetNames(object)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_set}{Character value. The name of the variable set of interest.}
}
\value{
A list of character vectors or a character vector of names.
}
\description{
Convenient access to defined sets of variables or names
mentioned sets.
}
