% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getVariableValues}
\alias{getVariableValues}
\title{Obtain variable overview}
\usage{
getVariableValues(object, phase = NULL, variable_name)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{variable_name}{Character value. Denotes the variable of interest. Valid inputs can be
obtained via the function \code{getVariableNames()}.}
}
\value{
A character vector or a data.frame of one row containing basic descriptive statistics.
}
\description{
If the variable denoted in \emph{variable_name} is categorical (character or factor)
all unique values/levels are returned. If the variable is numeric it is given to
\code{psych::describe()} which returns a statistical summary.
}
