% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT-getZ.R
\name{getWellLevels}
\alias{getWellLevels}
\alias{getWellLevels,WellPlate-method}
\alias{getWellLevels,ExperimentDesign-method}
\alias{getWellLevels,Cypro-method}
\title{Obtain well levels}
\usage{
getWellLevels(object)

\S4method{getWellLevels}{WellPlate}(object)

\S4method{getWellLevels}{ExperimentDesign}(object)

\S4method{getWellLevels}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A character vector.
}
\description{
Extracts a vector of well names in their natural order.
}
