% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT-getZ.R
\name{getWellPlate}
\alias{getWellPlate}
\alias{getWellPlate,ExperimentDesign-method}
\alias{getWellPlate,Cypro-method}
\title{Extract well plate}
\usage{
getWellPlate(object, well_plate = NULL)

\S4method{getWellPlate}{ExperimentDesign}(object, well_plate = NULL)

\S4method{getWellPlate}{Cypro}(object, well_plate = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
A data.frame.
}
\description{
Obtain object of class \code{WellPlate}.
}
