% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getWellPlateDf}
\alias{getWellPlateDf}
\alias{getWellPlateDf,Cypro-method}
\title{Extract cell well plate data (by cell id)}
\usage{
getWellPlateDf(object, ...)

\S4method{getWellPlateDf}{Cypro}(object, well_plate)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
A data.frame.
}
\description{
Obtain cell grouping variables according to the experiment
design.

Not to be confused with the well plate layout which is stored in data.frames
of class \code{layout_df} that can be obtained via \code{getLayoutDf()}.
}
