% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT-getZ.R
\name{getWellPlateDirectories}
\alias{getWellPlateDirectories}
\alias{getWellPlateDirectories,Cypro-method}
\title{Obtain well plate directories}
\usage{
getWellPlateDirectories(object, ...)

\S4method{getWellPlateDirectories}{Cypro}(object, well_plates = NULL, named = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{well_plates}{Character vector. The names of the well plates of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\description{
Extracts the directories assigned to each well plate.
}
