% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getWellPlateIndices}
\alias{getWellPlateIndices}
\alias{getWellPlateIndices,Cypro-method}
\title{Obtain well plate indices}
\usage{
getWellPlateIndices(object, ...)

\S4method{getWellPlateIndices}{Cypro}(object, return = "tibble")
}
\arguments{
\item{object}{A valid cypro object.}

\item{return}{Character value. Either \emph{'tibble'} or \emph{'vector'}.}
}
\value{
Depends on input for argument \code{return}.
}
\description{
Obtain well plate indices
}
