% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getWellPlateNames}
\alias{getWellPlateNames}
\alias{getWellPlateNames,Cypro-method}
\title{Obtain well plate names}
\usage{
getWellPlateNames(object)

\S4method{getWellPlateNames}{Cypro}(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
A character vector.
}
\description{
Obtain well plate names
}
