% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getWellPlates}
\alias{getWellPlates}
\alias{getWellPlates,ExperimentDesign-method}
\alias{getWellPlates,Cypro-method}
\title{Obtain list of WellPlate objects}
\usage{
getWellPlates(object, well_plates = NULL)

\S4method{getWellPlates}{ExperimentDesign}(object, well_plates = NULL)

\S4method{getWellPlates}{Cypro}(object, well_plates = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{well_plates}{Character vector. The names of the well plates of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
A named list of \code{WellPlate} objects.
}
\description{
Extracts a list of \code{WellPlate} objects.
}
