% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getWellPlateVariableNames}
\alias{getWellPlateVariableNames}
\alias{getWellPlateVariableNames,Cypro-method}
\title{Obtain well plate variable names of cell data}
\usage{
getWellPlateVariableNames(object, ...)

\S4method{getWellPlateVariableNames}{Cypro}(object, ...)
}
\arguments{
\item{object}{A valid cypro object.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}
}
\value{
A character vector.
}
\description{
Convenient access to the names of the data variables of your cell data. Useful to
obtain vectors of variable names as input for recurring arguments like \code{across} or
\code{grouping_variable}.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
