% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getWells}
\alias{getWells}
\alias{getWells,layout_df-method}
\alias{getWells,layout_df_mp-method}
\alias{getWells,WellPlate-method}
\alias{getWellRois}
\alias{getWellRois,layout_df-method}
\alias{getWellRois,layout_df_mp-method}
\alias{getWellRois,WellPlate-method}
\title{Obtain wells and well rois}
\usage{
getWells(object, info_status = info_status_levels)

\S4method{getWells}{layout_df}(object, info_status = info_status_levels)

\S4method{getWells}{layout_df_mp}(object, info_status = info_status_levels)

\S4method{getWells}{WellPlate}(object, info_status = info_status_level)

getWellRois(object, info_status = info_status_levels)

\S4method{getWellRois}{layout_df}(object, info_status = info_status_levels)

\S4method{getWellRois}{layout_df_mp}(object, info_status = info_status_levels)

\S4method{getWellRois}{WellPlate}(object, info_status = info_status_level)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{info_status}{Character vector. Valid options are \emph{'Complete', 'Incomplete', 'Missing'}. Refers
to the information status of a well regarding cell and condition.}
}
\value{
A character vector.
}
\description{
Extracts unique well or well regions of interest of a well plate
in form of a vector.
}
\details{
In all methods, data is eventually extracted from the layout data.frames.
}
