% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_add_data_to_cluster_object}
\alias{hlpr_add_data_to_cluster_object}
\title{Adds data to cluster object}
\usage{
hlpr_add_data_to_cluster_object(object, cluster_object, with_data, phase)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\description{
Adds data to cluster object
}
