% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_add_variable_specific_args}
\alias{hlpr_add_variable_specific_args}
\title{Complete argument list in compute_module_variables}
\usage{
hlpr_add_variable_specific_args(variable_info, args)
}
\description{
Complete argument list in compute_module_variables
}
