% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_assemble_directory}
\alias{hlpr_assemble_directory}
\title{Assemble a directory}
\usage{
hlpr_assemble_directory(input_list)
}
\arguments{
\item{input_list}{A named list: output of \code{shinyFiles::shinyDirButton()}}
}
\value{
A character string.
}
\description{
Assembles a single character string direction from the
output of \code{shinyFiles::shinyDirButton()}.
}
