% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_caption_add_on}
\alias{hlpr_caption_add_on}
\title{Clarifying caption}
\usage{
hlpr_caption_add_on(object, phase)
}
\arguments{
\item{object}{A valid cypro object.}
}
\description{
Returns a \code{ggplot2::labs()}-add on in which the
caption clarifies whether the plot refers to the times pan before the treatment,
after the treatment or to the entire experiment.

If there was not treatment or the cells were treated right from the beginning
no caption is returned.
}
