% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_coords_flip_add_on}
\alias{hlpr_coords_flip_add_on}
\alias{hlpr_plot_well_plate_fill}
\title{ggplot2 add on helpers}
\usage{
hlpr_coords_flip_add_on(flip_coords)

hlpr_plot_well_plate_fill(input)
}
\description{
Functions that either return an empty list
or the respective ggplot add-on.
}
