% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_create_meta_data}
\alias{hlpr_create_meta_data}
\title{Create Cell Meta Data}
\usage{
hlpr_create_meta_data(df, phase, verbose)
}
\description{
Filters meta variables from track data. To be used in purrr::map2().
}
