% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_create_track_list}
\alias{hlpr_create_track_list}
\title{Process Track Data.frame}
\usage{
hlpr_create_track_list(phase, track_df)
}
\description{
Filters the whole track data frame into it's phase-subparts
and recomputes the variable \emph{dfo (= Distance from origin)} if the
filtered part does not belong to the phase \code{'before_tmt'}.

To be used as input for \code{.f} in \code{purrr::map()}
inside the function \code{compileCto()}.
}
