% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_max_track_na_count}
\alias{hlpr_max_track_na_count}
\title{Assess number of missing values}
\usage{
hlpr_max_track_na_count(object)
}
\description{
Returns the maximum count of missing values per cell id
across all variables
}
