% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_merge_conditions}
\alias{hlpr_merge_conditions}
\alias{hlpr_merge_condition_by_id}
\title{Merge conditions}
\usage{
hlpr_merge_conditions(track_df, phase, across, verbose = TRUE)

hlpr_merge_condition_by_id(condition)
}
\description{
Merge conditions
}
