% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_module_name_pairs}
\alias{hlpr_module_name_pairs}
\title{Returns a vector of names in example named by the actual reference in cypro}
\usage{
hlpr_module_name_pairs(assembled_module_info)
}
\description{
Returns a vector of names in example named by the actual reference in cypro
}
