% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_phase_vertical_line}
\alias{hlpr_phase_vertical_line}
\title{Add vertical phase separating line to lineplot}
\usage{
hlpr_phase_vertical_line(object, phase, display_vline, vline_clr, vline_type)
}
\description{
Add vertical phase separating line to lineplot
}
