% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_pretty_colnames}
\alias{hlpr_pretty_colnames}
\alias{hlpr_pretty_value}
\alias{hlpr_pretty_vec}
\title{Make pretty column names}
\usage{
hlpr_pretty_colnames(df)

hlpr_pretty_value(value)

hlpr_pretty_vec(vec)
}
\arguments{
\item{df}{A data.frame that might contain variables for which prettier versions exist.}

\item{value}{A name for which a prettier version might exist.}

\item{vec}{A vector of variable names for which prettier versions might exist.}
}
\description{
Helper around the problem with concise vs. pretty columnnames
}
