% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_rename_df_cols}
\alias{hlpr_rename_df_cols}
\title{Rename cell tracker columns}
\usage{
hlpr_rename_df_cols(df, denoted_columns = NULL, additional_columns = NULL)
}
\value{
A renamed data.frame.
}
\description{
Rename cell tracker columns
}
