% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_wp_directories}
\alias{hlpr_wp_directories}
\alias{hlpr_wp_file_number_f}
\alias{hlpr_wp_file_number}
\alias{hlpr_wp_ambiguous_number_f}
\alias{hlpr_wp_ambiguous_number}
\alias{hlpr_wp_exp_file_number}
\title{Return directory of well plate}
\usage{
hlpr_wp_directories(wp_list)

hlpr_wp_file_number_f(wp_list)

hlpr_wp_file_number(...)

hlpr_wp_ambiguous_number_f(wp_list)

hlpr_wp_ambiguous_number(...)

hlpr_wp_exp_file_number(wp_list)
}
\arguments{
\item{wp_list}{A well plate list.}
}
\description{
A set of functions that extract information
from well plate data lists.
}
