% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{hovered_well_info_shiny}
\alias{hovered_well_info_shiny}
\alias{hovered_well_info_shiny_mp}
\title{Create printable well info df}
\usage{
hovered_well_info_shiny(df)

hovered_well_info_shiny_mp(df)
}
\description{
Create printable well info df
}
