% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-a-m.R
\name{initiateEmptyCyproObject}
\alias{initiateEmptyCyproObject}
\title{Initiate empty cypro object}
\usage{
initiateEmptyCyproObject(...)
}
\arguments{
\item{...}{}
}
\description{
Initiates an empty cypro object and
adds the version under which is is created.
}
