% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-initiation.R
\name{initiateHierarchicalClustering}
\alias{initiateHierarchicalClustering}
\alias{initiateKmeansClustering}
\alias{initiatePamClustering}
\title{Set up clustering objects with cypro}
\usage{
initiateHierarchicalClustering(
  object,
  variable_set,
  phase = NULL,
  scale = TRUE,
  force = FALSE,
  verbose = NULL
)

initiateKmeansClustering(
  object,
  variable_set,
  phase = NULL,
  scale = TRUE,
  force = FALSE,
  verbose = NULL
)

initiatePamClustering(
  object,
  variable_set,
  phase = NULL,
  scale = TRUE,
  force = FALSE,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{scale}{Logical value. If set to TRUE (the default) all variables will be scaled before clustering is
performed to ensure that all variables are weighted equally.}

\item{force}{Logical value. Needs to be set to TRUE if you want to overwrite an already existing
set up or already existing results.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{variables_subset}{Character vector or NULL. Denotes the numeric variables the subsequent clustering
steps will include which influences the clustering results.

If set to NULL all of them are chosen. You can prefix variables you do NOT want to influence the clustering
with a \emph{'-'}. (Saves writing if there are more variables you are interested in
than variables you are not interested in.)

Use \code{getNumericVariableNames()} to obtain all valid input options.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
These functions set up the necessary objects to perform clustering with the
respective algorithm. See details for more.
}
\details{
The clustering initiation functions set up the S4 object that is used by cypro to do the clustering.
Every downstream analysis function depends on the input you specify here. This means in particular the input for
argument \code{scale} and the input for argument \code{variables_subset}. The latter denotes the variables on which
the clustering will base on. Changing these might influence the clustering results. If you realize later on that
you want to change the set up use the respective \code{initiate*Clustering()} function again and set argument \code{force}
to TRUE in order to overwrite the current set up.
}
\seealso{
getHclustConv(), getKmeansConv(), getPamConv()
}
