% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{input_df}
\alias{input_df}
\title{input_df}
\usage{
input_df(input_df)
}
\arguments{
\item{input_df}{A data.frame that contains at least two variables:

\describe{
\item{\emph{cell_id}}{Character. The cells ids that are used to join the cluster variable unambiguously}
\item{\code{variable__name}}{Character or factor. The discrete variable that is to be joined.}
}}
}
\description{
input_df
}
