% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-is.R
\name{is_complete}
\alias{is_complete}
\alias{is_complete.cypro_df}
\alias{isComplete}
\alias{isComplete.cypro_df}
\title{Test completeness}
\usage{
is_complete(df)

\method{is_complete}{cypro_df}(df)

isComplete(df)

\method{isComplete}{cypro_df}(df)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the data.frame has been completed.
}
\details{
Only relevant for \code{cypro_df} data.frames of subclass
\code{stats_df} or \code{tracks_df}.
}
\seealso{
\code{completeStatsDf()}, \code{completeTracksDf()}
}
