% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-is.R
\name{is_cluster_df}
\alias{is_cluster_df}
\alias{is_cypro_df}
\alias{is_well_plate_df}
\alias{isClusterDf}
\alias{isCyproDf}
\title{Test known data.frame classes}
\usage{
is_cluster_df(df, stop_if_false = FALSE)

is_cypro_df(df, stop_if_false = FALSE)

is_well_plate_df(df, stop_if_false = FALSE)

isClusterDf(df, stop_if_false = FALSE)

isCyproDf(df, stop_if_false = FALSE)

isCyproDf(df, stop_if_false = FALSE)
}
\arguments{
\item{df}{A data.frame.}

\item{stop_if_false}{Logical value. If set to TRUE and the test evaluates to FALSE the
function is stopped via \code{stop()} printing an informative message. If FALSE the
resulting logical value is returned.}
}
\value{
TRUE or FALSE
}
\description{
Logical tests that check input data.frame for its class
and its content.
}
\details{
Tests if the input data.frame is of the respective class by
checking its attributes. Additionally tests if the content of the data.frame
matches the requirements. If both tests evaluate to TRUE the function
returns TRUE else it returns FALSE.
}
