% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{is_modified_wp_df}
\alias{is_modified_wp_df}
\title{Well plate data.frame check}
\usage{
is_modified_wp_df(wp_df)
}
\description{
Checks if there is a current well plate data.frame that has
been modified but not saved yet. Prevents 'New Well Plate' from overwriting
unsaved progress.
}
