% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-is.R
\name{isActive}
\alias{isActive}
\alias{isActive,Cypro-method}
\alias{isActive,AnalysisModule-method}
\title{Test module activity}
\usage{
isActive(object, ..., stop_if_false = FALSE)

\S4method{isActive}{Cypro}(object, module_name, ..., stop_if_false = FALSE)

\S4method{isActive}{AnalysisModule}(object, ..., stop_if_false = FALSE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{stop_if_false}{Logical value. If set to TRUE and the test evaluates to FALSE the
function is stopped via \code{stop()} printing an informative message. If FALSE the
resulting logical value is returned.}

\item{module_name}{Character value. The name of the analysis module. Use \code{validModuleNames()} or
\code{getModuleNames()} to obtain valid input options.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the module's slot @active is set to TRUE.
}
\details{
The slot @active indicates whether the user declared to want to
use the module during \code{assignVariables()}. A module becomes \emph{usable}
if all assigned variables are valid.
}
