% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-is.R
\name{isCompleteExampleDf}
\alias{isCompleteExampleDf}
\alias{isCompleteExampleDf,Cypro-method}
\title{Test example data.frame}
\usage{
isCompleteExampleDf(object)

\S4method{isCompleteExampleDf}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the data file loaded as an example data.frame during
\code{assignVariables()} covers the whole experiment.
}
