% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-test-is.R
\name{isLayoutDf}
\alias{isLayoutDf}
\title{Test layout data.frame}
\usage{
isLayoutDf(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
TRUE or FALSE.
}
\description{
Tests if the input data.frame is of class \code{layout_df}.
}
