% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-is.R
\name{isLoaded}
\alias{isLoaded}
\alias{isLoaded,DataFile-method}
\title{Test if loaded}
\usage{
isLoaded(object)

\S4method{isLoaded}{DataFile}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the data file has been loaded yet.
}
