% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-is.R
\name{isModifiedLayoutDf}
\alias{isModifiedLayoutDf}
\title{Test layout data.frame modification}
\usage{
isModifiedLayoutDf(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
TRUE or FALSE.
}
\description{
Checks if the input data.frame has been modified by \code{setWellInfo()}.
}
