% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-is.R
\name{isMultiplePhase}
\alias{isMultiplePhase}
\alias{isMultiplePhase,data.frame-method}
\alias{isMultiplePhase,layout_df-method}
\alias{isMultiplePhase,layout_df_mp-method}
\alias{isMultiplePhase,WellPlate-method}
\alias{isMultiplePhase,ExperimentDesign-method}
\alias{isMultiplePhase,Cypro-method}
\title{Test multiple phases}
\usage{
isMultiplePhase(object, stop_if_false = FALSE)

\S4method{isMultiplePhase}{data.frame}(object, stop_if_false = FALSE)

\S4method{isMultiplePhase}{layout_df}(object, stop_if_false = FALSE)

\S4method{isMultiplePhase}{layout_df_mp}(object, stop_if_false = FALSE)

\S4method{isMultiplePhase}{WellPlate}(object, stop_if_false = FALSE)

\S4method{isMultiplePhase}{ExperimentDesign}(object, stop_if_false = FALSE)

\S4method{isMultiplePhase}{Cypro}(object, stop_if_false = FALSE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{stop_if_false}{Logical value. If set to TRUE and the test evaluates to FALSE the
function is stopped via \code{stop()} printing an informative message. If FALSE the
resulting logical value is returned.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the provided object derives from a multiple phase experiment.
}
