% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-is.R
\name{isNested}
\alias{isNested}
\alias{isNested.layout_df}
\title{Test layout data.frame}
\usage{
isNested(df)

\method{isNested}{layout_df}(df)
}
\arguments{
\item{df}{A data.frame of class \code{layout_df}.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the layout data.frame exists in nested form.
}
