% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-test-is.R
\name{isSinglePhase}
\alias{isSinglePhase}
\alias{isSinglePhase,WellPlate-method}
\alias{isSinglePhase,ExperimentDesign-method}
\alias{isSinglePhase,Cypro-method}
\title{Test single phase}
\usage{
isSinglePhase(object, ...)

\S4method{isSinglePhase}{WellPlate}(object, stop_if_false = FALSE)

\S4method{isSinglePhase}{ExperimentDesign}(object, stop_if_false = FALSE)

\S4method{isSinglePhase}{Cypro}(object, stop_if_false = FALSE)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{stop_if_false}{Logical value. If set to TRUE and the test evaluates to FALSE the
function is stopped via \code{stop()} printing an informative message. If FALSE the
resulting logical value is returned.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the provided object derives from a single phase experiment.
}
