% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests-is.R
\name{isUsable}
\alias{isUsable}
\alias{isUsable,AnalysisModule-method}
\alias{isUsable,Cypro-method}
\title{Test module usability}
\usage{
isUsable(
  object,
  ...,
  fdb_if_false = FALSE,
  fdb_fn = "warning",
  verbose = FALSE
)

\S4method{isUsable}{AnalysisModule}(
  object,
  ...,
  fdb_if_false = FALSE,
  fdb_fn = "warning",
  verbose = FALSE
)

\S4method{isUsable}{Cypro}(
  object,
  module_name,
  ...,
  fdb_if_false = FALSE,
  fdb_fn = "warning",
  verbose = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{module_name}{Character value. The name of the analysis module. Use \code{validModuleNames()} or
\code{getModuleNames()} to obtain valid input options.}
}
\value{
TRUE or FALSE
}
\description{
Tests if the module's slot @active is set to TRUE AND
tests if all required variables have been assigned to a variable
in the input example data.frame that meets the variable specific requirements.
}
