% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-a-m.R
\name{joinWith}
\alias{joinWith}
\alias{joinWith,Cdata-method}
\alias{joinWith,CdataTimeLapseMP-method}
\title{Join cell data frames}
\usage{
joinWith(object, df, ...)

\S4method{joinWith}{Cdata}(
  object,
  df,
  with_cluster = FALSE,
  with_meta = FALSE,
  with_well_plate = FALSE,
  with_everything = FALSE
)

\S4method{joinWith}{CdataTimeLapseMP}(
  object,
  df,
  with_cluster = FALSE,
  with_meta = FALSE,
  with_well_plate = FALSE,
  with_everything = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{df}{A data.frame of class \code{cypro_df}.}

\item{with_cluster}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{with_meta}{Logical. If set to TRUE the discrete variables of the cluster data slot are added
to the output data.frame.}

\item{with_well_plate}{Logical values. Denoting
if the respective grouping information should be joined to the output data.frame
or not.}
}
\value{
A data.frame.
}
\description{
Joins different variable types of the cell data to a
data.frame by the key variable \emph{cell_id}.
}
