% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-helper.R
\name{labelsXrotate}
\alias{labelsXrotate}
\alias{labelsYrotate}
\alias{labelsXremove}
\alias{labelsYremove}
\title{Manipulate labels of ggplot output}
\usage{
labelsXrotate(angle = 90)

labelsYrotate(angle = 90)

labelsXremove()

labelsYremove()
}
\arguments{
\item{angle}{Numeric value. Denotes the angle with which
the labels are to be rotated.}
}
\value{
gg object that can be added to ggplots via the \code{+}-operator
}
\description{
Allows to rotate the labels of the x-/y-axis
of ggplot2 plots. Useful in case of overlap.
}
