% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-helper.R
\name{legendBottom}
\alias{legendBottom}
\alias{legendNone}
\alias{legendRight}
\alias{legendTop}
\title{Specify legend position of ggplot output}
\usage{
legendBottom(...)

legendNone(...)

legendRight(...)

legendTop(...)
}
\description{
Specify legend position of ggplot output
}
