% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{load_data_files_shiny}
\alias{load_data_files_shiny}
\title{Read in cell track data}
\usage{
load_data_files_shiny(
  wp_list,
  wp_name,
  session,
  object,
  assembled_module_info_lists,
  used_variable_names,
  mitosis_module_used
)
}
\arguments{
\item{wp_list}{A well plate list.}

\item{wp_name}{The name of the respective well plate.}
}
\value{
A nested list obtained by \code{purrr::safely()} with
successfully loaded files and failed ones including the error message.
}
\description{
Reads in excel- and csv-files of all valid
directories while displaying the progress via a progress bar.
}
