% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-initiation.R
\name{loadData}
\alias{loadData}
\title{Object initiation: Step 2}
\usage{
loadData(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
An updated cypro object.
}
\description{
Opens an interactive application in which the folders containing
the data files are assigned to the well plates and where the data files
are read in.
}
