% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-initiation.R
\name{loadDataFile}
\alias{loadDataFile}
\alias{loadDataFiles}
\title{Object initiation: Step 2}
\usage{
loadDataFile(object)

loadDataFiles(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
An updated cypro object.
}
\description{
Opens an interactive application in which the data to
be analyzed is loaded.

\code{loadDataFile()} assumes that the data
exist in one .csv, .xls, .xlsx or .txt file and that this data table
contains variables/columns that identify the well plate, the well
and  the region of interest.

\code{loadDataFiles()} assumes that the data tables exist for every well-region
of interest combination (every stack of images). Here you can assign a directory
to every well plate leading to a folder in which the files are placed.
Successful reading of the files requires them to be named according to
the syntax: \emph{well_roi.filetype}, e.g A1_1.xlsx, B12_4.csv
}
