% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDataFiles}
\alias{loadDataFiles}
\alias{loadDataFiles,Cypro-method}
\title{Data loading - Read data files}
\usage{
loadDataFiles(
  object,
  well_plates = NULL,
  check_vars = TRUE,
  check_additional = TRUE,
  in_shiny = FALSE
)

\S4method{loadDataFiles}{Cypro}(
  object,
  well_plates = NULL,
  check_vars = TRUE,
  check_additional = TRUE,
  in_shiny = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{well_plates}{Character vector. The names of the well plates of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\value{
The input object.
}
\description{
Second of three data loading steps. Reads data files and
validates their content based on the variables assigned during
\code{assignVariables()}. See details for more.
}
\details{
Every well plate that has been prepared via \code{prepareDataLoading()}
contains a list of \code{DataFile} objects in slot @files which again contain
the directories to each single file that is supposed to be loaded.

\code{DataFiles} that have already been read without any errors are skipped. Every
other \code{DataFiles} is read and the read content is validated regarding
the variable assignment during \code{assignVariables()}. The resulting named list
is then stored in slot @content of the respective \code{DataFile} object.
Error messages that occurred during the validation appear in form of \code{glue}
objects of length one in the list of slot @content.

\code{loadDataFiles()} can be called once after all directories have been set
via several calls to \code{prepareDataLoading()} or several times as files that
were loaded without errors are skipped. This allows to read files that
contained errors again without reading every other file, too.

Once all files are read, proceed with \code{processData()}.
}
