% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{loading_status_table_shiny}
\alias{loading_status_table_shiny}
\title{Display loading status}
\usage{
loading_status_table_shiny(well_plate_list)
}
\arguments{
\item{well_plate_list}{A list of well plate lists.}
}
\description{
Creates a data.frame that displays information
about the file availability evaluation undergone for every well
plate.
}
