% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{merge_condition}
\alias{merge_condition}
\alias{merge_condition.layout_df_mp}
\alias{mergeCondition.layout_df_mp}
\title{Merge multiple phase conditions}
\usage{
merge_condition(df, ...)

\method{merge_condition}{layout_df_mp}(df, collapse_with = " -> ", phases = NULL)

\method{mergeCondition}{layout_df_mp}(df, collapse_with = " -> ", phases = NULL)
}
\arguments{
\item{df}{A layout data.frame of class \code{layout_df_mp}.}
}
\value{
The input data.frame.
}
\description{
Merges the list variable \emph{condition}
of multiple phase layout data.frames to a character variable.
}
