% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designExperiment-shiny-module-server.R
\name{moduleExperimentDesignServer}
\alias{moduleExperimentDesignServer}
\title{Server logic: Experiment Design}
\usage{
moduleExperimentDesignServer(id, usage = "in_function")
}
\arguments{
\item{id}{Namespace ID}
}
\description{
Server logic to be used as input for \code{module}-argument
of function \code{shiny::moduleServer()}.
}
