% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData-shiny-module-server.R
\name{moduleLoadDataServer}
\alias{moduleLoadDataServer}
\title{Server logic: Load Data}
\usage{
moduleLoadDataServer(id, object)
}
\arguments{
\item{id}{Namespace ID}

\item{ed_input}{A reactive and named list. See value of \code{moduleExperimentSetUpServer()}.}
}
\value{
A named list:
}
\description{
Server logic to be used as input for \code{module}-argument
of function \code{shiny::moduleServer()}.
}
