% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-quality-check-server.R
\name{moduleQualityCheckServer}
\alias{moduleQualityCheckServer}
\title{Server logic: Quality Check}
\usage{
moduleQualityCheckServer(id, object)
}
\arguments{
\item{id}{Namespace ID}

\item{ld_input}{A reactive and named list. See value of \code{moduleLoadDataServer()}.}
}
\value{
A named list:
}
\description{
Server logic to be used as input for \code{module}-argument
of function \code{shiny::moduleServer()}.
}
