% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nAmbiguousFiles}
\alias{nAmbiguousFiles}
\alias{nAmbiguousFiles,WellPlate-method}
\title{Number of ambiguous file directories}
\usage{
nAmbiguousFiles(object)

\S4method{nAmbiguousFiles}{WellPlate}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A numeric value.
}
\description{
Returns the number of ambiguous file directories in
the folder assigned to slot @directory.
}
\details{
If the number of files in slot @file of \code{WellPlate} object
is zero the function returns NA.
}
