% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-n-z.R
\name{nestLayoutDf}
\alias{nestLayoutDf}
\alias{nestLayoutDf.layout_df}
\alias{unnestLayoutDf}
\alias{unnestLayoutDf.layout_df}
\title{Shift scope of well plate layout data.frame}
\usage{
nestLayoutDf(df)

\method{nestLayoutDf}{layout_df}(df)

unnestLayoutDf(df)

\method{unnestLayoutDf}{layout_df}(df)
}
\arguments{
\item{df}{A data.frame of class \code{layout_df}.}
}
\description{
Shift between layout data.frame scopes.
}
\details{
In nested layout data.frames every row represents a well.
They contain a list column named \emph{roi_info} that contains additional
information regarding the regions of interest of eachwell.

In unnested layout data.frames every row represents a region of interest of
a well. The list column \emph{roi_info} has been unnested into it's parts.

Which variables belong to \emph{roi_info} is stored in the layout data.frames
attributes and can change based on the experiment design and data loading modality.
}
