% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nExpectedFiles}
\alias{nExpectedFiles}
\alias{nExpectedFiles,WellPlate-method}
\title{Number of expected files}
\usage{
nExpectedFiles(object)

\S4method{nExpectedFiles}{WellPlate}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A numeric value.
}
\description{
Returns the number of files that is expected according to
the number of regions of interest, the number of wells with a complete information
status and the loading modality chosen - according to the input example.
}
