% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nFrames}
\alias{nFrames}
\title{Number of frames}
\usage{
nFrames(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A numeric value.
}
\description{
Returns the total number of frames as was denoted during \code{designExperiment()}.
}
