% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nMissingValuesStats}
\alias{nMissingValuesStats}
\title{Number of NAs by cell id  (in stats)}
\usage{
nMissingValuesStats(object, phase = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{variable_names}{Character vector or NULL. If character, denotes
the track variables of interest. If NULL all of them are chosen.}
}
\value{
A summarized data.frame. Variable \emph{na_count} refers to
the number of NAs across all stat variables. E.g a value of 8 in \emph{na_count}
means that the respective cell features missing values in 8 stat variables.
}
\description{
Sums up the number of missing values every cell id has
in the stats data.frame.
}
