% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nMissingValuesTracks}
\alias{nMissingValuesTracks}
\title{Number of NAs by cell id  (in tracks)}
\usage{
nMissingValuesTracks(object, variable_names, phase = NULL, verbose = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A summarized data.frame. Each row contains the number of missing
values a cell has in the respective variable.
}
\description{
Returns the number of missing values of every track
variable denoted in \code{variable_names} by cell id.
}
