% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeZscore}
\alias{normalizeZscore}
\title{Normalize data with zscore}
\usage{
normalizeZscore(object, across = NULL, verbose = TRUE)
}
\arguments{
\item{object}{}

\item{across}{Character value or NULL. If character, the normalization
is conducted separately for all groups of the grouping variable.
Must be a variable from the well plate data.frame.}
}
\value{
A cypro object.
}
\description{
Normalizes numeric data variables (apart from x-/y- coordinates)
by computing their zscore. In case of timelapse experiments the stat data.frame is affected.
}
