% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nPhases}
\alias{nPhases}
\alias{nPhases,layout_df_mp-method}
\alias{nPhases,CyproTimeLapseMP-method}
\title{Number of phases}
\usage{
nPhases(object)

\S4method{nPhases}{layout_df_mp}(object)

\S4method{nPhases}{CyproTimeLapseMP}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A numeric value
}
\description{
Returns the number of phases the experiment
design determines as a numeric value.
}
