% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nRois}
\alias{nRois}
\alias{nRois,WellPlate-method}
\alias{nRois,layout_df-method}
\title{Number of regions of interest}
\usage{
nRois(object)

\S4method{nRois}{WellPlate}(object)

\S4method{nRois}{layout_df}(object)
}
\arguments{
\item{object}{}
}
\value{
Numeric value.
}
\description{
Returns the number of regions of interest set during
\code{designExperiment()}.
}
