% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nSubsets}
\alias{nSubsets}
\alias{nSubsets,Cypro-method}
\title{Number of subsets}
\usage{
nSubsets(object)

\S4method{nSubsets}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Numeric value.
}
\description{
Returns the number of times the \code{Cypro}
object has undergone subsetting.
}
