% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nWellPlates}
\alias{nWellPlates}
\alias{nWellPlates,ExperimentDesign-method}
\alias{nWellPlates,Cypro-method}
\title{Number of well plates}
\usage{
nWellPlates(object)

\S4method{nWellPlates}{ExperimentDesign}(object)

\S4method{nWellPlates}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Numeric value.
}
\description{
Returns the number of well plates the experiment
design includes.
}
