% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nWells}
\alias{nWells}
\alias{nWells,layout_df-method}
\alias{nWells,WellPlate-method}
\alias{nWellRois}
\alias{nWellRois,layout_df-method}
\title{Number of wells and well rois}
\usage{
nWells(object, info_status = info_status_levels)

\S4method{nWells}{layout_df}(object, info_status = info_status_levels)

\S4method{nWells}{WellPlate}(object, info_status = info_status_levels)

nWellRois(object, info_status = info_status_levels)

\S4method{nWellRois}{layout_df}(object, info_status = info_status_levels)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{info_status}{Character vector. Valid options are \emph{'Complete', 'Incomplete', 'Missing'}. Refers
to the information status of a well regarding cell and condition.}
}
\value{
Numeric value.
}
\description{
Returns the absolute number of wells or well rois of a well plate.
}
