% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{OptionalVariable-class}
\alias{OptionalVariable-class}
\alias{OptionalVariable}
\title{The OptionalVariable Class}
\description{
Subclass of \code{DataVariable}. Similar architecture to
\code{RequiredVariable} but it differs from this class regarding what it represents,
hence, the additional class. See details for more information.
}
\details{
A variable is considered optional if it is not required for a module to
work and if it can neither be computed using required variables. This applies for the variables
that assign a cell to their localisation on the well plate: \emph{well_plate},
\emph{well}, \emph{well_roi}. These variables can be part of the input data files
but do not have to be as there are other options to add this information to the
\code{Cypro} object.
}
\seealso{
\code{AssignableVariable}, \code{RequiredVariable}, \code{SummaryVariable},
\code{loadDataFiles()}, \code{loadData()}.
}
