% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{plot_dim_red}
\alias{plot_dim_red}
\title{Title}
\usage{
plot_dim_red(
  object,
  variable_set,
  phase = NULL,
  method_dim_red = "pca",
  color_by = NULL,
  color_aes = "fill",
  pt_clrp = NULL,
  clrp_adjust = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_clrsp = NULL,
  pt_fill = NULL,
  pt_size = NULL,
  add_ons = list(ggplot2::labs(fill = NULL, color = NULL)),
  ...
)
}
\arguments{
\item{object}{}

\item{phase}{}

\item{method_dim_red}{}

\item{color_by}{}

\item{color_aes}{}

\item{clrp_adjust}{}

\item{pt_alpha}{}

\item{pt_clr}{}

\item{pt_fill}{}

\item{pt_size}{}

\item{clrp}{}

\item{clrsp}{}
}
\description{
Title
}
