% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{plot_qc_barplot_shiny}
\alias{plot_qc_barplot_shiny}
\title{Quality check barplot}
\usage{
plot_qc_barplot_shiny(df, aes_x, aes_fill, bar_position)
}
\arguments{
\item{df}{A data.frame}
}
\description{
Quality check barplot
}
