% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{plot_qc_histogram_shiny}
\alias{plot_qc_histogram_shiny}
\title{Quality check histogram}
\usage{
plot_qc_histogram_shiny(
  track_summary_df,
  aes_x = "skipped_meas",
  lab_x = "Measurements",
  legend_position = "none"
)
}
\arguments{
\item{track_summary_df}{Output of \code{quality_check_summary()}}

\item{aes_x}{Character value.}

\item{lab_x}{Character value.}

\item{aes_fill}{Character value.}

\item{lab_fill}{Character value.}
}
\description{
Quality check histogram
}
