% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{plot_well_plate_shiny}
\alias{plot_well_plate_shiny}
\title{Visualize the well plate}
\usage{
plot_well_plate_shiny(
  wp_df,
  selected_wells_df = NULL,
  aes_fill,
  fill_values = NULL,
  fill_guide = FALSE,
  aes_color,
  color_values,
  pt_size = 13.5,
  pt_stroke = 2,
  border = 0.75
)
}
\arguments{
\item{wp_df}{A well-plate data.frame.}

\item{selected_wells_df}{A subsetted well-plate data.frame or NULL.}

\item{aes_fill}{Character value. Variable of \code{wp_df} to map on the
fill-aesthetic.}

\item{aes_color}{Character value. Variable of \code{wp_df} to map on the
color-aesthetic.}

\item{pt_size}{Numeric value. Size of points that represent the wells.}

\item{pt_stroke}{Numeric value. Size of border that indicates what has been
mapped on the color-aesthetic.}

\item{border}{Numeric value. Distance between the outer wells and the plate's
margin.}
}
\value{
A ggplot.
}
\description{
Visualize the well plate
}
