% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-pam.R
\name{plotAvgSilWidths}
\alias{plotAvgSilWidths}
\alias{plotAvgSilWidths,Cypro-method}
\alias{plotAvgSilWidths,CyproTimeLapseMP-method}
\title{Plot avg sil-width data}
\usage{
plotAvgSilWidths(object, ...)

\S4method{plotAvgSilWidths}{Cypro}(
  object,
  ks,
  methods_pam = "euclidean",
  fset = "all_features",
  color = "steelblue",
  display_cols = TRUE,
  display_points = TRUE,
  display_line = TRUE,
  ncol = NULL,
  nrow = NULL
)

\S4method{plotAvgSilWidths}{CyproTimeLapseMP}(
  object,
  phase,
  ks,
  methods_pam = "euclidean",
  fset = "all_features",
  color = "steelblue",
  display_cols = TRUE,
  display_points = TRUE,
  display_line = TRUE,
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{methods_pam}{Character value. Specifies the pam-method of interest.
Use \code{validMethodsPam()} to obtain all valid input options.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{display_cols}{Logical value. If set to TRUE columns are used additionally
to display the results.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{display_line}{Logical value. If set to TRUE a line is used additionally
to display the results.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A ggplot.
}
\description{
Plots information about the average silhouette
widths of different clustering results with method \emph{pam}.
}
