% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-statistics.R
\name{plotBoxplot}
\alias{plotBoxplot}
\alias{plotDensityplot}
\alias{plotHistogram}
\alias{plotRidgeplot}
\alias{plotViolinplot}
\title{Plot numeric distribution and statistical tests}
\usage{
plotBoxplot(
  object,
  variables,
  phase = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  test_groupwise = NULL,
  test_pairwise = NULL,
  ref_group = NULL,
  step_increase = 0.01,
  vjust = 0,
  display_facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display_points = FALSE,
  pt_alpha = 0.8,
  pt_clr = "black",
  pt_num = 100,
  pt_size = 1.25,
  pt_shape = 21,
  verbose = TRUE,
  ...
)

plotDensityplot(
  object,
  variables,
  phase = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  clrp = "milo",
  clrp_adjust = NULL,
  display_facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  verbose = TRUE,
  ...
)

plotHistogram(
  object,
  variables,
  phase = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  verbose = TRUE,
  ...
)

plotRidgeplot(
  object,
  variables,
  phase = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  alpha = 0.8,
  clrp = "milo",
  clrp_adjust = NULL,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  verbose = TRUE,
  ...
)

plotViolinplot(
  object,
  variables,
  phase = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  test_groupwise = NULL,
  test_pairwise = NULL,
  ref_group = NULL,
  step_increase = 0.01,
  display_facets = TRUE,
  vjust = 0,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display_points = FALSE,
  pt_alpha = 0.8,
  pt_clr = "black",
  pt_num = 100,
  pt_size = 1.25,
  pt_shape = 21,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the choice of color.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{test_groupwise}{Character value or NULL. Specifies the groupwise statistical
test to be conducted. If character, one of \emph{'anova', 'kruskal.test'}. If set
to NULL the testing is skipped.}

\item{test_pairwise}{Character value or NULL. Specifies the pairwise statistical
test to be conducted. If character, one of \emph{'t.test', 'wilcox.test'}. If set
to NULL the testing is skipped.}

\item{ref_group}{Character value or NULL. Specifies the reference group against
which all other groups are compared in the test denoted in \code{test_groupwise}
is conducted. If set to NULL the first group found is taken.}

\item{step_increase}{Numeric value. Denotes the increase in fraction of total
height for every additional comparison to minimize overlap.}

\item{vjust}{Numeric value. Denotes the relative, vertical position of the results of
the test denoted in \code{test.groupwise}. Negative input highers, positive
input lowers the position.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{scales}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to the respective \code{ggplot2::geom_<plot_type>()}
function. E.g. \code{plotViolinplot()} relies on \code{ggplot2::geom_violin()}.}
}
\value{


A ggplot.
}
\description{
These functions display the distribution of numeric
variables for the whole sample or in a comparative manner if argument
\code{across} is specified. \code{plotViolinplot()} and \code{plotBoxplot}
allow for statistical tests such as t-test or ANOVA.
}
