% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-analysis.R
\name{plotCorrplot}
\alias{plotCorrplot}
\title{Plot a correlation plot}
\usage{
plotCorrplot(
  object,
  variable_set,
  method_corr = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  variables_subset = NULL,
  phase = NULL,
  plot_type = "lower",
  display_diagonal = TRUE,
  signif_level = NULL,
  clr_low = "darkred",
  clr_high = "steelblue",
  shape = "tile",
  shape_size = 15,
  size_aes = FALSE,
  display_values = FALSE,
  values_alpha = 0.9,
  values_clr = "black",
  values_digits = 2,
  values_size = 4,
  draw_grid = TRUE,
  grid_clr = "black",
  grid_size = 0.5,
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{method_corr}{Character value. Denotes the correlation method of interest. Either \code{'pearson'}
or \code{'spearman'}.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector, list of character vectors or NULL (see details for
more). Specifies the particular groups of interest the grouping variable specified in argument
\code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the choice of color.}

\item{variables_subset}{Character vector or NULL. Specifies the numeric variables you
want to be included in the correlation plot.

If set to NULL all of them are chosen. You can prefix variables you do NOT want to influence the clustering
with a \emph{'-'}. (Saves writing if there are more variables you are interested in
than variables you are not interested in.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{plot_type}{Character value. Either \emph{'upper'}, \emph{'lower'} or \emph{'complete'}. Specifies
how the correlation matrix is displayed.}

\item{display_diagonal}{Logical value. Specifies if the diagonal of the correlation matrix is supposed to be included
in the correlation plot.}

\item{signif_level}{Numeric value or NULL. If numeric, specifies the minimum significance level a correlation pair
must feature in order to be displayed. Insignificant correlation values are crossed out. Argument \code{shape_size} denotes
the size of the crosses.}

\item{clr_low}{Character value. Specifies the color used for the lower end of the colorspectrum (negative correlations).}

\item{clr_high}{Character value. Specifies the color used for the upper end of the colorspectrum (positive correlations).}

\item{shape}{Character value. Specifies the geometric objects with which to display the correlation pairs. Either \emph{'tile'} or one of
\emph{'circle'} and \emph{'rect'}. In the latter two cases the size of the geometric objects can be used to emphasize the correlation
in addtion to the colorspectrum if \code{size_aes} is set to TRUE.}

\item{shape_size}{Numeric value. Specifies the size with which to display the circles and rectangulars. If \code{shape_aes} is set to TRUE
it Specifies the maximum size possible.}

\item{size_aes}{Logical value. If set to TRUE the size of the circles or rectangulars is used to display the correlation.}

\item{display_values}{Logical value. If set to TRUE the actual correlation values are printed on the geometric objects.}

\item{values_alpha}{Numeric value. Specifies the transparency of the values.}

\item{values_clr}{Character value. Specifies the color of the values.}

\item{values_digits}{Numeric value. Specifies the number of digits to which the correlation values are rounded before beeing displayed.}

\item{values_size}{Numeric value. Specifies the size of the values.}

\item{draw_grid}{Logical value. If set to TRUE a grid is drawn separating all circles or shapes.}

\item{grid_clr}{Character value. Specifies the color of the grid.}

\item{grid_size}{Numeric value. Specifies the thickness of the grid's lines.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}
}
\value{


A ggplot.
}
\description{
Visualizes the correlation results computed by \code{correlateAll()} and \code{correlateAcross()}.
}
