% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R,
%   R/clustering-hierarchical.R
\name{plotDendrogram}
\alias{plotDendrogram}
\alias{plotDendrogram,Cypro-method}
\title{Plot dendrograms}
\usage{
plotDendrogram(object, ...)

\S4method{plotDendrogram}{Cypro}(
  object,
  phase,
  fset = "all_features",
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle",
  direction = "bt",
  branch_color = "black",
  branch_size = 1,
  display_labels = FALSE,
  labels_angle = 90,
  labels_hjust = 0,
  labels_nudge = 0.01,
  labels_size = 3,
  labels_vjust = 0.5,
  display_title = FALSE,
  clrp = "milo",
  clrp_adjust = NULL,
  simple = FALSE,
  facet_with = "grid",
  nrow = NULL,
  ncol = NULL,
  ...
)

\S4method{plotDendrogram}{Cypro}(
  object,
  phase,
  fset = "all_features",
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle",
  direction = "bt",
  branch_color = "black",
  branch_size = 1,
  display_labels = FALSE,
  labels_angle = 90,
  labels_hjust = 0,
  labels_nudge = 0.01,
  labels_size = 3,
  labels_vjust = 0.5,
  display_title = FALSE,
  clrp = "milo",
  clrp_adjust = NULL,
  simple = FALSE,
  facet_with = "grid",
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{methods_dist}{Character vector. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{methods_aggl}{Character vector. Specifies the agglomerative
methods of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{k}{Numeric value. Denotes the number of clusters. Must be bigger than 1. Must
not be bigger than the number of observations of the data set.}

\item{h}{Numeric value. Denotes the height at which the dendrogram is cut.}

\item{type}{Character value. Indicates the shape of the dendrogram.
Input \emph{'rectangle'} will draw rectangular lines, while \emph{'triangle'} will draw triangular lines.}

\item{display_labels}{Lgoical value. Indicates if labels are displayed.}

\item{labels_angle}{Numeric value. The angle with which labels are displayed.}

\item{labels_hjust}{Numeric values. Adjust labels positioning.}

\item{labels_size}{Numeric value. The size with which labels are displayed.}

\item{labels_vjust}{Numeric values. Adjust labels positioning.}

\item{display_title}{Logical values. Indicate if legend or
title is displayed.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector. Can be used to adjust colors
with which groups are displayed. Names indicate groups, values
indicate the colors with which specific groups are supposed to be displayed.}

\item{simple}{Logical value. If TRUE, the dendrogram is plotted with \code{base::plot()}.
This is way quicker but does not allow for ggplot2 specific adjustments.}

\item{facet_with}{Character value. Either 'grid' or 'wrap'. Specifies the function
with which the plot-facetting is created. If the number of input combinations for
\code{methods_dist} and \code{methods_aggl} length 2 or bigger and \code{facet_with} = \emph{'wrap'}
\code{ggplot2::facet_wrap()} is used. Else \code{ggplot2::facet_grid()} is used.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}
}
\value{
A ggplot or a base plot.
}
\description{
Plots a dendrogram with either ggplot or base plot.
}
