% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics.R
\name{plotDensityplot}
\alias{plotDensityplot}
\alias{plotDensityplot,CyproScreening-method}
\alias{plotDensityplot,CyproTimeLapse-method}
\alias{plotDensityplot,CyproTimeLapseMP-method}
\title{Plot numeric distribution and statistical tests}
\usage{
plotDensityplot(object, ...)

\S4method{plotDensityplot}{CyproScreening}(
  object,
  features,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  display_facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  verbose = TRUE,
  ...
)

\S4method{plotDensityplot}{CyproTimeLapse}(
  object,
  features,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  display_facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  verbose = TRUE,
  ...
)

\S4method{plotDensityplot}{CyproTimeLapseMP}(
  object,
  features,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  display_facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  phase = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to the respective \code{ggplot2::geom_<plot_type>()}
function. E.g. \code{plotViolinplot()} relies on \code{ggplot2::geom_violin()}.}

\item{features}{Character vector. Denotes the features of interest. \emph{Features} are numeric
variables in the cell data (slot @cdata) of the \code{Cypro} object.

Use \code{getFeatureNames()} to obtain all valid input options.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector, list of character vectors or NULL (see details for
more). Specifies the particular groups of interest the grouping variable specified in argument
\code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the choice of color.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{scales}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{


A ggplot.
}
\description{
These functions display the distribution of numeric
variables for the whole sample or in a comparative manner if argument
\code{across} is specified. \code{plotViolinplot()} and \code{plotBoxplot}
allow for statistical tests such as t-test or ANOVA.
}
