% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plot dimensional reduction}
\usage{
plotDimRed(
  object,
  dim_red = "umap",
  color_by = NULL,
  pt_size = 1,
  pt_alpha = 0.9
)
}
\description{
Visualizes the dimensional reduction method of choice.
}
