% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot descriptive statistics}
\usage{
plotDistribution(
  object,
  variables = "all",
  across = "cl_condition",
  across_subset = NULL,
  phase = "first_tmt",
  plot_type = "boxplot",
  binwidth = 1,
  display_points = FALSE,
  pt_size = 1.2,
  pt_alpha = 0.8,
  pt_color = "black",
  n_cells = 100,
  shape_to = NULL,
  test_pairwise = "none",
  test_groupwise = "none",
  ref_group,
  clrp = "milo",
  ...,
  pretty_names = TRUE,
  verbose = TRUE
)
}
\description{
These functions are deprecated in favor of \code{plotDensityplot(),
plotHistogram(), plotRidgplot(), plotBoxplot() and plotViolinplot()}.
}
