% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotDistributionDiscrete}
\alias{plotDistributionDiscrete}
\title{Distribution of discrete features}
\usage{
plotDistributionDiscrete(
  object,
  phase = "first_tmt",
  features,
  feature_compare = NULL,
  clrp = "milo",
  position = "fill",
  ...
)
}
\description{
This function is deprecated in favor of \code{plotBarchart()}.
}
