% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-pam.R
\name{plotPamMedoids}
\alias{plotPamMedoids}
\title{Plot medoid results}
\usage{
plotPamMedoids(
  object,
  variable_set,
  k,
  facet_by = "cluster",
  color_by = "variables",
  method_pam = NULL,
  phase = NULL,
  clrp = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{method_pam}{Character vector (or value see details for more.) Denotes the algorithm of interest.
Valid input options are \emph{'euclidean'} and \emph{'manhattan'}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{


A ggplot.
}
\description{
Visualizes the values of the observations that were identified
as the medoids giving insight in the overall qualities that constitute the
clusters.
}
