% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{plotPca}
\alias{plotPca}
\alias{plotTsne}
\alias{plotUmap}
\title{Plot dimensional reduction results}
\usage{
plotPca(
  object,
  variable_set,
  phase = NULL,
  color_by = NULL,
  color_aes = "fill",
  pt_clrp = NULL,
  pt_clrsp = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_fill = NULL,
  pt_size = NULL,
  clrp_adjust = NULL,
  add_ons = list(ggplot2::labs(fill = NULL, color = NULL)),
  ...
)

plotTsne(
  object,
  variable_set,
  phase = NULL,
  color_by = NULL,
  color_aes = "fill",
  pt_clrp = NULL,
  pt_clrsp = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_fill = NULL,
  pt_size = NULL,
  clrp_adjust = NULL,
  add_ons = list(ggplot2::labs(fill = NULL, color = NULL)),
  ...
)

plotUmap(
  object,
  variable_set,
  phase = NULL,
  color_by = NULL,
  color_aes = "fill",
  pt_clrp = NULL,
  pt_clrsp = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_fill = NULL,
  pt_size = NULL,
  clrp_adjust = NULL,
  add_ons = list(ggplot2::labs(fill = NULL, color = NULL)),
  ...
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{color_by}{Character value or NULL. If character, denotes either the numeric- or grouping variable whoose values
are displayed by color. If set to NULL the color is specified by the argument \code{pt_clr}.}

\item{color_aes}{Character value. Only relevant if \code{color_by} is specified.
Denotes the aesthetic with which colors are displayed. Either \emph{'color'} or \emph{'fill'}. Depending on that as well as the input for argument
\code{pt_shape} the design of the geometric objects (points) varies according to the rules
of the ggplot2-framework.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}
}
\value{


A ggplot.
}
\description{
Visualizes dimensional reduction in a scatterplot.
}
