% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-kmeans.R
\name{plotScreeplot}
\alias{plotScreeplot}
\title{Plot a scree plot}
\usage{
plotScreeplot(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{clr}{Character value. Denotes the color of the columns.}
}
\value{


A ggplot.
}
\description{
Visualizes the within-group sum of squares of every calculated kmeans
clustering.
}
\details{
If \code{method_kmeans} is a vector a subplot for all valid input values
is displayed via \code{ggplot2::facet_wrap()}.
}
