% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-pam.R
\name{plotSilWidths}
\alias{plotSilWidths}
\alias{plotSilWidths,Cypro-method}
\alias{plotSilWidths,CyproTimeLapseMP-method}
\title{Plot sil-width data}
\usage{
plotSilWidths(object, ...)

\S4method{plotSilWidths}{Cypro}(
  object,
  ks,
  method_pam = "euclidean",
  fset = "all_features",
  clrp = "milo",
  ncol = NULL,
  nrow = NULL,
  verbose = TRUE
)

\S4method{plotSilWidths}{CyproTimeLapseMP}(
  object,
  phase,
  ks,
  method_pam = "euclidean",
  fset = "all_features",
  clrp = "milo",
  ncol = NULL,
  nrow = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{method_pam}{Character value. Specifies the pam-method of interest.
Use \code{validMethodsPam()} to obtain all valid input options.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A ggplot.
}
\description{
Plots information about the silhouette width of
every observation in different clustering results with method \emph{pam}.
}
