% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics.R
\name{plotStatisticsInteractive}
\alias{plotStatisticsInteractive}
\alias{plotStatisticsInteractive,CyproScreening-method}
\alias{plotStatisticsInteractive,CyproTimeLapse-method}
\alias{plotStatisticsInteractive,CyproTimeLapseMP-method}
\title{Plot statistics related plots interactively}
\usage{
plotStatisticsInteractive(object, ...)

\S4method{plotStatisticsInteractive}{CyproScreening}(object, ...)

\S4method{plotStatisticsInteractive}{CyproTimeLapse}(object, ...)

\S4method{plotStatisticsInteractive}{CyproTimeLapseMP}(object, phase = NULL, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\description{
Opens an interactive application in which statistic related
plots like violinplots, boxplots as well as statistical tests across
groups can be conducted.
}
