% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-lapse.R
\name{plotTimeHeatmap}
\alias{plotTimeHeatmap}
\title{Visualize changes of cell characteristics over time}
\usage{
plotTimeHeatmap(
  object,
  variable,
  across = "condition",
  across_subset = NULL,
  relevel = NULL,
  phase = "all",
  n_cells = 100,
  clrp = NULL,
  clrp_adjust = NULL,
  color = NA,
  colors = viridis::viridis(15),
  smooth = TRUE,
  smooth_span = 0.25,
  arrange_rows = "maxima",
  set_seed = NULL,
  verbose = TRUE,
  ...,
  in_shiny = FALSE
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the choice of color.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{n_cells}{Numeric calue. Determines the number of cells that are randomly chosen from
every group to be displayed. Useful to keep plots insightful and aesthetically pleasing.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{colors}{A vector of colors that is used as the heatmap's color spectrum.}

\item{smooth}{Logical. If set to TRUE the values are smoothed.}

\item{smooth_span}{NUmeric value. Denotes the smoothing span used.}

\item{arrange_rows}{Character value. Either \emph{'maxima'} or \emph{'none'}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{pheatmap::pheatmap()}.}
}
\description{
Uses the design of a heatmap to visualize feature
dynamics over time across a grouping variable of interest.
}
