% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-lapse.R
\name{plotTimeLineplot}
\alias{plotTimeLineplot}
\title{Visualize changes of cell characteristics over time}
\usage{
plotTimeLineplot(
  object,
  variables,
  across = "condition",
  across_subset = NULL,
  summarize_with = "mean",
  phase = NULL,
  linesize = 1,
  scales = "free_y",
  nrow = NULL,
  ncol = NULL,
  smooth = TRUE,
  smooth_method = "loess",
  smooth_span = 0.25,
  smooth_se = FALSE,
  display_vline = TRUE,
  vline_clr = "lightgrey",
  vline_type = "dashed",
  clrp = "milo",
  verbose = TRUE,
  ...,
  in_shiny = FALSE
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variables}{Character vector. Denotes the track variables whoose summary you
want to display over time. Use \code{getTrackVariableNames()} to obtain all
valid input options.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{summarize_with}{Character value. One of \emph{'mean'}, \emph{'median'},
\emph{'max'} or \emph{'min'}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{linesize}{Numeric value. Denotes the size of the lines drawn.}

\item{scales}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{smooth}{Logical. If set to TRUE the values are smoothed.}

\item{smooth_method}{Character value. Denotes the model type used to display the line. Defaults to
\emph{'lm'} (linear model). Given to argument \code{method} of function \code{ggplot2::geom_smooth()}.}

\item{smooth_span}{NUmeric value. Denotes the smoothing span used.}

\item{smooth_se}{Logical. If set to TRUE the standard error will be displayed.}

\item{display_vline}{Logical value. If set to TRUE and more than one phase
is referred to a vertical line is displayed hihglighting the phase separation.}

\item{vline_clr}{Character value. Color of the vertical line.}

\item{vline_type}{Character value. Defaults to \emph{'dashed'}.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{


A ggplot.
}
\description{
Uses the design of a lineplot to visualize feature
dynamics over time across a grouping variable of interest.
}
