% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/dimensional-reduction.R
\name{plotUMAP}
\alias{plotUMAP}
\alias{plotUMAP,Cypro-method}
\alias{plotUMAP,CyproTimeLapseMP-method}
\title{Plot UMAP}
\usage{
plotUMAP(object, ...)

\S4method{plotUMAP}{Cypro}(
  object,
  fset = "all_features",
  alpha_by = NULL,
  color_by = NULL,
  shape_by = NULL,
  size_by = NULL,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_fill = "black",
  pt_shape = 19,
  pt_size = 1,
  color_aes = "color",
  clrp = "milo",
  clrp_adjust = NULL,
  clrsp = "inferno",
  ...
)

\S4method{plotUMAP}{CyproTimeLapseMP}(
  object,
  phase,
  fset = "all_features",
  alpha_by = NULL,
  color_by = NULL,
  shape_by = NULL,
  size_by = NULL,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_fill = "black",
  pt_shape = 19,
  pt_size = 1,
  color_aes = "color",
  clrp = "milo",
  clrp_adjust = NULL,
  clrsp = "inferno",
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{alpha_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{color_by}{Character value or NULL. If character, specifies the
variable that is displayed by color.}

\item{shape_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{size_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{pt_alpha}{Numeric value. Specifies the transparency of points.}

\item{pt_color}{Character value. Specifies the color with which all points are displayed .}

\item{pt_shape}{Numeric or character value. If numeric, the respective
shape is taken for all points. If character, the respective variable is
mapped to the shape-aesthetic.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{color_aes}{Either \emph{'color'} or \emph{'fill'}. Specifies the
aesthetic that is used to visualize the variable that is specified in \code{color_by}.
Ignored if \code{color_by} is NULL.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector. Can be used to adjust colors
with which groups are displayed. Names indicate groups, values
indicate the colors with which specific groups are supposed to be displayed.}

\item{clrsp}{Character value. Specfies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A ggplot.
}
\description{
Plots UMAP results in a scatterplot.
}
