% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWellPlate.R
\name{plotWellPlate}
\alias{plotWellPlate}
\alias{plotWellPlate,layout_df-method}
\title{Visualize well plates}
\usage{
plotWellPlate(
  object,
  clr_by = "file_status",
  plot_type = "tile",
  well_alpha = 0.9,
  well_size = 13.5,
  display_border = TRUE,
  border_clr = "black",
  border_size = 1,
  display_edge = TRUE,
  edge_clr = "black",
  edge_size = 1,
  display_labels = TRUE,
  labels_clr = "black",
  clrp = "milo",
  clrp_adjust = NULL,
  ncol = NULL,
  nrow = NULL,
  ...
)

\S4method{plotWellPlate}{layout_df}(
  object,
  clr_by = "file_status",
  plot_type = "tile",
  well_alpha = 0.9,
  well_size = 13.5,
  display_border = TRUE,
  border_clr = "black",
  border_size = 1,
  display_edge = TRUE,
  edge_clr = "black",
  edge_size = 1,
  display_labels = TRUE,
  labels_clr = "black",
  clrp = "milo",
  clrp_adjust = NULL,
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Plots a well plate.
}
