% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{prepareDataLoading}
\alias{prepareDataLoading}
\alias{prepareDataLoading,Cypro-method}
\title{Data loading - Prepare data loading}
\usage{
prepareDataLoading(
  object,
  well_plate,
  dir_input,
  valid_filetypes,
  recursive = FALSE,
  in_shiny = FALSE
)

\S4method{prepareDataLoading}{Cypro}(
  object,
  well_plate,
  dir_input,
  valid_filetypes,
  recursive = FALSE,
  in_shiny = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}
}
\description{
First of three data loading steps. Assigns a folder/file to every
well plate. See details for more.
}
\details{
Has to be called iteratively for every well plate.

Loading modality \emph{by_roi} or \emph{by_well}:

Sets \code{dir_input} in the slot @directory of well plate
denoted in \code{well_plate}. Screens the folder (and subfolders if \code{recursive} = TRUE)
for files that end with the valid file endings denoted in \code{valid_filetypes}.
Then it uses regular expressions to assign the remaining files to the wells / well rois.
The \emph{file_status} is then declared: Every well / well roi that afterwards contain one assigned
file is considered \emph{Valid}. Wells / well rois with more matching files are considered \emph{Ambiguous}
and those with no matching files are considered \emph{Missing}. For every valid file an object
of class \code{DataFile} is created in the list of slot @files of the \code{WellPlate} object
containing the assigned file directory.

Afterwards \code{loadDataFiles()} uses the prepared \code{DataFile} objects to read every file.
}
\seealso{
\code{readataFiles()}, \code{processData()}
}
