% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{processData}
\alias{processData}
\alias{processData,CyproScreening-method}
\alias{processData,CyproTimeLapse-method}
\title{Data loading - Process data}
\usage{
processData(object, verbose = TRUE)

\S4method{processData}{CyproScreening}(object, verbose = TRUE)

\S4method{processData}{CyproTimeLapse}(object, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Third of three data loading steps A pipeline that does all
the processing after the data has been loaded via \code{loadDataFiles()}.
See details fore more.
}
\details{
Depending on the method the pipeline consists of the following steps:

\code{transferData()}: All files that have been loaded with \code{loadDataFiles()} without
any error are merged to a data.frame. Depending on the loading modality - \emph{by_roi},
\emph{by_well}, \emph{by_well_plate} - the missing ID variables are added to the
data.frame. It is then split into the functional parts the \code{Cdata} object
that corresponds to the experiment design determines. The \code{Cdata} object
is then set in slot @cdata of the \code{Cypro} object.

\code{completeTracksDf()}: The method for time lapse experiments completes the data.frame by adding
observations were \emph{cell_id}+\code{frame_num} combinations are missing (due
to failed recognition of a cell in one ore more frames). Values of numeric variables
are filled with NAs for these observations. Which observation was added is indicated
by the variable \emph{frame_added}.

\code{computeModuleVariabels()}: Computable, module related variables that were
not assigned during \code{assignVariables()} are computed and added to the feature
data.frame \code{CyproScreening} / tracks data.frame
\code{CyproTimeLapse()}.

\code{summarizeDataByCellId()}: The method for time lapse experiments
summarizes the time lapse data by cell ID to create the stats data.frame.

\code{summarizeModuleVariables()}: The method for time lapse experiments
additionally summarizes all \code{SummaryVariables} of the used analysis modules
(e.g. \emph{mgr_efficiency}) and adds them to the stats data.frame.

The output is a \code{Cypro} object that is ready for analysis.
}
