% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-initiation.R
\name{processDataX}
\alias{processDataX}
\title{Object initiation: Step 3}
\usage{
processDataX(
  object,
  summarize_with = c("max", "min", "mean", "median"),
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Processes the data and constructs all needed slots. Afterwards the
cypro object is set for all subsequent analysis and visualization steps.
}
