% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-input-options.R
\name{protectedVariables}
\alias{protectedVariables}
\title{Protected variables}
\usage{
protectedVariables()
}
\value{
Character vector.
}
\description{
Obtain a character vector of names from variables
that must not be renamed and whose content must not be changed.
}
