% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{quality_check_summary_shiny}
\alias{quality_check_summary_shiny}
\title{Summarize tracking quality}
\usage{
quality_check_summary_shiny(track_df)
}
\arguments{
\item{track_df}{A track data.frame}
}
\description{
Summarize tracking quality
}
