% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{read_data_files_shiny}
\alias{read_data_files_shiny}
\title{Reads cell track data}
\usage{
read_data_files_shiny(directory, progress_n, progress, object)
}
\arguments{
\item{directory}{Character value. The directory from which to read the data.frame.
of \code{.x} contains.}

\item{progress_n}{Numeric value. Indicates the iterating progress.}

\item{progress}{A shiny - R6 progress object.}
}
\value{
The read in data.frame with three additional informative columns: \emph{well_image, condition}
and \emph{cell_line}.
}
\description{
A function to be used for argument \code{.f} of \code{purrr::map_df}.
Calls the respective reading function depending on the directory's ending.
}
