% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{read_table_shiny}
\alias{read_table_shiny}
\title{wrapper around several table reading functions}
\usage{
read_table_shiny(directory, ...)
}
\description{
wrapper around several table reading functions
}
