% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier-detection.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove outliers from object}
\usage{
removeOutliers(
  object,
  method_outlier,
  new_name = "outlier_removed",
  phase = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method_outlier}{Character vector. Specifies the method/algorithm of interest.

Use \code{validOutlierDetectionMethods()} to obtain all valid input options.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Removes the data from the cell ids detected by \code{detectOutliers()} from
the object. As this changes the data set all clustering-, correlation- and dimensional
reduction results are reset.
}
\details{
Cell are removed via a reversed application of \code{subsetByCellId()}.
You can therefore keep track of conducted outlier removal by using function \code{printSubsetHistory()}
which prints summary text on how, when and why you subsetted the cypro object.
}
\seealso{
detectOutliers()
}
