% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{renameGroups}
\alias{renameGroups}
\title{Rename groups}
\usage{
renameGroups(object, grouping_variable = NULL, ..., phase = NULL)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{grouping_variable}{Character value. The name of the grouping variable
whose groups you want to rename.}

\item{...}{The groups to be renamed specified according to the following
syntax: \emph{'new_cluster_name'} \code{=} \emph{'old_cluster_name'}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Allows to rename particular groups within a grouping variable
of the cypro object's cell data.
}
\details{
Renaming groups of variables \emph{well_plate_name, well_plate_index, well}
and \emph{well_image} is not allowed and will result in an error.

Use \code{getGroupNames()} to check if renaming resulted in the desired output.
}
