% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-content.R
\name{renameTracksDf}
\alias{renameTracksDf}
\alias{renameTracksDfWith}
\title{Rename track variables}
\usage{
renameTracksDf(object, ...)

renameTracksDfWith(object, ...)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{...}{Input for function \code{dplyr::rename_with()}. This includes the arguments
\code{.fn} and \code{.cols}.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Implementations of the functions \code{dplyr::rename()} and
\code{dplyr::rename_with()} that can be used to rename the variables of the cell track
data.frame.

\code{rename*Df()} changes the names of individual variables using \emph{new_name} = \emph{old_name}
syntax. \code{rename*DfWith()} renames variables with a function specified in argument \code{.fn}.
}
\details{
Renaming the tracks data.frame affects all slots of the cypro object that refer to
track variables (correlation analysis, variable sets etc.). All are renamed according to
the input.

Use \code{getTrackVariableNames()} to check if renaming resulted in
the desired output.
}
\note{
Make sure not to rename protected variables. Use the function \code{protectedVariables()}
to obtain variable names that must not be changed and must not be added. Doing
so will result in erroneous analysis results as many functions rely on these
variables to carry specific meaning.
}
