% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{RequiredVariable-class}
\alias{RequiredVariable-class}
\alias{RequiredVariable}
\title{The RequiredVariable Class}
\description{
Subclass of \code{DataVariable}. See details for more information.
}
\details{
A variable is considered to be required
if it can not be computed or summarized and thus represents the core of an
analysis module. An analysis module can only be used if all of it's required variables
exist in the input data. In case of the migration analysis the variables \emph{x_coords}
and \emph{y_coords} must be part of the input data as they can not be computed based
on other variables.
}
\seealso{
\code{AssignableVariable}, \code{ComputableVariable}, \code{SummarizableVariable}
}
