% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{run_dim_red}
\alias{run_dim_red}
\title{Title}
\usage{
run_dim_red(
  object,
  variable_set,
  phase = NULL,
  method_dim_red = "pca",
  force = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{force}{Logical value. Needs to be set to TRUE if you want to overwrite an already existing
set up or already existing results.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Title
}
