% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{runPca}
\alias{runPca}
\alias{runTsne}
\alias{runUmap}
\title{Compute dimensional reductions}
\usage{
runPca(object, variable_set, phase = NULL, force = FALSE, verbose = NULL, ...)

runTsne(object, variable_set, phase = NULL, force = FALSE, verbose = NULL, ...)

runUmap(object, variable_set, phase = NULL, force = FALSE, verbose = NULL, ...)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{force}{Logical value. Needs to be set to TRUE if you want to overwrite an already existing
set up or already existing results.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{variables_subset}{Character vector or NULL. Specifies the numeric variables the dimensional reduction
algorithms will include.

If set to NULL all of them are chosen. You can prefix variables you do NOT want to influence the clustering
with a \emph{'-'}. (Saves writing if there are more variables you are interested in
than variables you are not interested in.)

Use \code{getNumericVariableNames()} to obtain all valid input options.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Reduces the dimensions of all specified numeric variables using the respective
algorithm.
}
