% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{saveCyproObject}
\alias{saveCyproObject}
\alias{saveCyproObject,Cypro-method}
\title{Save \code{Cypro} object}
\usage{
saveCyproObject(object, ...)

\S4method{saveCyproObject}{Cypro}(object, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An invisible TRUE if saving occured without any error.
}
\description{
Savesa the object under its storage directory.
}
