% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/miscellaneous-n-z.R
\name{scaleData}
\alias{scaleData}
\alias{scaleData,CyproScreening-method}
\alias{scaleData,CyproTimeLapse-method}
\alias{scaleData,CyproTimeLapseMP-method}
\title{Scale data}
\usage{
scaleData(object, ...)

\S4method{scaleData}{CyproScreening}(object, verbose = TRUE, ...)

\S4method{scaleData}{CyproTimeLapse}(object, verbose = TRUE, ...)

\S4method{scaleData}{CyproTimeLapseMP}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Scales numeric data of the \code{Cypro} object. In case
of screening experiments the numeric data from slot @features is taken.
In case of time lapse experiments the data form slot @features stats is taken.
}
