% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-a-m.R, R/miscellaneous-n-z.R
\name{deselectWells}
\alias{deselectWells}
\alias{deselectWells.layout_df}
\alias{selectWells}
\alias{selectWells.layout_df}
\title{Select and deselect wells}
\usage{
deselectWells(df, wells = NULL)

\method{deselectWells}{layout_df}(df, wells = NULL)

selectWells(df, wells = NULL)

\method{selectWells}{layout_df}(df, wells = NULL)
}
\arguments{
\item{df}{Data.frame of class \code{layout_df}.}

\item{wells}{Character vector or NULL. If character, denotes wells which
are supposed to be (de-)selected. If NULL, all wells are considered.}
}
\value{
A layout data.frame of the same class as the input.
}
\description{
\itemize{
\item{\code{selectWells()}}{ Selects wells by changing the value
of variable \emph{selected} of all wells found in input for argument
\code{wells} to TRUE.}
\item{\code{deselectWells()}}{ De selects wells by changing the value
of variable \emph{selected} of all wells found in input for argument
\code{wells} to FALSE.}
}

The selection status of wells that are not denoted in the input for argument \code{wells} stays
as is.
}
