% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{set_up_vdata}
\alias{set_up_vdata}
\title{Computes variable summaries}
\usage{
set_up_vdata(object, verbose = TRUE)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Computes variable summaries
}
