% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setAnalysisAspect}
\alias{setAnalysisAspect}
\alias{setAnalysisAspect,Cypro-method}
\alias{setAnalysisAspect,CyproTimeLapseMP-method}
\title{Set analysis aspect}
\usage{
setAnalysisAspect(object, ...)

\S4method{setAnalysisAspect}{Cypro}(object, analysis_aspect, fset_name)

\S4method{setAnalysisAspect}{CyproTimeLapseMP}(object, analysis_aspect, fset_name, phase)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
The input object.
}
\description{
Sets analysis aspect and removes data.
}
