% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setAnalysisList}
\alias{setAnalysisList}
\alias{setAnalysisList,Cypro-method}
\title{Set empty analysis list}
\usage{
setAnalysisList(object, ...)

\S4method{setAnalysisList}{Cypro}(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
The input object.
}
\description{
Sets empty analysis list in slot @analysis.
}
