% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setCdata}
\alias{setCdata}
\alias{setCdata,CyproScreening-method}
\alias{setCdata,CyproTimeLapse-method}
\alias{setCdata,CyproTimeLapseMP-method}
\title{Set cdata slot}
\usage{
setCdata(object, df, verbose = TRUE, ...)

\S4method{setCdata}{CyproScreening}(object, df, verbose = TRUE, ...)

\S4method{setCdata}{CyproTimeLapse}(object, df, verbose = TRUE)

\S4method{setCdata}{CyproTimeLapseMP}(object, df, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{df}{A data.frame that contains the whole data set of the experiment
named according to the \code{cypro} naming convention. Can (should) be obtained
via \code{getMergedDf()} after loading all the data.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{directory}{Character value. The directory under which
the object is supposed to be stored via \code{saveCyproObject()}.}
}
\description{
Sets slot @cdata of the \code{Cypro} object by splitting
the data.frame provided with \code{df} into its functional subcategories.
}
