% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setCellDf}
\alias{setCellDf}
\title{Set cell data.frame}
\usage{
setCellDf(object, slot, df, phase)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Set cell data.frame
}
