% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setCellIDs}
\alias{setCellIDs}
\alias{setCellIDs,Cypro-method}
\title{Set cell IDs}
\usage{
setCellIDs(object, ...)

\S4method{setCellIDs}{Cypro}(object, cell_ids)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{cell_ids}{Character vector. All cell IDs.}
}
\description{
Sets a vector of all cell IDs the \code{Cypro} object
contains.
}
