% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setCellLine}
\alias{setCellLine}
\alias{setCellLine.layout_df}
\title{Set layout cell line}
\usage{
setCellLine(df, ...)

\method{setCellLine}{layout_df}(df, cell_line, in_shiny = FALSE, verbose = TRUE)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}

\item{cell_line}{Character value, NA or NULL. The cell line seeded in the wells denoted in
\code{selected_wells}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame.
}
\description{
Sets the cell line of all wells that are selected
according to variable \emph{selected} of the \code{layout_df}.
}
