% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setClusterConv}
\alias{setClusterConv}
\alias{setDimRedConv}
\alias{setCorrConv}
\title{Set analysis objects}
\usage{
setClusterConv(object, cluster_object, method, variable_set, phase)

setDimRedConv(object, dim_red_object, method, variable_set, phase)

setCorrConv(object, corr_object, variable_set, phase)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Sets a convenient object and empties its data slot.
}
