% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setCondition}
\alias{setCondition}
\alias{setCondition.layout_df}
\alias{setCondition.layout_df_mp}
\title{Set layout condition}
\usage{
setCondition(df, ...)

\method{setCondition}{layout_df}(df, condition, in_shiny = FALSE, verbose = TRUE)

\method{setCondition}{layout_df_mp}(df, condition, in_shiny = FALSE, verbose = TRUE)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}

\item{condition}{Character value, NA, or NULL. The condition of the well. If multiple phase
experiment a character vector of length equal to the number of phases of
the experiment design.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame.
}
\description{
Sets the condition of all wells that are selected
according to variable \emph{selected} of the \code{layout_df}.
}
