% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R
\name{setData}
\alias{setData}
\title{Set data and key variables}
\usage{
setData(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
The input object.
}
\description{
Sets the objects slot @data and slot @key_name and makes
sure that the key variable exists and is valid.
}
\details{
If \code{key_name} is NULL the key variable is constructed
as a combination of the input data.frames rownames and the string
provided with \code{key_prefix}. If the the datas rownames are NULL
or contain \emph{""} the rownumbers are used instead. The newly constructed
key variable is named \emph{data_ids}.
}
