% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setDataStatus}
\alias{setDataStatus}
\alias{setDataStatus.layout_df}
\alias{setInfoStatus.layout_df_mp}
\title{Set layout data status}
\usage{
setDataStatus(df)

\method{setDataStatus}{layout_df}(df)

\method{setInfoStatus}{layout_df_mp}(df)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}
}
\value{
A data.frame.
}
\description{
Sets the variable \emph{data_status} of the layout
data.frame according to the variables \emph{info_status} and
\emph{file_status}.
}
