% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setDefaultInstructions}
\alias{setDefaultInstructions}
\title{Set cypro default}
\usage{
setDefaultInstructions(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Set cypro default
}
