% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setExperimentDesign}
\alias{setExperimentDesign}
\alias{setExperimentDesign,Cypro-method}
\title{Set experiment design}
\usage{
setExperimentDesign(object, exp_design)

\S4method{setExperimentDesign}{Cypro}(object, exp_design)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{exp_design}{An object of class \code{ExperimentDesign}.}
}
\value{
The input object.
}
\description{
Sets the experiment design of the \code{Cypro} object.
}
