% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setExperimentName}
\alias{setExperimentName}
\alias{setExperimentName,Cypro-method}
\title{Set experiment name}
\usage{
setExperimentName(object, name)

\S4method{setExperimentName}{Cypro}(object, name)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{name}{Character value. The experiment name.}
}
\value{
The input object.
}
\description{
Sets the content for slot @experiment in the \code{Cypro} object and
all S4 objects that have an @experiment slot.
}
