% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setFeatureDf}
\alias{setFeatureDf}
\alias{setFeatureDf,CyproScreening-method}
\title{Set feature data.frame}
\usage{
setFeatureDf(object, df)

\S4method{setFeatureDf}{CyproScreening}(object, df)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
The input object.
}
\description{
Set feature data.frame
}
