% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setImageDirDf}
\alias{setImageDirDf}
\alias{setImageDirDf,Cypro-method}
\alias{setImageDirDf,CyproTimeLapse-method}
\title{Sets image directory data.frame}
\usage{
setImageDirDf(object, df = NULL, force = FALSE)

\S4method{setImageDirDf}{Cypro}(object, df = NULL, force = FALSE)

\S4method{setImageDirDf}{CyproTimeLapse}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{df}{The image directory data.frame.}

\item{force}{Logical value. Needs to be set to TRUE if you want to overwrite an already existing
set up or already existing results.}
}
\value{
The input object.
}
\description{
Sets an \bold{empty} data.frame in slot @image_directories
of the \code{Cypro} object. Do not confuse with \code{setImageDirectories()}.
}
