% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setInfoStatus}
\alias{setInfoStatus}
\alias{setInfoStatus.layout_df}
\alias{setInfoStatus.layout_df_mp}
\title{Sets info status}
\usage{
setInfoStatus(df)

\method{setInfoStatus}{layout_df}(df)

\method{setInfoStatus}{layout_df_mp}(df)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}
}
\value{

}
\description{
Sets variable \emph{info_status} of the \code{layout_df}
to one of \emph{'Missing', 'Incomplete', 'Complete'} depending on the
information found in variables \emph{cell_line} and \emph{condition}.
}
