% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setInputExample}
\alias{setInputExample}
\alias{setInputExample,ExperimentDesign-method}
\alias{setInputExample,Cypro-method}
\title{Set input example}
\usage{
setInputExample(object, df, dir)

\S4method{setInputExample}{ExperimentDesign}(object, df, dir)

\S4method{setInputExample}{Cypro}(object, df, dir)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{df}{The example data.frame}

\item{dir}{Character value. The file directory of the loaded data.frame.}
}
\value{
The input object.
}
\description{
Sets the input example during \code{assignVariables()} in
slot @example_df of the S4-class \code{ExperimentDesign}.
}
