% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setLayoutAttributes}
\alias{setLayoutAttributes}
\title{Set layout data.frame attributes}
\usage{
setLayoutAttributes(df, attr)
}
\arguments{
\item{df}{A data.frame of class \code{layout_df}.}

\item{attr}{A named list to be added to the attributes of input data.frame.}
}
\value{
A data.frame.
}
\description{
Safely sets layout data.frame specific attributes.
}
