% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setLayoutDf}
\alias{setLayoutDf}
\alias{setLayoutDf,WellPlate-method}
\alias{setLayoutDf,ExperimentDesign-method}
\alias{setLayoutDf,Cypro-method}
\title{Set layout data.frame}
\usage{
setLayoutDf(object, layout_df, ...)

\S4method{setLayoutDf}{WellPlate}(object, layout_df, ...)

\S4method{setLayoutDf}{ExperimentDesign}(object, layout_df, well_plate)

\S4method{setLayoutDf}{Cypro}(object, layout_df, well_plate)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}

\item{df}{Data.frame of class \code{layout_df}.}
}
\description{
Places the well plate layout data.frame in the correct slot
of the object provided.
}
