% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setLoadingModality}
\alias{setLoadingModality}
\alias{setLoadingModality,Cypro-method}
\title{Set loading modality}
\usage{
setLoadingModality(object)

\S4method{setLoadingModality}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
The input object.
}
\description{
Sets the slot @loading_modality of each well plate based on the content of the input
example data.frame.
}
