% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-M.R
\name{setModuleActivity}
\alias{setModuleActivity}
\alias{setModuleActivity,Cypro-method}
\title{Set module activity}
\usage{
setModuleActivity(object, ...)

\S4method{setModuleActivity}{Cypro}(object, activity, in_shiny = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{activity}{Logical vector. Must be named according to the modules for which
slot @activity is supposed to be set.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object
}
\description{
Sets slot @used of the analysis modules which activates
or inactivates the respective modules.
}
