% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setO-Z.R
\name{setOutlierDetection}
\alias{setOutlierDetection}
\alias{setOutlierDetection,Cypro-method}
\title{Set object of class \code{OutlierDetection}}
\usage{
setOutlierDetection(object, ...)

\S4method{setOutlierDetection}{Cypro}(object, outlier_object = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{outlier_object}{Object of class \code{OutlierDetection}.}
}
\value{
The input object.
}
\description{
Sets object of class \code{OutlierDetection}. Removes
content of slot @data beforehand to prevent accumulation of equal
data.frames in the \code{Cypro} object.
}
