% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setProgress}
\alias{setProgress}
\alias{setProgress,Cypro-method}
\title{Set progress status}
\usage{
setProgress(object, ...)

\S4method{setProgress}{Cypro}(
  object,
  designExperiment = NULL,
  assignVariables = NULL,
  loadData = NULL,
  processData = NULL,
  verbose = TRUE
)
}
\arguments{
\item{designExperiment, assignVariables, loadData, processData}{TRUE or FALSE.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Sets logical values in the object's slot @progress to keep
track of the progress made during the preparation process.
}
