% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setO-Z.R
\name{setStorageDirectory}
\alias{setStorageDirectory}
\alias{setStorageDirectory,Cypro-method}
\title{Set storage directory}
\usage{
setStorageDirectory(object, ...)

\S4method{setStorageDirectory}{Cypro}(object, directory, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{directory}{Character value. The file directory. Must end with \emph{'.RDS'}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Sets the directory under which the \code{Cypro} object
is saved by default with \code{saveCyproObject()}.
}
