% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setStatsDf}
\alias{setStatsDf}
\alias{setTracksDf}
\alias{setTracksDf,CdataTimeLapse-method}
\alias{setTracksDf,CyproTimeLapse-method}
\title{Set cypro data.frames}
\usage{
setStatsDf(object, df)

setTracksDf(object, df)

\S4method{setTracksDf}{CdataTimeLapse}(object, df)

\S4method{setTracksDf}{CyproTimeLapse}(object, df)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{df}{The data.frame to be set.}
}
\value{
The input object.
}
\description{
Functions to safely set the data.frames containing
data about the cells.
}
