% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-up-family.R
\name{setUpWellPlateDf}
\alias{setUpWellPlateDf}
\title{Well Plate Data.frame}
\usage{
setUpWellPlateDf(type = "8x12 (96)", phases = NULL)
}
\arguments{
\item{type}{Character value. One of \emph{'2x3 (6)', '3x4 (12)', '4x6 (24)', '6x8 (48)', '8x12 (96)'}}
}
\description{
Sets up a data.frame in which each observation refers
to a well.
}
