% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setO-Z.R
\name{setVariableAssignmentShiny}
\alias{setVariableAssignmentShiny}
\alias{setVariableAssignmentShiny,AnalysisModule-method}
\alias{setVariableAssignmentShiny,Cypro-method}
\title{Set variable assinment}
\usage{
setVariableAssignmentShiny(object, ...)

\S4method{setVariableAssignmentShiny}{AnalysisModule}(object, input_list)

\S4method{setVariableAssignmentShiny}{Cypro}(object, input_list, modules = NULL, validate = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{input_list}{The list to which one refers as \code{input} in shiny sessions
outputted by \code{shiny::reactiveValuesToList()}.}
}
\value{
The input object.
}
\description{
Only to use within \code{assignVariables()} as it is relying
on the input list that exists in active shiny sessions.

Picks the variable assignment from the input list of the shiny session and
assigns it to the respective variable of the module.
}
