% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setO-Z.R
\name{setWellInfo}
\alias{setWellInfo}
\alias{setWellInfo.layout_df}
\title{Set well plate layout information}
\usage{
setWellInfo(df, ...)

\method{setWellInfo}{layout_df}(
  df,
  wells = character(0),
  cell_line = NULL,
  condition = NULL,
  verbose = TRUE,
  in_shiny = FALSE
)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}

\item{wells}{Character vector. Denotes the wells for which the information
is supposed to be set.}

\item{cell_line}{Character value, NA or NULL. The cell line seeded in the wells denoted in
\code{selected_wells}.}

\item{condition}{Character value, NA, or NULL. The condition of the well. If multiple phase
experiment a character vector of length equal to the number of phases of
the experiment design.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame.
}
\description{
Subsequently fills the well plate \code{layout_df} with information.
}
\details{
Sets well cell line and condition for all wells that are selected -
meaning for which their value in variable \emph{selected} is TRUE. This
can be manipulated with argument \code{wells} with which wells are selected.
Wells selected this way are deselected right before the data.frame is returned.
}
\note{
\code{setWellInfo()} sets the provided information for all selected
wells. If the input data.frame already contains selected wells these are
included in addition to those denoted in argument \code{wells}.
}
\seealso{
\code{selectWells()}, \code{deselectWells()}
}
