% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setWellPlate}
\alias{setWellPlate}
\alias{setWellPlate,ExperimentDesign-method}
\title{Set well plate S4 object}
\usage{
setWellPlate(object, ...)

\S4method{setWellPlate}{ExperimentDesign}(object, well_plate)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{well_plate}{Object of class \code{WellPlate}.}
}
\value{
The input object.
}
\description{
Sets an object of class \code{WellPlate} in the respective
slot of input for argument \code{object}.
}
\details{
Slots @name and @experiment of the \code{WellPlate} object must not be empty.
}
\seealso{
\code{WellPlate-class}
}
