% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-functions.R
\name{starts_with}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{all_of}
\alias{any_of}
\title{Imported selection helpers from package 'tidyselect'}
\usage{
starts_with(match, ignore.case = TRUE, vars = NULL)

ends_with(match, ignore.case = TRUE, vars = NULL)

contains(match, ignore.case = TRUE, vars = NULL)

matches(match, ignore.case = TRUE, perl = FALSE, vars = NULL)

all_of(x)

any_of(x, ..., vars = NULL)
}
\arguments{
\item{match}{A character vector. If length > 1, the union of the
matches is taken.

For \code{starts_with()}, \code{ends_with()}, and \code{contains()} this is an exact
match. For \code{matches()} this is a regular expression, and can be a
stringr pattern.}

\item{ignore.case}{If \code{TRUE}, the default, ignores case when matching
names.}

\item{vars}{A character vector of variable names. If not supplied,
the variables are taken from the current selection context (as
established by functions like \code{select()} or \code{pivot_longer()}).}

\item{perl}{Should Perl-compatible regexps be used?}
}
\description{
These \link[tidyselect:language]{selection helpers} match variables according
to a given pattern.
\itemize{
\item \code{\link[tidyselect:starts_with]{starts_with()}}: Starts with an exact prefix.
\item \code{\link[tidyselect:ends_with]{ends_with()}}: Ends with an exact suffix.
\item \code{\link[tidyselect:contains]{contains()}}: Contains a literal string.
\item \code{\link[tidyselect:matches]{matches()}}: Matches a regular expression.
\item \code{\link[tidyselect:num_range]{num_range()}}: Matches a numerical range like x01, x02, x03.
}
}
