% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetBy.R
\name{subsetByCellID}
\alias{subsetByCellID}
\alias{subsetByCellID,CyproScreening-method}
\alias{subsetByCellID,CyproTimeLapse-method}
\title{Create data subset by cell ids}
\usage{
subsetByCellID(object, ...)

\S4method{subsetByCellID}{CyproScreening}(
  object,
  cell_ids,
  new_name,
  new_directory = NA,
  reasoning = NULL,
  verbose = TRUE,
  ...
)

\S4method{subsetByCellID}{CyproTimeLapse}(
  object,
  cell_ids,
  new_name,
  new_directory = NA,
  reasoning = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{cell_ids}{Character vector. Denotes the cells to keep unambiguously with their cell IDs.}

\item{new_name}{Character value. Denotes the name of the output object. If set to NULL the name of the
input object is taken and suffixed with \emph{'_subset'}.}

\item{new_directory}{Character value. Storage directory for the new object.}

\item{reasoning}{Character value. Allows two write a short description of how the cell ids according
to which the object is filtered were selected. This description is included in the output of \code{printSubsetHistory()}.
Ignored if set to NULL.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object containing a data subset.
}
\description{
Subset functions allow to conveniently split your data by certain characteristics such
as cell lines, conditions, cluster etc. or for specific cell ids. This might be useful if you want apply some machine learning
algorithms such as clustering and correlation on only a subset of cells. See details for more information.
}
\details{
Creating subsets of your data affects analysis results such as clustering, dimensional redurction and correlation which
is why these results are reseted in the subsetted object and must be computed again. To prevent inadvertent overwriting
the default directory is reset as well. Make sure to set a new one via \code{setDefaultDirectory()}.

The mechanism with which you create the subset is stored in the output object. Use \code{printSubsetHistory()}
to reconstruct the way from the original object to the current one.
}
