% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetBy.R
\name{subsetByGroup}
\alias{subsetByGroup}
\alias{subsetByGroup,CyproScreening-method}
\alias{subsetByGroup,CyproTimeLapse-method}
\title{Create data subset by group}
\usage{
subsetByGroup(object, ...)

\S4method{subsetByGroup}{CyproScreening}(
  object,
  grouping_variable,
  groups,
  reasoning = NA_character_,
  verbose = TRUE
)

\S4method{subsetByGroup}{CyproTimeLapse}(
  object,
  grouping_variable,
  groups,
  reasoning = NA_character_,
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{grouping_variable}{Character value. Denotes the categorical variable - the grouping of cells -
of interest. Use \code{getGroupingVariableNames()} to obtain all valid input options.}

\item{groups}{Character vector. The group names of the variable denoted in
\code{grouping_variable} whose cells are supposed to be kept.}

\item{reasoning}{Character value. Allows two write a short description of how the cell ids according
to which the object is filtered were selected. This description is included in the output of \code{printSubsetHistory()}.
Ignored if set to NULL.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object containing a data subset.
}
\description{
Subset functions allow to conveniently split your data by certain characteristics such
as cell lines, conditions, cluster etc. or for specific cell ids. This might be useful if you want apply some machine learning
algorithms such as clustering and correlation on only a subset of cells. See sections below for more information.
}
\details{
Creating subsets of your data affects analysis results such as clustering,
dimensional reduction and correlation which is why slot @analysis is reset.

Within the data of slot @cdata, the cluster data.frame is emptied. Additionally,
levels of grouping variables that are not used any longer after the subsetting
are dropped.

To prevent inadvertent overwriting the default directory is reset as well.
Make sure to set a new one via \code{setDefaultDirectory()}.

The name of the returned \code{Cypro} object (slot @experiment) is suffixed
with \emph{_Subset\code{<nth_subset>}}.

The mechanism and the information with which you create the subset is stored in the output object.
Use \code{printSubsetHistory()} to reconstruct the way from the original object
to the current one.
}
\section{Subset mechanism}{


The cell IDs that belong to the groups denoted in \code{groups} are gathered.
The object is then subsetted by the mechanism described in the documentation
for function \code{subsetByCellID}.
}

