% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetByQuality}
\alias{subsetByQuality}
\title{Create data subset according to coverage quality}
\usage{
subsetByQuality(
  object,
  new_name = NULL,
  suffix = FALSE,
  suffix_sep = "_",
  verbose = NULL
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{new_name}{Character value. Denotes the name of the output object. If set to NULL the name of the
input object is taken and suffixed with \emph{'_subset'}.}

\item{suffix}{Logical value. If set to TRUE the character value provided with arugment \code{new_name} is
used to suffix the original name as well as the current default directory.}

\item{suffix_sep}{Character value. Denotes the string to insert between the object name / storage directory
and the input value for \code{new_name} if \code{suffix} is set to TRUE.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Subset functions allow to conveniently split your data. \code{subsetByQuality()}
opens a shiny application in which histograms of aspects are displayed that summarize
the quality of a cells coverage.

See details for more information.
}
\details{
Creating subsets of your data affects analysis results such as clustering and correlation which
is why these results are reset in the subsetted object and must be computed again. To prevent inadvertent overwriting
the default directory is reset as well. Make sure to set a new one via \code{setDefaultDirectory()}.

The mechanism with which you create the subset is stored in the output object. Use \code{printSubsetHistory()}
to reconstruct the way from the original object to the current one.

The histograms you see in the application provide insights into the distribution
of coverage quality assessments. (e.g. the  distribution of numbers of frames that
have been skipped by cells or the first frame the cells appeared in.) You can
select the columns that contain the cells that match the quality requirements
of your choice. Eventually cells that match all the requirements you specified
are selected and the object is subsetted by \code{subsetByCellId()}.

The requirements you set up are included in the message constructed and
printed by \code{printSubsetHistory()}.
}
