% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-object.R
\name{subsetByQualityCheck}
\alias{subsetByQualityCheck}
\title{Create data subset according to coverage quality}
\usage{
subsetByQualityCheck(object, new_name = NULL, verbose = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Subset functions allow to conveniently split your data. \code{subsetByQualiteCheck()}
opens a shiny application in which histograms of aspects are displayed that summarise
the quality of a cells coverage.

See details for more information.
}
\details{
Creating subsets of your data affects analysis results such as clustering and correlation which
is why these results are reset in the subsetted object and must be computed again. To prevent inadvertent overwriting
the default directory is reset as well. Make sure to set a new one via \code{setDefaultDirectory()}.

The mechanism with which you create the subset is stored in the output object. Use \code{printSubsetHistory()}
to reconstruct the way from the original object to the current one.

The histograms you see in the application provide insights into the distribution
of coverage quality assessments. (e.g. the  distribution of numbers of frames that
have been skipped by cells or the first frame the cells appeared in.) You can
select the columns that contain the cells that match the quality requirements
of your choice. Eventually cells that match all the requirements you specified
are selected and the object is subsetted by \code{subsetByCellId()}.

Use \code{printSubsetHistory()} to have the requirements you set up printed
in the R-console.
}
