% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{suggestLoadingModality}
\alias{suggestLoadingModality}
\alias{suggestLoadingModality,Cypro-method}
\title{Suggest the proper loading function}
\usage{
suggestLoadingModality(object)

\S4method{suggestLoadingModality}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Character value.
}
\description{
Suggests the loading function to use depending on the
example data and the variable assignment during \code{assignVariables()}.
}
