% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeDataByCell}
\alias{summarizeDataByCell}
\alias{summarizeDataByCell,CyproTimeLapse-method}
\title{Summarize time lapse data}
\usage{
summarizeDataByCell(object, verbose = TRUE)

\S4method{summarizeDataByCell}{CyproTimeLapse}(object, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Summarizes time lapse data (the \code{tracks_df}) by cell id and sets up
the \code{stats_df} - the cellular statistics data.frame.
}
