% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeDataByWell}
\alias{summarizeDataByWell}
\alias{summarizeDataByWell,CyproScreening-method}
\title{Summarize data by well}
\usage{
summarizeDataByWell(object, ...)

\S4method{summarizeDataByWell}{CyproScreening}(
  object,
  summarize_with = "mean",
  features = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{features}{Character or NULL. If character, selects the numeric
variables that are summarized. If NULL, all are summarized.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Summarizes cell based data by well for every well plate and
adds the summarized variables to the layout data.frame of every well plate.
}
\details{
Summarized data variables that already exist under the same name in the
layout data.frames are overwritten without asking.
}
\seealso{
\code{getLayoutVariableNames()}
}
