% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeModuleVariables}
\alias{summarizeModuleVariables}
\alias{summarizeModuleVariables,CyproTimeLapse-method}
\title{Summarize module variables}
\usage{
summarizeModuleVariables(object, ...)

\S4method{summarizeModuleVariables}{CyproTimeLapse}(object, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Summarizes the summary variables of all active analysis modules
in the \code{Cypro} object and adds them to \code{Cypro} objects data.
}
