% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-timelapse-data.R
\name{summarizeTrackDfWith}
\alias{summarizeTrackDfWith}
\title{Summarize time lapse data variables}
\usage{
summarizeTrackDfWith(
  object,
  summarize_with = c("max", "mean", "median", "min"),
  verbose = NULL
)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Sets up the stat data.frame based on the data variables
of slot @tracks. This includes summary via the denoted functions
in argument \code{summarize_with} as well as all module based
variables.
}
\note{
As this function affects and changes the whole stat data.frame all analysis
results are discarded.
}
