% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-timelapse-data.R
\name{summarizeTrackVariables}
\alias{summarizeTrackVariables}
\title{Summarize time lapse data variables}
\usage{
summarizeTrackVariables(
  object,
  ...,
  overwrite = FALSE,
  phase = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{...}{Name-value pairs according to the syntax of \code{dplyr::summarize()}.
This can be single or several expressions as well as usage of \code{dplyr::across()} if several
variables are supposed to be affected. See details for output requirements.}

\item{overwrite}{Logical value. Must be set to TRUE in case of overlapping
variable names.

(Note that overwriting stat variables leads to all analysis progress of affected
variable sets being discarded.)}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
This function allows to summarize time lapse data variables
stored in slot @tracks and to add the summarized variables to the
data.frame in slot @stats.
}
\details{
Prior to summarizing the tracks data.frame is grouped by \emph{cell_id}
via \code{dplyr::group_by()}. All data variables of the output data.frame of
\code{dplyr::summarize(...)} is joined to the stats data.frame via
\code{addStatVariables()}.
}
