% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{SummaryVariable-class}
\alias{SummaryVariable-class}
\alias{SummaryVariable}
\title{The SummaryVariable Class}
\description{
Subclass of \code{DataVariable}. See details for more information.
}
\details{
Summary variables represent variables of time lapse experiments that can be
computed by summarizing other variables.

E.g. the total distance a cell has traveled throughout the imaging process is part of the
migration analysis module and referred to in \code{cypro} as \emph{total_dist}.
It can be summarized by adding up all distances a cell has
traveled from frame to frame. (The so called distance from last point is referred to as
\emph{dflp} in \code{cypro} and is a computable variable of the migration analysis module).
}
\section{Slots}{

\describe{
\item{\code{summarize_with}}{function. A function that takes the input data as the first
argument \code{df}, the summarized data as the second argument \code{stat_df}
and the \code{Cypro} object as the third argument. It uses
the information of both to compute the variable. The function must only return the
stat data.frame now containing the additional variable that has just been summarized.}
}}

\seealso{
\code{AssignableVariable}, \code{RequiredVariable}, \code{ComputableVariable}
}
