% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-n-z.R
\name{transferData}
\alias{transferData}
\alias{transferData,CyproScreening-method}
\alias{transferData,CyproTimeLapse-method}
\alias{transferData,CyproTimeLapseMP-method}
\title{Transfer loaded data}
\usage{
transferData(object, ...)

\S4method{transferData}{CyproScreening}(object, verbose = TRUE)

\S4method{transferData}{CyproTimeLapse}(object, verbose = TRUE)

\S4method{transferData}{CyproTimeLapseMP}(object, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{Cypro} object with the data files merged to an appropriate
\code{cypro_df} in an appropriate \code{Cdata} object set in slot @cdata.
}
\description{
Extracts the content of all loaded files across
all well plates, merges them into one data.frame. Then sets up the appropriate
\code{Cdata} object of slot @cdata.
}
\details{
All files that have been loaded with \code{loadDataFiles()} without
any error are merged to a data.frame. Depending on the loading modality - \emph{by_roi},
\emph{by_well}, \emph{by_well_plate} - the missing ID variables are added to the
data.frame. It is then split into the functional parts the \code{Cdata} object
that corresponds to the experiment design determines. The \code{Cdata} object
is then set in slot @cdata of the \code{Cypro} object.
}
