% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-experiment-design-server.R
\name{update_condition}
\alias{update_condition}
\alias{update_condition_df}
\title{Title}
\usage{
update_condition(var_condition, var_selected, input_condition)

update_condition_df(
  var_condition_df,
  var_selected,
  input_condition,
  input_phases
)
}
\description{
Take shiny input and update the condition and the condition_df variables
of the well plate data.frame.
}
