% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-cypro-object.R
\name{updateCyproObject}
\alias{updateCyproObject}
\title{Keep cypro object up to data}
\usage{
updateCyproObject(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
A cypro object whose architecture is up to date with the latest version.
}
\description{
Ensures that the cypro object is constructed according
to the latest version of the package.
}
