% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{used_variable_names_shiny}
\alias{used_variable_names_shiny}
\title{Title}
\usage{
used_variable_names_shiny(assembled_module_info, test_overlap = FALSE)
}
\arguments{
\item{assembled_module_info}{output of assemble_module_info()-functions (module information list)}

\item{test_overlap}{if set to TRUE stops execution if var names are not uniquely
assigned}
}
\value{
Extracts information about the name of variables carried by the example file from
the module information list in form of a character vector. Checks for duplicated
variable name assignemnt
}
\description{
Title
}
