% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-design-experiment-server.R
\name{valid_example_cols}
\alias{valid_example_cols}
\title{Title}
\usage{
valid_example_cols(example_list)
}
\description{
Make sure that all denoted columns appear only one time.
}
