% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-input-options.R
\name{validAgglomerationMethods}
\alias{validAgglomerationMethods}
\alias{validColorPalettes}
\alias{validColorSpectra}
\alias{validDistanceMethods}
\alias{validKmeansMethods}
\title{Valid input options}
\usage{
validAgglomerationMethods()

validColorPalettes()

validColorSpectra()

validDistanceMethods()

validKmeansMethods()
}
\value{
Character vector.
}
\description{
Helper functions that to not take any arguments but simply return
a character vector of valid input options for certain arguments.
}
