% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{validate_analysis_module_variables_shiny}
\alias{validate_analysis_module_variables_shiny}
\title{Title}
\usage{
validate_analysis_module_variables_shiny(
  assembled_module_info,
  df,
  mitosis_module_used,
  object
)
}
\arguments{
\item{assembled_module_info}{the respective output of assemble_module_info*()
functions (module information list)}

\item{mitosis_module_used}{logical, skip if lineages have to be reconstructed}

\item{object}{}

\item{example_df}{the example file loaded in prepare data loading}
}
\value{
List of variables, converted if necessary, named according to the variable. If
conversion was not possible or the variable was missing the respective slot
contains the error message of the thrown error instead.
}
\description{
makes sure that the denoted identifier variables are valid.
to be used in observeEvent() of pdl_step4 action button 'Save & Proceed'
}
