% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-functions.R
\name{validate_identifier_variables_shiny}
\alias{validate_identifier_variables_shiny}
\title{Title}
\usage{
validate_identifier_variables_shiny(
  df,
  assembled_module_info,
  mitosis_module_used,
  object,
  in_shiny
)
}
\arguments{
\item{assembled_module_info}{the respective output of assemble_module_info*()
functions (module information list)}

\item{mitosis_module_used}{logical, skip if lineages have to be reconstructed}

\item{object}{}

\item{example_df}{the example file loaded in prepare data loading}
}
\value{
Variables (list of variables if time lapse), converted if necessary. If
conversion not possible or variable missing the error message of the thrown
error.
}
\description{
makes sure that the denoted identifier variables are valid.
to be used in observeEvent() of pdl_step4 action button 'Save & Proceed'
}
