% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateVariableAssignment}
\alias{validateVariableAssignment}
\alias{validateVariableAssignment,Cypro-method}
\title{Validate variable assignment}
\usage{
validateVariableAssignment(object, ...)

\S4method{validateVariableAssignment}{Cypro}(
  object,
  modules = NULL,
  in_shiny = FALSE,
  stop_if_false = FALSE,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{stop_if_false}{Logical value. If set to TRUE and the test evaluates to FALSE the
function is stopped via \code{stop()} printing an informative message. If FALSE the
resulting logical value is returned.}
}
\value{
TRUE or FALSE
}
\description{
Validates that each data variable has been assigned a unique
variable from the input data.
}
