% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validateWells}
\alias{validateWells}
\alias{validateRois}
\alias{validateWellRois}
\title{Validate well plate information}
\usage{
validateWells(wells, cypro_nc = FALSE)

validateRois(rois, cypro_nc = FALSE)

validateWellRois(well_rois, cypro_nc = FALSE)
}
\arguments{
\item{cypro_nc}{Logical value. Denotes which regular expression to use and thus, the valid
naming conventions. If TRUE, only the naming convention of the \code{cypro}-package is considered valid.}
}
\value{
A named logical vector of the same length as the input vector.
}
\description{
Validates the content of variables regarding the
well plate layout: \emph{well}, \emph{roi} and \emph{well_roi} by
using regular expressions.
}
