% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{WellPlate-class}
\alias{WellPlate-class}
\alias{WellPlate}
\title{The WellPlate Class}
\description{
S4-objects that abstracts a well plate.
It additionally carries information about if and how the related data has been
loaded.
}
\section{Slots}{

\describe{
\item{\code{experiment}}{character. The name of the \code{Cypro} object it is part of.}

\item{\code{files}}{character. All spreadsheet-files directories that have been loaded
for that well plate.}

\item{\code{folder}}{character. Only relevant in case of usage of \code{loadFilesByWellRoi()}.
Contains the directory to the folder from which the files have been loaded.}

\item{\code{loaded_by}}{character. The function that has been used to load data for this
well plate.}

\item{\code{index}}{numeric. The ordinal well plate number. E.g. if @index = 2, it was the second
well plate that has been designed.}

\item{\code{name}}{character. The well plate name.}

\item{\code{set_up}}{data.frame. A data.frame carrying and additonal S3-class \code{wp_design_df} or
\code{wp_design_mp_df} in case of multiple phase experiments.}

\item{\code{type}}{character. The well plate type. Currently one of
\emph{'2x3 (6)', '3x4 (12)', '4x6 (24)', '6x8 (48)','8x12 (96)'}.}
}}

