% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addVariableSet}
\alias{addVariableSet}
\title{Add predefined set of variables}
\usage{
addVariableSet(object, variable_names, set_name, overwrite = FALSE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{variable_names}{Character vector. Denotes the variable names that are part
of the variable set.}

\item{set_name}{Character value. The name of the set of variable names.}

\item{overwrite}{Logical value. Must be set to TRUE in case of overlapping
variable names.

(Note that overwriting stat variables leads to all analysis progress of affected
variable sets being discarded.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Allows to store predefined set of variable names in the object that
can be easily accessed via \code{getVariableSet()}.
}
