% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{AnalysisModule-class}
\alias{AnalysisModule-class}
\alias{AnalysisModule}
\title{The AnalysisModule Class}
\description{
An analysis module in \code{cypro} represents a biological aspect
of cells that can be described and explored using a fixed set of data variables.
See details for more information and examples.
}
\details{
Example: Cellular migration
Cellular migration is a commonly explored aspect in time lapse image analysis.
}
\section{Slots}{

\describe{
\item{\code{descr}}{character. Text that describes what the analysis module is about.}

\item{\code{name_in_app}}{character. Pretty name to be used as reference in application.}

\item{\code{name_in_cypro}}{character. character. Short name to be used as reference throughout
\code{cypro}. Should follow underscore-naming-convention (e.g. \emph{x_coords} instead of
\emph{xCoords} or \emph{x-coords})}

\item{\code{variables_computed}}{list. Named list of computable variables represented by the S4-Class
\code{ComputableVariable}.  Names of list correspond to the respective slot @name_in_cypro.
Variable must be ordered according to the order in which they are supposed to be computed.}

\item{\code{variables_required}}{list. Named list of required variables represented by the S4-Class
\code{RequiredVariable}.  Names of list correspond to the respective slot @name_in_cypro.}
}}

