% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_.R
\name{as_layout_df}
\alias{as_layout_df}
\alias{as_layout_df_mp}
\alias{as_layout_df.data.frame}
\alias{as_layout_df_mp.layout_df}
\alias{asLayoutDf}
\alias{asLayoutDf.data.frame}
\alias{asLayoutDfMP}
\alias{asLayoutDfMP.layout_df}
\title{Coerce to \code{layout_df} data.frame}
\usage{
as_layout_df(df, well_plate_type = NULL)

as_layout_df_mp(df, well_plate_type = NULL, n_phases = NULL)

\method{as_layout_df}{data.frame}(df, well_plate_type = NULL)

\method{as_layout_df_mp}{layout_df}(df, well_plate_type = NULL, n_phases = NULL)

asLayoutDf(df, well_plate_type)

\method{asLayoutDf}{data.frame}(df, well_plate_type = NULL)

asLayoutDfMP(df, well_plate_type = NULL, n_phases = NULL)

\method{asLayoutDfMP}{layout_df}(df, well_plate_type = NULL, n_phases = NULL)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
The input data.frame with adjusted attributes.
}
\description{
Coerces input data.frame to layout data.frames.
}
