% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-classes.R
\docType{class}
\name{AssignableVariable-class}
\alias{AssignableVariable-class}
\alias{AssignableVariable}
\title{The AssignableVariable Class}
\description{
An assignable variable is defined as a variable that can be part of
the input data. See details for more information.
}
\details{
To x- and y-coordinates can be referred to in many ways like \emph{x-coordinates,
x_coordinates, x_coords, x, etc.}. In \code{cypro} one refers to x-coordinates
via \emph{x_coords}. If the data set uploaded refers to x-coordinates
with \emph{x} instead of \emph{x_coords} problems would arise during all functions that need
the information regarding x- (and -y) coordinates since it is unclear which variable
should be used as the names are not identical. Therefore, the user denotes
the variable \emph{x} as the variable containing the x-coordinate information in
\code{assignVariables()}. Afterwards, while being loaded every data file
is checked for the variable \emph{x}, which is then renamed to \code{x_coords} and added
to the cell data of the \code{Cypro} object.

All data variables that have a specific meaning and are recurrently used in \code{cypro} are
handled this way. Further (S4-Classes) exist that differentiate between three types of
assignable data variables: \code{RequiredVariable}, \code{OptionalVariable}, \code{ComputableVariable}.
}
\section{Slots}{

\describe{
\item{\code{check_candidate}}{function. A function that takes a vector as input and
returns a single TRUE if the vector meets the requirements described in
slot @descr_requirements or a single FALSE if not. This only refers to superficial things
like class and types.

The function must have only one argument, namely \code{var}.}

\item{\code{check_var}}{function. A more sophisticated function that takes several things
into consideration to check if the content of the assigned variable fits the requirements.
It takes a data.frame, extracts the assigned variable, checks the variable's content,
adjusts it if necessary/possible and returns the variable or an informative message about
the content's flaws.

See details of the \code{check_var_*()}-function documentation for more information.}

\item{\code{descr_requirements}}{character. Text that describes the class requirements (e.g.
numeric, character).}

\item{\code{descr_variable}}{character. Text that describes what the data variable represents.}

\item{\code{name_in_example}}{character. Name the variable carried in the example
data set uploaded in \code{assignVariables()}.}

\item{\code{valid}}{logical. Indicates if the content of the assigned variable meets the
variable specific requirements and is valid.}
}}

\seealso{
\code{RequiredVariable}, \code{ComputableVariable}, \code{OptionalVariable}, \code{SummaryVariable}
}
