% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignVariables.R
\name{assignVariables}
\alias{assignVariables}
\title{Step 2: Assign variables}
\usage{
assignVariables(object = createEmptyCyproObject())
}
\arguments{
\item{object}{An object of class \code{Cypro}.}
}
\value{
The input object.
}
\description{
something
}
