% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-effect-detection.R
\name{detectBatchEffects}
\alias{detectBatchEffects}
\title{Detect batch effects}
\usage{
detectBatchEffects(object, verbose = TRUE)
}
\arguments{
\item{object}{A valid cypro object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Batch effects result from undesired technical variation and can lead to
misenterpretation and false conclusion. This function computes the distances of each
well median profile across all well plates.
}
