% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier-detection.R
\name{detectOutliers}
\alias{detectOutliers}
\title{Detect outlier cells}
\usage{
detectOutliers(
  object,
  method_outlier = "iqr",
  threshold_pval = 0.001,
  variable_names = NULL,
  across = NULL,
  phase = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method_outlier}{Character vector. Specifies the method/algorithm of interest.

Use \code{validOutlierDetectionMethods()} to obtain all valid input options.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
There are several methods according to which outlier cells can be detected or
defined. Which one to use strongly depends on the researchers question at hand. See details
for more information on which method does what exactly.
}
\details{
This function only detects outliers. It does not remove them from the object. Use \code{removeOutliers()}
for that matter. To get more insights in the outlier detection results before removing them
use \code{getOutlierResults()}.

Method: Interquartile Range (iqr)

Every stat variable denoted in argument \code{variable_names} is screened via
R's built in function \code{grDevices::boxplot.stats()}. Every ID that is
recognized as an outlier is stored in a list slot named after the stat variable
under which it was detected as an outlier.

Method: mahalanobis

Multivariate outlier detection method. Mahalanobis distance is computed including
all stat variables denoted in argument \code{variable_names} via \code{stats::mahalanobis()}.
P-values are calculated via \code{stats::pchisq()} and cells are filtered according
to the threshold set with argument \code{threshold_pval}. Cells with a mahalanobis distance
p-value lower or equal than the threshold are defined as outliers.
}
\seealso{
removeOutliers()
}
