% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests.R
\name{existOutlierResults}
\alias{existOutlierResults}
\alias{isTimeLapseExp}
\alias{isUsable}
\alias{multiplePhases}
\title{Logical Tests}
\usage{
existOutlierResults(object, phase = NA, method_outlier = NA, verbose = FALSE)

isTimeLapseExp(object)

isUsable(object, module, verbose = FALSE)

multiplePhases(object)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{method_outlier}{Character vector. Specifies the method/algorithm of interest.

Use \code{validOutlierDetectionMethods()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
TRUE or FALSE
}
\description{
Logical tests that return TRUE if the specified functions/algorithms
have been used on the object and FALSE if not.
}
