% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getAdditionalVariableNames}
\alias{getAdditionalVariableNames}
\alias{getAdditionalVariableNames,Cypro-method}
\title{Obtain additional variable names}
\usage{
getAdditionalVariableNames(object, with_grouping = TRUE, with_numeric = TRUE)

\S4method{getAdditionalVariableNames}{Cypro}(object, with_grouping = TRUE, with_numeric = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{with_grouping, with_numeric}{Logical values. Indicate the additional
variable types that are supposed to be included in the output list.}
}
\value{
A list.
}
\description{
Extracts variable names that have been denoted during
\code{assignVariables()} as unknown to \code{cypro} but of interest
to the user.
}
