% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getAnalysisAspect}
\alias{getAnalysisAspect}
\title{Obtain analysis aspect}
\usage{
getAnalysisAspect(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{aspect}{Character value. The analysis aspect.}
}
\value{
An object of class \code{AnalysisAspect}.
}
\description{
Extracts S4 object of class \code{AnalysisAspect} as described in
package \code{confuns}.
}
