% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getCdata}
\alias{getCdata}
\alias{getCdata,Cypro-method}
\title{Extract cell data}
\usage{
getCdata(object)

\S4method{getCdata}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
An object of class \code{Cdata}.
}
\description{
Obtain cell data in form of an S4-class called \code{Cdata}.
}
