% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getCellIDs}
\alias{getCellIDs}
\title{Obtain cell IDs}
\usage{
getCellIDs(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A character vector.
}
\description{
Extracts cell IDs in form of a character vector.
}
