% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family-A-M.R
\name{getCellLines}
\alias{getCellLines}
\alias{getCellLines,Cypro-method}
\alias{getConditions}
\alias{getConditions,Cypro-method}
\alias{getConditions,CyproTimeLapseMP-method}
\title{Extract cell line and condition names}
\usage{
getCellLines(object)

\S4method{getCellLines}{Cypro}(object)

getConditions(object, ...)

\S4method{getConditions}{Cypro}(object)

\S4method{getConditions}{CyproTimeLapseMP}(object, phase = NULL)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
Character vector.
}
\description{
Obtain names of the \emph{cell_line}- or the \emph{condition}-variable
in form of a character vector..
}
