% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getClusterDf}
\alias{getClusterDf}
\alias{getClusterDf,Cdata-method}
\alias{getClusterDf,CdataTimeLapseMP-method}
\alias{getClusterDf,Cypro-method}
\alias{getClusterDf,CyproTimeLapseMP-method}
\title{Obtain cluster data (by cell id)}
\usage{
getClusterDf(object, ...)

\S4method{getClusterDf}{Cdata}(object, ...)

\S4method{getClusterDf}{CdataTimeLapseMP}(object, phase, ...)

\S4method{getClusterDf}{Cypro}(object, verbose = NULL, ...)

\S4method{getClusterDf}{CyproTimeLapseMP}(object, phase = NULL, verbose = NULL, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame.
}
\description{
Obtain a data.frame that contains variables
with which cells are grouped according to clustering results.
}
