% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getData}
\alias{getData}
\title{Obtain data slots}
\usage{
getData(object, data_slot, phase)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{data_slot}{Character value. One of \emph{'stats', 'tracks', 'meta'} or \emph{'cluster'}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
The data.frame of interest.
}
\description{
A wrapper around \code{purr::map_df()} and the respective
list of the data slot of interest.
}
