% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family-A-M.R
\name{getDefaultInstructions}
\alias{getDefaultInstructions}
\alias{getDefaultInstructions,Cypro-method}
\title{Extract the object's default instructions}
\usage{
getDefaultInstructions(object, ...)

\S4method{getDefaultInstructions}{Cypro}(object, return = "list")
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{return}{Character value. Either \emph{'list'} (default) or \emph{'S4'}.}
}
\value{
A list or an S4 object.
}
\description{
Obtain what is currently defined as the default in form of
a list or the original S4-object.
}
