% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getFeatureDf}
\alias{getFeatureDf}
\alias{getFeatureDf,CdataScreening-method}
\alias{getFeatureDf,CyproScreening-method}
\alias{getFeatureDf,CyproTimeLapse-method}
\alias{getFeatureDf,CyproTimeLapseMP-method}
\title{Extract numeric cell data}
\usage{
getFeatureDf(object, ...)

\S4method{getFeatureDf}{CdataScreening}(object, ...)

\S4method{getFeatureDf}{CyproScreening}(
  object,
  drop_na = FALSE,
  with_grouping = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  with_well_plate = NULL,
  verbose = NULL,
  ...
)

\S4method{getFeatureDf}{CyproTimeLapse}(
  object,
  slot = "stats",
  drop_na = FALSE,
  with_grouping = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  with_well_plate = NULL,
  verbose = NULL
)

\S4method{getFeatureDf}{CyproTimeLapseMP}(
  object,
  slot = "stats",
  drop_na = FALSE,
  with_grouping = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  with_well_plate = NULL,
  phase = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{with_cluster}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{with_meta}{Logical. If set to TRUE the discrete variables of the cluster data slot are added
to the output data.frame.}

\item{with_well_plate}{Logical values. Denoting
if the respective grouping information should be joined to the output data.frame
or not.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
Obtain numeric cell features of screening experiments in
form of a data.frame.
}
\seealso{
\code{getTracksDf()} and \code{getStatsDf()} to obtain numeric data
of \code{Cypro} objects from time lapse experiments.

\code{getTracksDf()}, \code{getStatsDf()}
}
