% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getF.R
\name{getFeatureDf}
\alias{getFeatureDf}
\alias{getFeatureDf,CdataScreening-method}
\alias{getFeatureDf,CyproScreening-method}
\title{Extract numeric cell data}
\usage{
getFeatureDf(object, ...)

\S4method{getFeatureDf}{CdataScreening}(
  object,
  with_cluster = FALSE,
  with_meta = FALSE,
  with_well_plate = FALSE,
  with_everything = FALSE,
  ...
)

\S4method{getFeatureDf}{CyproScreening}(
  object,
  with_cluster = NULL,
  with_meta = NULL,
  with_well_plate = NULL,
  with_everything = NULL,
  with_non_data = TRUE,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{with_cluster}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{with_meta}{Logical. If set to TRUE the discrete variables of the cluster data slot are added
to the output data.frame.}

\item{with_well_plate}{Logical values. Denoting
if the respective grouping information should be joined to the output data.frame
or not.}
}
\value{
A data.frame.
}
\description{
Obtain numeric cell features of screening experiments in
form of a data.frame.
}
\seealso{
\code{getTracksDf()} and \code{getStatsDf()} to obtain numeric data
of \code{Cypro} objects from time lapse experiments.

\code{getTracksDf()}, \code{getStatsDf()}
}
