% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFeatureSet}
\alias{getFeatureSet}
\alias{getFeatureSet,Cypro-method}
\alias{getFeatureSets,Cypro-method}
\alias{getFeatureSetNames,Cypro-method}
\title{Obtain defined set of data variables}
\usage{
getFeatureSet(object, feature_set)

\S4method{getFeatureSet}{Cypro}(object)

\S4method{getFeatureSets}{Cypro}(object)

\S4method{getFeatureSetNames}{Cypro}(object)
}
\arguments{
\item{object}{A valid cypro object.}

\item{feature_set}{Character value. The name of the feature set of interest.}
}
\value{
A character vector.
}
\description{
Convenient access to defined sets of variables or names
mentioned sets. Useful to obtain vectors for recurrent arguments like
\code{variables}.
}
