% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getFeatureSet}
\alias{getFeatureSet}
\alias{getFeatureSet,Cypro-method}
\alias{getFeatureSets,Cypro-method}
\alias{getFeatureSetNames,Cypro-method}
\title{Extract a feature set}
\usage{
getFeatureSet(object, feature_set)

\S4method{getFeatureSet}{Cypro}(object)

\S4method{getFeatureSets}{Cypro}(object)

\S4method{getFeatureSetNames}{Cypro}(object)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{feature_set}{Character value. The name of the feature set of interest.}
}
\value{
A character vector.
}
\description{
Obtain defined sets of numeric data variables in form of vectors.
Useful to conveniently obtain vectors for recurring arguments like
\code{variables}.
}
