% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFeatureSetNames}
\alias{getFeatureSetNames}
\title{Obtain names of defined sets of data variables}
\usage{
getFeatureSetNames(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
A character vector.
}
\description{
Obtain names of defined sets of data variables
}
