% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getFeatureSets}
\alias{getFeatureSets}
\title{Obtain defined sets of data variables}
\usage{
getFeatureSets(object)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}
}
\value{
A list of character vectors.
}
\description{
Obtain defined sets of data variables in form of a list of
character vectors.
}
