% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFeatureSets}
\alias{getFeatureSets}
\title{Obtain defined sets of data variables}
\usage{
getFeatureSets(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
A list of character vectors.
}
\description{
Convenient access to defined sets of variables or names
mentioned sets. Useful to obtain vectors for recurrent arguments like
\code{variables}.
}
