% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFrameTimeSeq}
\alias{getFrameTimeSeq}
\alias{getFrameTimeSeq,CyproTimeLapse-method}
\alias{getFrameTimeSeq,CyproTimeLapseMP-method}
\title{Obtain frame time sequence}
\usage{
getFrameTimeSeq(object, ...)

\S4method{getFrameTimeSeq}{CyproTimeLapse}(object)

\S4method{getFrameTimeSeq}{CyproTimeLapseMP}(object, phase = "all", ...)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A numeric vector.
}
\description{
Convenient access to the time line as a numeric vector.
}
