% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getGroupingDf}
\alias{getGroupingDf}
\alias{getGroupingDf,Cypro-method}
\alias{getGroupingDf,CyproTimeLapseMP-method}
\title{Extract cell grouping data}
\usage{
getGroupingDf(object, ...)

\S4method{getGroupingDf}{Cypro}(object, verbose = NULL, ...)

\S4method{getGroupingDf}{CyproTimeLapseMP}(object, phase = NULL, verbose = NULL, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
Obtain a data.frame that contains variables that group
cells.
}
