% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getGroupingVariableNames}
\alias{getGroupingVariableNames}
\alias{getGroupingVariableNames,Cypro-method}
\alias{getGroupingVariableNames,CyproTimeLapseMP-method}
\title{Obtain grouping variable names of cell data}
\usage{
getGroupingVariableNames(object, ...)

\S4method{getGroupingVariableNames}{Cypro}(object, ..., named = FALSE, verbose = NULL)

\S4method{getGroupingVariableNames}{CyproTimeLapseMP}(
  object,
  ...,
  named = FALSE,
  phase = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}

\item{named}{Logical value. If set to TRUE the grouping variables are named
according to their grouping type (cluster, meta or well_plate).}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A character vector.
}
\description{
Extracts the names of variables that group cells. Useful to obtain
valid input options for recurring arguments like \code{across} or \code{grouping_variable}.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
