% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/deprecated.R
\name{getGroupNames}
\alias{getGroupNames}
\alias{getGroupNames,Cypro-method}
\alias{getGroupNames,CyproTimeLapseMP-method}
\alias{getGroups}
\title{Obtain group names a grouping variable contains}
\usage{
getGroupNames(object, ...)

\S4method{getGroupNames}{Cypro}(object, grouping_variable, ...)

\S4method{getGroupNames}{CyproTimeLapseMP}(object, grouping_variable, ..., phase = NULL)

getGroups(object, option)
}
\arguments{
\item{object}{A valid cypro object.}

\item{grouping_variable}{Character value. Denotes the discrete variable - the grouping of cells -
of interest. Use \code{getGroupingVariableNames()} to obtain all valid input options.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
Character vector of group names.
}
\description{
This function returns the names of the groups in which a specific grouping
variable groups the cells. Useful to obtain input options for arguments like \code{across_subset}.
}
