% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getHclustConv}
\alias{getHclustConv}
\alias{getHclustObject}
\alias{getKmeansConv}
\alias{getKmeansObject}
\alias{getPamConv}
\alias{getPamObject}
\title{Obtain cypros clustering objects}
\usage{
getHclustConv(object, variable_set, phase = NULL, with_data = TRUE)

getHclustObject(object, variable_set, phase = NULL, with_data = TRUE)

getKmeansConv(object, variable_set, phase = NULL, with_data = TRUE)

getKmeansObject(object, variable_set, phase = NULL, with_data = TRUE)

getPamConv(object, variable_set, phase = NULL, with_data = TRUE)

getPamObject(object, variable_set, phase = NULL, with_data = TRUE)
}
\arguments{
\item{object}{A valid cypro object.}

\item{variable_set}{Character value. Denotes the variable set of interest. Use \code{getVariableSetNames()}
to obtain all names of currently stored variable sets in your object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
An S4 object of \emph{'hclust_conv'}, \emph{'kmeans_conv'} or \emph{'pam_conv'}.
}
\description{
Obtain cypros clustering objects
}
