% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-getM.R
\name{getMergedDf}
\alias{getMergedDf}
\alias{getMergedDf,Cypro-method}
\title{Obtain merged data.frame}
\usage{
getMergedDf(object, verbose = TRUE)

\S4method{getMergedDf}{Cypro}(object, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame of class \code{cypro_df}.
}
\description{
Merges @slot content of all objects of class \code{DataFiles}
that contain data and that do not contain errors to a single data.frame.
}
