% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family-A-M.R
\name{getMetaDf}
\alias{getMetaDf}
\alias{getMetaDf,Cypro-method}
\alias{getMetaDf,CyproTimeLapseMP-method}
\alias{getMetaDf,Cdata-method}
\alias{getMetaDf,CdataTimeLapseMP-method}
\title{Extract cell meta data}
\usage{
getMetaDf(object, ...)

\S4method{getMetaDf}{Cypro}(object)

\S4method{getMetaDf}{CyproTimeLapseMP}(object, phase = NULL)

\S4method{getMetaDf}{Cdata}(object, ...)

\S4method{getMetaDf}{CdataTimeLapseMP}(object, phase, ...)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
Obtain cell meta data in form of a data.frame that contains
data variables with which cells are grouped.
}
\seealso{
\code{getClusterDf()}, \code{getWellPlateDf()}
}
