% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family-A-M.R
\name{getMetaVariableNames}
\alias{getMetaVariableNames}
\alias{getMetaVariableNames,Cypro-method}
\alias{getMetaVariableNames,CyproTimeLapseMP-method}
\title{Extract meta variable names of cell data}
\usage{
getMetaVariableNames(object, ...)

\S4method{getMetaVariableNames}{Cypro}(object, ..., named = FALSE, verbose = NULL)

\S4method{getMetaVariableNames}{CyproTimeLapseMP}(object, ..., named = FALSE, phase = NULL, verbose = NULL)
}
\value{
A character vector.
}
\description{
Obtain the names of variables in the meta data.frame. Useful to obtain
valid input options for recurring arguments like \code{across} or \code{grouping_variable}.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
