% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getMissingValuesDf}
\alias{getMissingValuesDf}
\title{Obtain missing value counts}
\usage{
getMissingValuesDf(object, phase = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
This function returns a data.frame giving insight into
the number of missing values every cell has across all variables.
}
