% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getMissingValuesDf}
\alias{getMissingValuesDf}
\alias{getMissingValuesDf,Cypro-method}
\alias{getMissingValuesDf,CyproTimeLapseMP-method}
\title{Obtain missing value counts}
\usage{
getMissingValuesDf(object, ...)

\S4method{getMissingValuesDf}{Cypro}(object, ...)

\S4method{getMissingValuesDf}{CyproTimeLapseMP}(object, phase = NULL, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A data.frame.
}
\description{
This function returns a data.frame giving insight into
the number of missing values every cell has across all variables.
}
