% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getOutlierDetection}
\alias{getOutlierDetection}
\alias{getOutlierDetection,Cypro-method}
\title{Obtain object of class \code{OutlierDetection}}
\usage{
getOutlierDetection(object, ...)

\S4method{getOutlierDetection}{Cypro}(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{}
}
\value{
An object of S4-class \code{OutlierDetection}.
}
\description{
Extracts the complete \code{OutlierDetection} S4-object from
slot @quality_checks.
}
