% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getPhases}
\alias{getPhases}
\alias{getPhases,CyproTimeLapseMP-method}
\title{Obtain vector of phase names}
\usage{
getPhases(object)

\S4method{getPhases}{CyproTimeLapseMP}(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
Character vector.
}
\description{
Obtain vector of phase names
}
