% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getS.R
\name{getPhases}
\alias{getPhases}
\alias{getPhases,CyproTimeLapseMP-method}
\title{Extract phase names}
\usage{
getPhases(object)

\S4method{getPhases}{CyproTimeLapseMP}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Numeric vector.
}
\description{
Obtain a numeric vector of valid phases.
}
\seealso{
\code{getPhaseNames()}
}
