% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getStatsDf}
\alias{getStatsDf}
\alias{getStatsDf,CdataTimeLapse-method}
\alias{getStatsDf,CdataTimeLapseMP-method}
\alias{getStatsDf,CyproTimeLapse-method}
\alias{getStatsDf,CyproTimeLapseMP-method}
\title{Extract stats data.frame}
\usage{
getStatsDf(object, ...)

\S4method{getStatsDf}{CdataTimeLapse}(object, ...)

\S4method{getStatsDf}{CdataTimeLapseMP}(object, phase, ...)

\S4method{getStatsDf}{CyproTimeLapse}(
  object,
  drop_na = TRUE,
  with_grouping = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  with_well_plate = NULL,
  verbose = NULL,
  ...
)

\S4method{getStatsDf}{CyproTimeLapseMP}(
  object,
  drop_na = TRUE,
  with_grouping = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  with_well_plate = NULL,
  phase = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{with_cluster, with_meta, with_well_plate}{Logical value. Denoting
if the respective grouping information should be joined to the stats data.frame
or not.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame with all numeric variables summarizing the measurements of
the track data.frame.
}
\description{
Obtain summarized numeric cell features of time lapse experiments in form of a data.frame.
}
