% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getStatVariableNames}
\alias{getStatVariableNames}
\alias{getStatVariableNames,CyproTimeLapse-method}
\alias{getStatVariableNames,CyproTimeLapseMP-method}
\title{Extract numeric variables of cell stat data}
\usage{
getStatVariableNames(object, ...)

\S4method{getStatVariableNames}{CyproTimeLapse}(object, ...)

\S4method{getStatVariableNames}{CyproTimeLapseMP}(object, phase = NULL, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A character vector.
}
\description{
Obtain the names of the numeric data variables of your
cell data. Useful to obtain vectors of variable names as input for recurring
arguments like \code{variables}.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
