% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getS.R
\name{getStorageDirectory}
\alias{getStorageDirectory}
\alias{getStorageDirectory,Cypro-method}
\title{Obtain storage directory}
\usage{
getStorageDirectory(object)

\S4method{getStorageDirectory}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Character value.
}
\description{
Extract the directory under which the \code{Cypro} object
is currently stored by default using \code{saveCyproObject()}.
}
