% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTracksDf}
\alias{getTracksDf}
\title{Obtain track data.frame.}
\usage{
getTracksDf(
  object,
  phase = NULL,
  with_grouping = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  with_well_plate = NULL,
  drop_na = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{with_cluster}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{with_meta}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame in which each observation refers to a cell at a given frame.
}
\description{
Obtain track data.frame.
}
