% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getTracksDf}
\alias{getTracksDf}
\alias{getTracksDf,CdataTimeLapse-method}
\alias{getTracksDf,CdataTimeLapseMP-method}
\alias{getTracksDf,CyproTimeLapse-method}
\alias{getTracksDf,CyproTimeLapseMP-method}
\title{Extract track data.frame}
\usage{
getTracksDf(object, ...)

\S4method{getTracksDf}{CdataTimeLapse}(object, ...)

\S4method{getTracksDf}{CdataTimeLapseMP}(object, phase, ...)

\S4method{getTracksDf}{CyproTimeLapse}(
  object,
  with_grouping = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  with_well_plate = NULL,
  verbose = NULL
)

\S4method{getTracksDf}{CyproTimeLapseMP}(
  object,
  with_grouping = NULL,
  with_cluster = NULL,
  with_meta = NULL,
  with_well_plate = NULL,
  phase = "all",
  phase_add = TRUE,
  verbose = NULL
)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{with_cluster}{Logical. If set to TRUE the discrete variables of the meta data slot are added
to the output data.frame.}

\item{with_meta}{Logical. If set to TRUE the discrete variables of the cluster data slot are added
to the output data.frame.}

\item{with_well_plate}{Logical values. Denoting
if the respective grouping information should be joined to the output data.frame
or not.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame in which each observation refers to a cell at a given frame.
}
\description{
Obtain numeric cell features of time lapse experiments in
form of a data.frame.
}
