% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getTrackVariableNames}
\alias{getTrackVariableNames}
\alias{getTrackVariableNames,CyproTimeLapse-method}
\alias{getTrackVariableNames,CyproTimeLapseMP-method}
\title{Extract numeric variables of cell track data}
\usage{
getTrackVariableNames(object, ...)

\S4method{getTrackVariableNames}{CyproTimeLapse}(object, only_numeric = TRUE, ...)

\S4method{getTrackVariableNames}{CyproTimeLapseMP}(object, phase = NULL, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A character vector.
}
\description{
Obtain the names of cell features of time lapse experiments.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
