% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getWellPlateDf}
\alias{getWellPlateDf}
\alias{getWellPlateDf,Cypro-method}
\title{Obtain well plate data (by cell id)}
\usage{
getWellPlateDf(object)

\S4method{getWellPlateDf}{Cypro}(object)
}
\arguments{
\item{object}{A valid cypro object.}
}
\value{
A data.frame.
}
\description{
This function lets you extract a data.frame that contain variables
with which cells are grouped.
}
