% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getWellPlateDf}
\alias{getWellPlateDf}
\alias{getWellPlateDf,Cdata-method}
\alias{getWellPlateDf,Cypro-method}
\title{Extract cell well plate data (by cell id)}
\usage{
getWellPlateDf(object, ...)

\S4method{getWellPlateDf}{Cdata}(object)

\S4method{getWellPlateDf}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A data.frame.
}
\description{
Obtain cell grouping variables according to the experiment
design.

Not to be confused with the well plate layout which is stored in data.frames
of class \code{layout_df} that can be obtained via \code{getLayoutDf()}.
}
