% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family-N-Z.R
\name{getWellPlateIndices}
\alias{getWellPlateIndices}
\alias{getWellPlateIndices,Cypro-method}
\title{Extract well plate indices}
\usage{
getWellPlateIndices(object, ...)

\S4method{getWellPlateIndices}{Cypro}(object, return = "tibble")
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{return}{Character value. Either \emph{'tibble'} or \emph{'vector'}.}
}
\value{
Either a tibble or a vector.
}
\description{
Obtain the indices of well plates. Well plates are indexed
according to the order in which their layout is designed during the
session in \code{designExperimen()}.
}
