% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getWellPlateNames}
\alias{getWellPlateNames}
\alias{getWellPlateNames,ExperimentDesign-method}
\alias{getWellPlateNames,Cypro-method}
\title{Extract well plate names}
\usage{
getWellPlateNames(object)

\S4method{getWellPlateNames}{ExperimentDesign}(object)

\S4method{getWellPlateNames}{Cypro}(object)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}
}
\value{
A character vector.
}
\description{
Obtain names of the well plate stored in the experiment design
in form of a character vector.
}
