% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-getZ.R
\name{getWellPlateType}
\alias{getWellPlateType}
\alias{getWellPlateType,WellPlate-method}
\alias{getWellPlateType,layout_df-method}
\title{Obtain well plate type}
\usage{
getWellPlateType(object, ...)

\S4method{getWellPlateType}{WellPlate}(object, ...)

\S4method{getWellPlateType}{layout_df}(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{

}
\description{
Extracts the type of the well plate.
}
