% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family-N-Z.R
\name{getWellPlateVariableNames}
\alias{getWellPlateVariableNames}
\alias{getWellPlateVariableNames,Cypro-method}
\title{Extract well plate variable names of cell data}
\usage{
getWellPlateVariableNames(object, ...)

\S4method{getWellPlateVariableNames}{Cypro}(object, ...)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}
}
\value{
A character vector.
}
\description{
Obtain names of variables that group cells according to
the well plate layouts.
}
\seealso{
starts_with(), ends_with(), contains(), matches()
}
