% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-family.R
\name{hlpr_add_data_to_cluster_object}
\alias{hlpr_add_data_to_cluster_object}
\title{Adds data to cluster object}
\usage{
hlpr_add_data_to_cluster_object(object, cluster_object, with_data, phase)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\description{
Adds data to cluster object
}
