% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nCells}
\alias{nCells}
\alias{nCellLines}
\alias{nConditions}
\title{Number of miscellaneous content}
\usage{
nCells(object, across = NULL, drop_na = FALSE, phase = NULL, ...)

nCellLines(object)

nConditions(object, phase = NULL)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A numeric value. If \code{across} is specified a data.frame.
}
\description{
These functions return a numeric value describing the number
of \code{n*()} the object contains. (E.g. \code{nCells()} returns the total
number of cells about which the object contains data, \code{nConditions()}
returns the total number of conditions, etc. )
}
