% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nCells}
\alias{nCells}
\alias{nCellLines}
\alias{nConditions}
\title{Number of miscellaneous content}
\usage{
nCells(object)

nCellLines(object)

nConditions(object, phase = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A numeric value.
}
\description{
These functions return a numeric value describing the number
of \code{n*()} the object contains. (E.g. \code{nCells()} returns the total
number of cells about which the object contains data, \code{nConditions()}
returns the total number of conditions, etc. )
}
