% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nFiles}
\alias{nFiles}
\alias{nFrames,ExperimentDesign-method}
\alias{nFrames,Cypro-method}
\title{Number of files read in}
\usage{
nFiles(object)

\S4method{nFrames}{ExperimentDesign}(object)

\S4method{nFrames}{Cypro}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A numeric vector named by well plate. Values indicate the number of valid directories
that have been read in via \code{loadData()}
}
\description{
Returns the number of cells the \code{Cypro} object
contains data about.

These functions return a numeric value describing the number
of \code{n*()} the object contains. (E.g. \code{nCells()} returns the total
number of cells about which the object contains data, \code{nConditions()}
returns the total number of conditions, etc. )
}
