% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nMissingValuesTracks}
\alias{nMissingValuesTracks}
\title{Number of NAs by cell id  (in tracks)}
\usage{
nMissingValuesTracks(object, variable_names, phase = NULL, verbose = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A summarized data.frame. Each row contains the number of missing
values a cell has in the respective variable.
}
\description{
Returns the number of missing values of every track
variable denoted in \code{variable_names} by cell id.
}
