% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nWellPlates}
\alias{nWellPlates}
\alias{nWellPlates,ExperimentDesign-method}
\title{Number of NAs by cell id  (in stats)}
\usage{
nWellPlates(object)

\S4method{nWellPlates}{ExperimentDesign}(object)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}
}
\value{
Numeric value.
}
\description{
Returns the number of well plates.
}
