% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module-migration.R
\name{plotAllTracks}
\alias{plotAllTracks}
\alias{plotRosePlot}
\title{Plot scaled cell migration}
\usage{
plotAllTracks(
  object,
  across = "condition",
  across_subset = NULL,
  relevel = NULL,
  time_subset = NULL,
  phase = NULL,
  n_cells = 100,
  color_by = across[1],
  linetype = "solid",
  linesize = 0.75,
  clrp = "milo",
  clrp_adjust = NULL,
  display_annotation = TRUE,
  set_seed = NULL,
  ...,
  verbose = TRUE
)

plotRosePlot(
  object,
  across = "condition",
  across_subset = NULL,
  relevel = NULL,
  time_subset = NULL,
  phase = NULL,
  n_cells = 100,
  color_by = across[1],
  linetype = "solid",
  linesize = 0.75,
  clrp = "milo",
  clrp_adjust = NULL,
  display_annotation = TRUE,
  set_seed = NULL,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector, list of character vectors or NULL (see details for
more). Specifies the particular groups of interest the grouping variable specified in argument
\code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the choice of color.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{n_cells}{Numeric calue. Determines the number of cells that are randomly chosen from
every group to be displayed. Useful to keep plots insightful and aesthetically pleasing.}

\item{linetype}{Character value. Valid options are \emph{'solid', 'twodash', 'longdash', 'dotted'}
and \emph{'dotdash'}.}

\item{linesize}{Numeric value. Denotes the size of the lines drawn.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{display_annotation}{Logical. If set to TRUE the number of cells (input
for argument \code{n_cells}) is displayed in the right upper corner of each plot.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A ggplot.
}
\description{
Visualizes the cells migration in a typical migration plot.
Scales the cell's x- and y-coordinates such that all cell's migration
paths start from the same position.
}
\details{
Argument \code{across} can be a single character value or a character vector of
length two. If character vector, e.g. \code{across = c(\emph{'cell_line', 'condition'})},
argument \code{across_subset} - if specified - should be a corresponding list:
\code{across_subset = list(cell_line = \emph{c('GCL24', 'GCL38')}, condition = c(\emph{'Ctrl', 'chlorambucil'}))}.
}
