% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-miscellaneous.R
\name{plotCellCount}
\alias{plotCellCount}
\title{Plot cell count}
\usage{
plotCellCount(object, across, color_by = across, phase = NULL)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{across}{Character value, character vector or NULL (see details for more).
Specifies the grouping variable of interest.

Use \code{getGroupingVariableNames()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{color_by}{Character value. Denotes the discrete variable with
which to color the columns of the plot. Defaults to input for
argument \code{across}.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\value{
A ggplot.
}
\description{
Visualizes the number and distribution of cells across
a discrete feature of choice.
}
