% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{plotCorrelationSD}
\alias{plotCorrelationSD}
\title{Plot correlation variance across groups}
\usage{
plotCorrelationSD(
  object,
  variable_set,
  method_corr = NULL,
  phase = NULL,
  across = NULL,
  relevel = NULL,
  variables_subset = NULL,
  signif_level = NULL,
  clrsp = NULL
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method_corr}{Character value. Specifies the correlation
method of interest. Use \code{validMethodsCorrelation()} to obtain all valid
input options.}

\item{across}{Character vector (!). Denotes all grouping variables of interest.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across.subset} and
\code{variables} determines the order in which the groups of interest are displayed.}
}
\value{


A ggplot.
}
\description{
Uses the style of correlation matrices to visualize
the standard deviation of the correlation values across groups of
a grouping variable - indicating which grouping is responsible for changes
in correlation.
}
