% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-hierarchical.R
\name{plotDendrogram}
\alias{plotDendrogram}
\title{Visualize hierarchical clustering}
\usage{
plotDendrogram(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{ggdendro::ggdendrogram()}.}

\item{k}{Numeric value or NULL. If numeric the branches are colored according to their cluster assignment
whereby the number of clusters is equal to \code{k}.}

\item{h}{Numeric value or NULL. If numeric the branches are colored according to their cluster assignment
whereby the number of clusters is equal to the results of cutting the hierarchical tree at height \code{h}..}

\item{branch_size}{Numeric value. Denotes the thickness of the branches.}
}
\value{


A ggplot.
}
\description{
Plots the hierarchical tress agglomerated by \code{agglomerateHierarchicalCluster()}.
}
\details{
Iterates over all valid combinations of the \code{method_*} arguments and creates a dendrogram
if the respective clustering results are found. \code{nrow} and \code{ncol} can be used to
align the plots.
}
