% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/plot-miscellaneous.R
\name{plotScatterplot}
\alias{plotScatterplot}
\title{Plot a scatterplot}
\usage{
plotScatterplot(object, ...)

plotScatterplot(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{variable_x, variable_y}{Character values. The numeric variables to be plotted on the x-axis and
y-axis. Use \code{getNumericVariableNames()} to obtain all valid input options.}

\item{display_corr}{Logical value. If set to TRUE the correlation of the x- and y-variable is calculated
and displayed on the plot (or each subplot if \code{across} is specified.)}

\item{corr_method}{Character value. Denotes the method with which to compute the correlation values if
\code{display_corr} is set to TRUE. Defaults to \emph{'pearson'}.}

\item{corr_pmin}{Numeric value. The minimum p-value that is displayed as a number. Everything below it is
displayed as \emph{ < \code{corr_pmin}}.}

\item{corr_pos_x, corr_pos_y}{Numeric values or NULL. Specify the exact position of the text used to display the correlation
results. If set to NULL defaults to left upper corner of the plot.}

\item{corr_text_sep}{Character value. The string that separates correlation value and respective p-value.}

\item{corr_text_size}{Numeric value. The size used to print the correlation results.}
}
\value{
A ggplot.
}
\description{
Convenient wrapper around a variety of scatterplot functionalities. See details for more.
}
\details{
Argument \code{across} can be specified as a character vector of length 2. If so, argument
\code{across_subset} must be a list of character vectors whereby the names of the list correspond to the input for \code{across}.
}
