% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-generics.R, R/clustering-kmeans.R
\name{plotScreeplot}
\alias{plotScreeplot}
\alias{plotScreeplot,Cypro-method}
\title{Plot a screeplot}
\usage{
plotScreeplot(object, ...)

\S4method{plotScreeplot}{Cypro}(
  object,
  phase,
  ks,
  fset = "all_features",
  methods_kmeans = "Hartigan-Wong",
  color = "steelblue",
  display_cols = TRUE,
  display_line = TRUE,
  display_points = TRUE
)

\S4method{plotScreeplot}{Cypro}(
  object,
  phase,
  ks,
  fset = "all_features",
  methods_kmeans = "Hartigan-Wong",
  color = "steelblue",
  display_cols = TRUE,
  display_line = TRUE,
  display_points = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{fset}{Character value. The feature set of interest. Obtain all added feature set names
with \code{getFeatureSetNames()}.}

\item{methods_kmeans}{Character vector. Specifies the kmeans methods
of interest. Use \code{validMethodsKmeans()} to obtain all valid input options.}

\item{display_cols}{Logical value. If set to TRUE columns are used additionally
to display the results.}

\item{display_line}{Logical value. If set to TRUE a line is used additionally
to display the results.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}
}
\value{
A ggplot.
}
\description{
Plots quality of clustering results
suggested by kmeans with a screeplot.
}
\details{
If \code{ks} and \code{methods_kmeans} are specified only
the resulting combinations are included. Else all found combinations
are included.
}
