% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module-migration.R
\name{plotSingleTracks}
\alias{plotSingleTracks}
\title{Plot single cell migration}
\usage{
plotSingleTracks(
  object,
  cell_ids,
  phase = "all",
  size = 1,
  color_by = NULL,
  scales = "free"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{cell_ids}{Character vector. Denotes the cell ids of interest.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{scales}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}
}
\value{
A ggplot.
}
\description{
Visualizes the migration of single cells of interest.
}
