% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-statistics.R
\name{plotStatisticsInteractive}
\alias{plotStatisticsInteractive}
\title{Plot statistics related plots interactively}
\usage{
plotStatisticsInteractive(object, phase = NULL, drop_na = TRUE)
}
\arguments{
\item{object}{A valid cypro object.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\description{
Opens an interactive application in which statistic related
plots like violinplots, boxplots as well as statistical tests across
groups can be conducted.
}
