% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics.R
\name{plotStatisticsInteractive}
\alias{plotStatisticsInteractive}
\title{Plot statistics related plots interactively}
\usage{
plotStatisticsInteractive(object, phase = NULL, drop_na = TRUE)
}
\arguments{
\item{object}{Any S4-object for whose class a method has been defined.}

\item{phase}{Character or numeric. If character, the ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest or \emph{'all'}. If numeric, the number referring to the phase.

If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}
}
\description{
Opens an interactive application in which statistic related
plots like violinplots, boxplots as well as statistical tests across
groups can be conducted.
}
