% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-lapse.R
\name{plotTimeHeatmap}
\alias{plotTimeHeatmap}
\title{Visualize changes of cell characteristics over time}
\usage{
plotTimeHeatmap(
  object,
  variable,
  across = "condition",
  across_subset = NULL,
  phase = "all",
  n_cells = 100,
  color = NA,
  colors = viridis::viridis(15),
  smooth = TRUE,
  smooth_span = 0.25,
  arrange_rows = "maxima",
  verbose = TRUE,
  ...,
  in_shiny = FALSE
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
cells of your experiment in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{phase}{Character value. The ordinal value referring to the phase of interest (e.g. \emph{'first'}, \emph{'second'} etc.).
referring to the phase of interest. If set to NULL takes the phase denoted as default with \code{adjustDefault()}.

Ignored if the experiment design contains only one phase.}

\item{n_cells}{Numeric calue. Determines the number of cells that are randomly chosen from
every group to be displayed. Useful to keep plots insightful and aesthetically pleasing.}

\item{colors}{A vector of colors that is used as the heatmap's color spectrum.}

\item{smooth}{Logical. If set to TRUE the values are smoothed.}

\item{smooth_span}{NUmeric value. Denotes the smoothing span used.}

\item{arrange_rows}{Character value. Either \emph{'maxima'} or \emph{'none'}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{pheatmap::pheatmap()}.}
}
\description{
Uses the design of a heatmap to visualize feature
dynamics over time across a grouping variable of interest.
}
