% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWellPlate.R
\name{plotWellPlate}
\alias{plotWellPlate}
\alias{plotWellPlate,layout_df-method}
\alias{plotWellPlate,layout_df_mp-method}
\alias{plotWellPlate,Cypro-method}
\title{Visualize well plates}
\usage{
plotWellPlate(object, ...)

\S4method{plotWellPlate}{layout_df}(
  object,
  color_by = NULL,
  fill_by = NULL,
  plot_type = "tile",
  well_alpha = 0.9,
  well_color = "black",
  well_size = 13.5,
  well_stroke = 2,
  display_border = TRUE,
  border_color = "black",
  border_size = 1,
  display_edge = TRUE,
  edge_color = "black",
  edge_size = 1,
  display_labels = TRUE,
  labels_color = "black",
  clrp = "milo",
  clrp_adjust = NULL,
  ncol = NULL,
  nrow = NULL,
  selected_wells = NULL
)

\S4method{plotWellPlate}{layout_df_mp}(
  object,
  color_by = NULL,
  fill_by = NULL,
  plot_type = "tile",
  well_alpha = 0.9,
  well_color = "black",
  well_size = 13.5,
  well_stroke = 2,
  display_border = TRUE,
  border_color = "black",
  border_size = 1,
  display_edge = TRUE,
  edge_color = "black",
  edge_size = 1,
  display_labels = TRUE,
  labels_color = "black",
  clrp = "milo",
  clrp_adjust = NULL,
  ncol = NULL,
  nrow = NULL,
  selected_wells = NULL
)

\S4method{plotWellPlate}{Cypro}(
  object,
  well_plate = NULL,
  color_by = NULL,
  fill_by = NULL,
  plot_type = "tile",
  well_alpha = 0.9,
  well_color = "black",
  well_size = 13.5,
  well_stroke = 2,
  display_border = TRUE,
  border_color = "black",
  border_size = 1,
  display_edge = TRUE,
  edge_color = "black",
  edge_size = 1,
  display_labels = TRUE,
  labels_color = "black",
  clrp = "milo",
  clrp_adjust = NULL,
  ncol = NULL,
  nrow = NULL,
  selected_wells = NULL
)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Plots a well plate.
}
