% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-miscellaneous.R
\name{plotWellPlate}
\alias{plotWellPlate}
\title{Plot the well plate set up}
\usage{
plotWellPlate(
  object,
  well_plate = NULL,
  color_by = "condition",
  size_well = 13.5,
  display_labels = TRUE,
  clrp_adjust = NULL,
  make_pretty = NULL,
  ...
)
}
\arguments{
\item{object}{A valid cypro object.}

\item{well_plate}{Character value. The name of the well plate of interest. Valid inputs can be obtained
via \code{getWellPlateNames()}.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}
}
\value{
A ggplot.
}
\description{
Plot the well plate set up
}
