% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-subset-history.R
\name{printSubsetHistory}
\alias{printSubsetHistory}
\title{Print subset information}
\usage{
printSubsetHistory(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A printed message via \code{writeLines()}.
}
\description{
If the object is the results of one or more
subsetting processes you can print the respective information
into the console to keep track of the objects origin.
}
