% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{Progress-class}
\alias{Progress-class}
\alias{Progress}
\title{The Progress Class}
\description{
S4-object that keeps track of the necessary function
calls that have to be made in order to use a \code{Cypro}-object.
}
\section{Slots}{

\describe{
\item{\code{experiment}}{character. The name of the \code{Cypro} object it is part of.}

\item{\code{designExperiment}}{logical.}

\item{\code{assignVariables}}{logical.}

\item{\code{loadData}}{logical.}
}}

