% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier-detection.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove outliers from object}
\usage{
removeOutliers(object, method_outlier = NULL, verbose = NULL)
}
\arguments{
\item{object}{A valid cypro object.}

\item{method_outlier}{Character vector. Specifies the method/algorithm of interest.

Use \code{validOutlierDetectionMethods()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Removes the data from the cell ids detected by \code{detectOutliers()} from
the object. As this changes the data set all clustering-, correlation- and dimensional
reduction results are reset.
}
\seealso{
detectOutliers()
}
