% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{saveCyproObject}
\alias{saveCyproObject}
\title{Save cypro object}
\usage{
saveCyproObject(object, verbose = TRUE)
}
\arguments{
\item{object}{A valid cypro object.}
}
\description{
Save cypro object
}
