% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setStorageDirectory}
\alias{setStorageDirectory}
\title{Set default storage directory}
\usage{
setStorageDirectory(object, directory)
}
\arguments{
\item{object}{A valid cypro object.}

\item{directory}{Character value. The directory under which
the object is supposed to be stored via \code{saveCyproObject()}.}
}
\value{
An updated version of the input \code{cypro}-object.
}
\description{
Set default storage directory
}
