% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{setUpWellPlateDf}
\alias{setUpWellPlateDf}
\title{Create data.frame representing a well plate}
\usage{
setUpWellPlateDf(type = "8x12 (96)", phases = NULL)
}
\arguments{
\item{type}{Character value. One of \emph{'2x3, '3x4', '4x6, '6x8', '8x12'}}
}
\description{
Sets up a data.frame in which each observation refers
to a well.
}
