% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setWellInfo}
\alias{setWellInfo}
\alias{setWellInfo.layout_df}
\title{Set well plate layout information}
\usage{
setWellInfo(df, ...)

\method{setWellInfo}{layout_df}(
  df,
  selected_wells,
  cell_line = NULL,
  condition = NULL,
  verbose = TRUE,
  in_shiny = FALSE
)
}
\arguments{
\item{df}{A data.frame of a class for which a method has been defined.}

\item{selected_wells}{Character vector. Denotes the wells for which the information
is supposed to be set.}

\item{cell_line}{Character value, NA or NULL. The cell line seeded in the wells denoted in
\code{selected_wells}.}

\item{condition}{Character value, NA, or NULL. The condition of the well. If multiple phase
experiment a character vector of length equal to the number of phases of
the experiment design.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame.
}
\description{
Subsequently fills the well plate \code{layout_df} with information.
}
