% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetBy.R
\name{subsetByCellID}
\alias{subsetByCellID}
\alias{subsetByCellID,CyproScreening-method}
\alias{subsetByCellID,CyproTimeLapse-method}
\title{Create data subset by cell IDs}
\usage{
subsetByCellID(object, ...)

\S4method{subsetByCellID}{CyproScreening}(
  object,
  cell_ids,
  reasoning = NA_character_,
  prev_subset = NULL,
  verbose = TRUE,
  ...
)

\S4method{subsetByCellID}{CyproTimeLapse}(
  object,
  cell_ids,
  reasoning = NA_character_,
  prev_subset = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{cell_ids}{Character vector. Denotes the cells to keep unambiguously with their cell IDs.}

\item{reasoning}{Character value. Allows two write a short description of how the cell ids according
to which the object is filtered were selected. This description is included in the output of \code{printSubsetHistory()}.
Ignored if set to NULL.}

\item{prev_subset}{Object of class \code{CyproSubset} that inherits from \code{CyproSubsetByCellID} or NULL.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{new_name}{Character value. Denotes the name of the output object. If set to NULL the name of the
input object is taken and suffixed with \emph{'_subset'}.}

\item{new_directory}{Character value. Storage directory for the new object.}
}
\value{
The input object containing a data subset.
}
\description{
Subset functions allow to conveniently split your data by certain characteristics such
as cell lines, conditions, cluster etc. or for specific cell ids. This might be useful if you want apply some machine learning
algorithms such as clustering and correlation on only a subset of cells. See sections below for more information.
}
\details{
Creating subsets of your data affects analysis results such as clustering,
dimensional reduction and correlation which is why slot @analysis is reset.

Within the data of slot @cdata, the cluster data.frame is emptied. Additionally,
levels of grouping variables that are not used any longer after the subsetting
are dropped.

To prevent inadvertent overwriting the default directory is reset as well.
Make sure to set a new one via \code{setDefaultDirectory()}.

The name of the returned \code{Cypro} object (slot @experiment) is suffixed
with \emph{_Subset\code{<nth_subset>}}.

The mechanism and the information with which you create the subset is stored in the output object.
Use \code{printSubsetHistory()} to reconstruct the way from the original object
to the current one.
}
\section{Subset mechanism}{


Cells with cell IDs that were not part of the cell IDs in argument \code{cell_ids} are discarded.
}

